==> scripts/01_Sample_metadata.Rmd <==
---
title: "Initial analysis"
date: "2023-04-12"
output: 
  html_document:
    toc: true
    toc_float:
      collapsed: true
    toc_depth: 3
    number_sections: false
    theme: lumen
---

# Background

We have cohort of samples profiled for proteomics. 

In regards to proteomics. Samples were profiled ;

1) Psomagen cohort
2) Olink 2023 cohort (Q-13356) - Also called Project 2
3) Olink 2024 cohort (Q-15806) - Also called Project 4

# Objectives

1. Gather data
2. Clean up clinical data
3. Viz of data

# Conclusion 

We extracted the metadata from 3 proteomics experiment. Identified samples to be used in future analysis. We have the final patient and sample numbers. 

# Pre-processing 

## Loading packages

```{r, message=FALSE, warning=FALSE}
library(here)
library(tidyverse)
library(readxl)   
library(readr)
library(stringr)

library(tidyverse)
library(ggplot2)
library(VennDiagram)
library(ggvenn)
```


## Directories

```{r}
wd <- list()
wd$main <- here()
wd$data <- file.path(wd$main, "data")
wd$manishData <- file.path(wd$data, "raw/data_fromManish")
wd$d2024 <- file.path(wd$data, "updated_2024")
wd$d2024_npx <- file.path(wd$d2024, "NPX_data")
wd$output <- file.path(wd$main, "output")
wd$script <- file.path(wd$main, "scripts")

wd$outData <- file.path(wd$output, "data")
wd$outCurr <- file.path(wd$output, "01_InitialAnalysis")
```


Create directories

```{r}
if (!file.exists(wd$outCurr)) {
  dir.create(wd$outCurr)
} else {
  print("Directory already exists")
}
```

Load functions

```{r}
source(file.path(wd$script, "functions.R"))
```


# Objective 1

**Gather data**

We start with the information that there are 312 samples sent for Proteomics profiling. For each patient, we have the HCI collection ID (HCI_cID) and the sample ID by Olink. 

We read the master file of all the HCI_cID into R

```{r}
df_hci_id <- read.csv(file.path(wd$d2024, "clinical_data/MetaData_Proteomics.csv"))
nrow(df_hci_id)
```

We also specify in the clinical file provided by Enos and Becky. This file is available in the cloud as well at - 

```{r}
meta_file <- file.path(wd$d2024, "clinical_data/Proteomic Demographic 06_23_25.xlsx")
```

This excel file contains a few sheets, each sheet represent the particular disease cohort. 

Next, for each the proteomics batch, we read in the NPX reading. 

## Psomagen 

**E3072_AN00003660_AN00003661_NPX_2022-04-21-columns**

This file is the NPX values from Olink. Samples are listed in the `SampleID` column

```{r}
# Read the files
dat_NPX <- readxl::read_excel(file.path(wd$d2024_npx, "batch_01_E3072_AN00003660_AN00003661_NPX_2022-04-21-columns.xlsx"))
```

Get Sample ID (listed in the `SampleID` column)

```{r}
b1_id <- unique(dat_NPX$SampleID)
length(b1_id)
```

There are `length(b1_id)` samples assayed by Olink. This includes "Sample Control" as well, an internal control used by Olink. 

Find the number of samples not including the internal control

```{r}
b1_noCtrl <- b1_id[!grepl("SC", b1_id)]
length(b1_noCtrl)
```

Lets check if there are equal number in our master sample sheet

```{r}
df_hci_id %>% filter(Psomagen == "Y") %>% nrow()
```

Its the same number of samples from NPX data and our sample information sheet. 

## Olink Q-13356 

**Q-13356_Kohli_EXTENDED_NPX_2024-03-11**

This file is the NPX values from Olink. Samples are listed in the `SampleID` column

Read the NPX data. Initially we read in this file - but this file was not "intensity normalized" rather it was "plate control" normalized. 

```{r}
# Read the files
# dat_NPX2 <- readr::read_delim(file.path(wd$d2024_npx, "batch_02_Q-13356_Kohli_EXTENDED_NPX_2024-03-11.csv"), delim = ";") 
```

We requested Olink to resend the "intensity normalzed" NPX, to match Psomagen data

```{r}
dat_NPX2 <- readr::read_delim(file.path(wd$d2024_npx, "Re_delivery_Intensity_based/Q-13356_Data Delivery/Q-13356_Kohli_Intensity_EXTENDED_NPX_2024-10-30.csv"), delim = ";")
```
Get Sample ID

```{r}
b2_id <- unique(dat_NPX2$SampleID)
length(b2_id)
```

There are `r `length(b2_id)` samples assayed by Olink. This includes "Sample Control" as well, an internal control used by Olink. 

Find the number of samples not including the internal control

```{r}
b2_noCtrl <- b2_id[!grepl("CONTROL_SAMPLE|NEG_CTRL|PLATE_CTRL", b2_id)]
length(b2_noCtrl)
```

Lets check if there are equal number of samples in our master sample sheet

```{r}
df_hci_id %>% filter(Q_13356 == "Y") %>% nrow()
```


Based on the NPX, we have `r length(b2_noCtrl)` samples. Meanwhile in the clinical file, we have 87. 

There is one less patient in the clinical sheet. 

Clarification was given by Claire 

> From: Claire Hanson <Claire.Hanson@hci.utah.edu>   
> Sent: Thursday, June 20, 2024 9:33 AM.  
To: Zaki Wilmot <zaki.wilmot@hci.utah.edu>; Manish Kohli <Manish.Kohli@hci.utah.edu>; Enos Ampaw <Enos.Ampaw@hci.utah.edu>   
> Subject: Re: Overlapping files
> 
> Hi Zaki, 
> 
> The duplicate sample is 34 and 78 (2021-2165). We used sample ID 34 as one of the bridging samples as well. 
> 
>Thanks, 
Claire 

We noted here that there is one sample missing a clinical ID. We go back to the email from Claire that mentioned the sample "2021-2165" was profiled twice and should have the assigned sample_id : 34 and 78 

```{r}
#filter(d2, is.na(HCI_cID))
```

We see sample ID 78 is missing clinical data. So we just exclude sample 78 for now

```{r}
#d2 <- d2 %>% 
#  filter(!sample_id_Q13356 == 78)
```



## Olink Q-15806


Read the NPX data. Samples are listed in the `SampleID` column

Again, similar to Q-13356, we have two NPX file based on the differnt normalization approach. We use the "intensity based" to keep it similar to Psomagen,

```{r}
# Read the files
#dat_NPX3 <- readr::read_delim(file.path(wd$d2024_npx, "Q-15806_Kohli_NPX_2024-07-29.csv"), delim = ";")
dat_NPX3 <- readr::read_delim(file.path(wd$d2024_npx, "Re_delivery_Intensity_based/Q-15806_Data Delivery/Q-15806_Kohli_EXTENDED_NPX_2024-10-30.csv"), delim = ";")
```

Get Sample ID

```{r}
b3_id <- unique(dat_NPX3$SampleID)
length(b3_id)
```

There are `r `length(b3_id)` samples assayed by Olink. This includes "Sample Control" as well, an internal control used by Olink. 

Find the number of samples not including the internal control.

```{r}
b3_noCtrl <-  b3_id[!grepl("CONTROL_SAMPLE|NEG_CTRL|PLATE_CTRL", b3_id)]
length(b3_noCtrl)
```

Lets check if there are equal number of samples in our master sample sheet

```{r}
df_hci_id %>% filter(Q_15806 == "Y") %>% nrow()
```


## Match protein IDs

Make a table of proteins and OlinkIDs

```{r}
d1 <- dat_NPX %>% select(OlinkID, Assay) %>% distinct() %>% 
  mutate(ID_assay = paste(OlinkID, Assay, sep="_"))
d2 <- dat_NPX2 %>% filter(!str_detect(Assay, "control")) %>% select(OlinkID, Assay) %>% distinct() %>% 
  mutate(ID_assay = paste(OlinkID, Assay, sep="_"))

d_all <- rbind(d1,d2) %>% distinct()


d_all <- d_all %>% 
  mutate(inPsom = if_else(ID_assay %in% d1$ID_assay, "Y", "N"),
         inQ1335 = if_else(ID_assay %in% d2$ID_assay, "Y", "N"))

write.csv(d_all, file.path(wd$outCurr, "Proteins_overlap.csv"))

```

We note there are three proteins with mis-match annotation, lets correct this

```{r}
d_all <- d_all %>% 
  mutate(
    Assay = case_when(
      OlinkID == "OID20125" ~ "NT-proBNP",
      OlinkID == "OID20857" ~ "CERT1",
      OlinkID == "OID21084" ~ "WARS1",
      TRUE ~ Assay))
  
# Create a list of Olink paired with protein ID
df_id_pro <- d_all %>% select(OlinkID, Assay) %>% distinct()

write.csv(df_id_pro, file.path(wd$outCurr, "Olink_assay_table.csv"))
```

# Objective 2

**Clean up clinical data**

Matt provided us with an cohort information in 30 May 2025. This contained information defining pre- and post- treated samples

```{r}
df_cohort <- read_xlsx(file.path(wd$d2024, "clinical_data/Cohort_grouping_2025_05_30.xlsx")) %>% 
  select(mrn:txt_stat) %>% 
  rename(Psomagen = Discovery,
         Q_13356 = Project_2,
         Q_15806 = Project_4)
```


## Cohort summary

Enos provided us with an updated clinical data. We stored this as `meta_file`.

Tab 1 contains cohort A. 
Tab 2 contains cohort B. 
Tab 3 contains cohort C. 

### Cohort A

For cohort A, lets get the HCI_cID and collection date

```{r}
df_clin <- read_xlsx(meta_file, sheet = 1)

dd <- df_clin
df_a <- data.frame(
  mrn = dd$`MRN (UUHSC)`,
  age = dd$Age_at_time_of_enrollment,
  HCI_cID = c(dd$Sample1_HCI_Num, dd$Sample2_HCI_Num),
  date = c(dd$Sample1_date, dd$`Sample 2 Start date`)
  ) %>% filter(!is.na(HCI_cID)) %>% 
  mutate(cohort = "A") %>% 
  filter(HCI_cID %in% df_hci_id$HCI_cID)


df_a2 <- data.frame(
    mrn = dd$`MRN (UUHSC)`,
    age = dd$Age_at_time_of_enrollment,
    HCI_cID = c(dd$Sample1_HCI_Num, dd$Sample2_HCI_Num),
    date = c(dd$Sample1_date, dd$`Sample 2 Start date`),
    psa = c(dd$sample1_psa, dd$sample2_psa),
    alk_phos = c(dd$sample1_alk_phos, dd$sample2_alk_phos),
    ldh = c(dd$sample1_ldh, dd$sample2_ldh),
    testo = NA,  # No testosterone data for Cohort A
    hemoGlo = c(dd$`Hemoglobin at s1`, rep(NA, length(dd$Sample1_HCI_Num))),  # Only sample 1 has hemoglobin data
    surgery_type = dd$Surgery_Type,
    surgery_date = dd$Surgery_date,
    radiation_type = dd$Radiation_Type,
    radiation_date = dd$Radiation_date,
    glsn_diag = dd$diag_glsn,
    death_date = c(dd$`Death date`),
    death_status = dd$`Deceased?`)  %>%
  filter(!is.na(HCI_cID)) %>%
  mutate(cohort = "A") %>%
  mutate(
    psa = as.numeric(gsub("<", "", psa)),
    testo = as.numeric(gsub("<", "", testo))
  ) %>%
  filter(HCI_cID %in% df_hci_id$HCI_cID) %>% 
  # Convert dates to Date type
  mutate(
    date = as.Date(date),
    surgery_date = as.Date(surgery_date),
    radiation_date = as.Date(radiation_date),
    # Calculate days from treatment to sample (positive if treatment was before sample)
    days_since_surgery = abs(as.numeric(difftime(date, surgery_date, units = "days"))),
    days_since_radiation = abs(as.numeric(difftime(date, radiation_date, units = "days"))),
    # Create sur_or_rad column: pick treatment that happened before sample and is closest to sample date
    sur_or_rad = case_when(
    # Both treatments happened before sample - pick the closer one (smaller time difference)
    !is.na(days_since_surgery) & days_since_surgery > 0 &
    !is.na(days_since_radiation) & days_since_radiation > 0 ~
    ifelse(days_since_surgery < days_since_radiation, surgery_type, radiation_type),
    # Only surgery happened before sample
    !is.na(days_since_surgery) & days_since_surgery > 0 ~ surgery_type,
    # Only radiation happened before sample
    !is.na(days_since_radiation) & days_since_radiation > 0 ~ radiation_type,
    # No treatment before sample
    TRUE ~ NA_character_)
    ) 
```

### Cohort B

For cohort B, lets get the HCI_cID and collection date. We also want to add the treatment date. 

```{r}
df_clin <- read_xlsx(meta_file, sheet = 2)

dd <- df_clin

HCI_cID = c(dd$Sample1_HCI_Num, dd$Sample2_HCI_Num, dd$Sample3_HCI_Num)


df_b <- data.frame(
  mrn = dd$`MRN (UUHSC)`,
  age = dd$Age_at_time_of_enrollment,
  HCI_cID = c(dd$Sample1_HCI_Num, dd$Sample2_HCI_Num, dd$Sample3_HCI_Num),
  date = c(dd$Sample1_date, dd$`Sample 2 Start date`, dd$`sample3 date`),
  psa = c(dd$sample1_psa, dd$sample2_psa, dd$sample3_psa),
  alk_phos = c(dd$sample1_alk_phos, dd$sample2_alk_phos, dd$sample3_alk_phos),
  ldh = c(dd$sample1_ldh, dd$sample2_ldh, dd$sample3_ldh),
  testo = c(dd$sample1_testos, dd$`Sample 2 Testerone`, dd$`Sample 3 Testosterone`),
  hemoGlo = c(dd$`Hemoglobin at s1`, rep(NA, (length(HCI_cID) - length(dd$`Hemoglobin at s1`)  ) )),
  surgery_type = dd$Surgery_Type,
  surgery_date = dd$Surgery_date,
  radiation_type = dd$Radiation_Type,
  radiation_date = dd$Radiation_date,
  glsn_diag = dd$diag_glsn,
  mets_stage = dd$`De Novo Mets`,
  mets_volume = dd$chaarted_volume_at_mhspc,
  txt_ori = dd$txt_stat,
  death_date = c(dd$`Death date`),
  death_status = dd$`Deceased?`,
  # Info on HSPC treatment
  txt_nhspc_1st_type = dd$`1st_nmhspc_rx`,
  txt_nhspc_1st_date = dd$`1st_nmhspc_rx_date`,
  txt_mhspc_1st_type = dd$`1st_mhspc_rx`,
  txt_mhspc_1st_date = dd$`1st_mhspc_rx_date`,
  txt_mhspc_2nd_type = dd$`2nd_mhspc_rx`,
  txt_mhspc_2nd_date = dd$`2nd_mhspc_rx_date`,
  # Info on CRPC progression
  progressed_to_crpc = dd$`Progressed to CRPC`,
  crpc_biochem_date = dd$biochem_crpc_date,
  txt_nmcrpc_1st_date = dd$`1st_nmcrpc_rx_date`,
  txt_mcrpc_1st_date = dd$`1st_mcrpc_rx_date`,
  txt_mcrpc_2nd_date = dd$`2nd_mcrpc_rx_date`
) %>%
  filter(!is.na(HCI_cID)) %>%
  mutate(
    cohort = "B",
    psa = as.numeric(gsub("<", "", psa)),
    testo = as.numeric(gsub("<", "", testo))
  ) %>%
  filter(HCI_cID %in% df_hci_id$HCI_cID) %>% 
    # Convert dates to Date type
  mutate(
    date = as.Date(date),
    surgery_date = as.Date(surgery_date),
    radiation_date = as.Date(radiation_date),
    # Calculate days from treatment to sample (positive if treatment was before sample)
    days_since_surgery = abs(as.numeric(difftime(date, surgery_date, units = "days"))),
    days_since_radiation = abs(as.numeric(difftime(date, radiation_date, units = "days"))),
    # Create sur_or_rad column: pick treatment that happened before sample and is closest to sample date
    sur_or_rad = case_when(
    # Both treatments happened before sample - pick the closer one (smaller time difference)
    !is.na(days_since_surgery) & days_since_surgery > 0 &
    !is.na(days_since_radiation) & days_since_radiation > 0 ~
    ifelse(days_since_surgery < days_since_radiation, surgery_type, radiation_type),
    # Only surgery happened before sample
    !is.na(days_since_surgery) & days_since_surgery > 0 ~ surgery_type,
    # Only radiation happened before sample
    !is.na(days_since_radiation) & days_since_radiation > 0 ~ radiation_type,
    # No treatment before sample
    TRUE ~ NA_character_)
    )
```

#### ADT treatment

We want to figure out when is the first ADT date that is given. Lets first list all the treatments given and figure out which one is ADT.

```{r}
# Create columns indicating if ADT was given and the earliest date of ADT given for each patient
df_b2 <- df_b %>%
  mutate(
    # Identify if each treatment type is ADT
    nhspc_adt = str_detect(tolower(txt_nhspc_1st_type), "adt"),
    mhspc1_adt = str_detect(tolower(txt_mhspc_1st_type), "adt"),
    mhspc2_adt = str_detect(tolower(txt_mhspc_2nd_type), "adt"),
    # Convert all treatment dates to Date type
    nhspc_adt_date = if_else(nhspc_adt, as.Date(txt_nhspc_1st_date), as.Date(NA)),
    mhspc1_adt_date = if_else(mhspc1_adt, as.Date(txt_mhspc_1st_date), as.Date(NA)),
    mhspc2_adt_date = if_else(mhspc2_adt, as.Date(txt_mhspc_2nd_date), as.Date(NA))
  ) %>%
  rowwise() %>%
  mutate(
    # Was ADT given in any of the three?
    adt_given = any(c(nhspc_adt, mhspc1_adt, mhspc2_adt), na.rm = TRUE),
    # Find the earliest date of ADT given
    adt_earliest_date = min(c(nhspc_adt_date, mhspc1_adt_date, mhspc2_adt_date), na.rm = TRUE)
  ) %>%
  ungroup() %>%
  mutate(
    # If no ADT was given, set adt_earliest_date to NA
    adt_earliest_date = if_else(adt_given, adt_earliest_date, as.Date(NA))
  )

df_b2 <- df_b %>%
  mutate(
    # Identify if each treatment type is ADT
    nhspc_adt = str_detect(tolower(txt_nhspc_1st_type), "adt"),
    mhspc1_adt = str_detect(tolower(txt_mhspc_1st_type), "adt"),
    mhspc2_adt = str_detect(tolower(txt_mhspc_2nd_type), "adt"),
    # Convert all treatment dates to Date type
    nhspc_adt_date = if_else(nhspc_adt, as.Date(txt_nhspc_1st_date), as.Date(NA)),
    mhspc1_adt_date = if_else(mhspc1_adt, as.Date(txt_mhspc_1st_date), as.Date(NA)),
    mhspc2_adt_date = if_else(mhspc2_adt, as.Date(txt_mhspc_2nd_date), as.Date(NA))
  ) %>%
  rowwise() %>%
  mutate(
    # Was ADT given in any of the three?
    adt_given = any(c(nhspc_adt, mhspc1_adt, mhspc2_adt), na.rm = TRUE),
    # Find the earliest and latest date of ADT given
    adt_earliest_date = min(c(nhspc_adt_date, mhspc1_adt_date, mhspc2_adt_date), na.rm = TRUE),
    adt_latest_date   = max(c(nhspc_adt_date, mhspc1_adt_date, mhspc2_adt_date), na.rm = TRUE)
  ) %>%
  ungroup() %>%
  mutate(
    # If no ADT was given, set adt_earliest_date and adt_latest_date to NA
    adt_earliest_date = if_else(adt_given, adt_earliest_date, as.Date(NA)),
    adt_latest_date   = if_else(adt_given, adt_latest_date, as.Date(NA)))
```

Now we want to create a column to indicate if the patient was given ADT within 30 days of the sample collecton date. For samples before the 30 days, label as Pre and after 30 days label as post. 

For some patients (with >90 days since ADT) Bekcy provided us with an updated ADT date as well. We read this updated dates

```{r}
df_adt <- read_xlsx(file.path(wd$d2024, "clinical_data/ADT_timeline_HSPC.xlsx"), sheet=1)
```

Based on the HCI_cID, we want to replace the ADT collection date to Becky's date. We replace the adt_earliest_date with the dates provided by Becky.

```{r}
df_b2 <- df_b2 %>%
  left_join(df_adt %>% select(HCI_cID, Becky_ADT_date), by = "HCI_cID") %>%
  mutate(
    adt_earliest_date = if_else(!is.na(Becky_ADT_date), as.Date(Becky_ADT_date), adt_earliest_date)
  ) %>%
  select(-Becky_ADT_date)
```

Lets calcualte the time of sample collection to ADT

```{r}
df_b2 <- df_b2 %>%
  mutate(
    # Calculate days from sample collection to ADT
    days_to_adt_early = as.numeric(difftime(date, adt_earliest_date, units = "days")),
    days_to_adt_late  = as.numeric(difftime(date, adt_latest_date, units = "days")),
    # Label as Pre if sample is within 30 days before/after earliest ADT, Post if >30 days after earliest ADT
    adt_status_early = case_when(
      is.na(adt_earliest_date) ~ "Pre",                # No ADT date available
      days_to_adt_early >= -30 & days_to_adt_early <= 30 ~ "Pre",  # Within 30 days before/after earliest ADT
      days_to_adt_early > 30 ~ "Post",                       # More than 30 days after earliest ADT
      days_to_adt_early < -30 ~ "Pre",                       # More than 30 days before earliest ADT 
      TRUE ~ NA_character_),
    adt_status_late = case_when(
      is.na(adt_latest_date) ~ "Pre",                # No ADT date available
      days_to_adt_late >= -30 & days_to_adt_late <= 30 ~ "Pre",  # Within 30 days before/after earliest ADT
      days_to_adt_late > 30 ~ "Post",                       # More than 30 days after earliest ADT
      days_to_adt_late < -30 ~ "Pre",                       # More than 30 days before earliest ADT 
      TRUE ~ NA_character_)
  )
```

When we compare if the Pre- or Post- is the same using early or late. 3 patients were different ;

- 2021-1600 
- 2020-3493 
- 2021-2049 

For now, we are using the days_to_adt_early to define the ADT status. Additionally, we are now considering the testosterone level as well. 

Anything labeled initially as "Pre" and have a testo value <50 will be considered "Post" 

Anything labeled initially as "Post" and have a testo value >50 will be considered as "Pre"

```{r}
df_b3 <- df_b2 %>% 
  mutate(
    updated_adt_early = case_when(
      adt_status_early == "Pre" & testo  <= 50 ~ "Post",
      adt_status_early == "Post" & testo  > 50 ~ "Pre",
      TRUE ~ adt_status_early
    )
  )
```

Number of samples in pre- or post-

```{r}
table(df_b3$updated_adt_early)
```

Lets check how many pre- post- pairs we have

```{r}
mrn_pre_post <- df_b3 %>%
  group_by(mrn) %>%
  summarise(
    has_pre = as.integer(any(updated_adt_early == "Pre", na.rm = TRUE)),
    has_post = as.integer(any(updated_adt_early == "Post", na.rm = TRUE))
  )

# View only those with both Pre and Post
mrn_pre_post %>%
  filter(has_pre == 1 & has_post == 1) %>% 
  dim()
```

#### CRPC progression

We want to also identify the earliest date of CRPC progression

```{r}
df_b3 <- df_b3 %>%
  mutate(
    # Convert all relevant CRPC treatment date columns to Date type
    crpc_biochem_date = as.Date(crpc_biochem_date),
    txt_nmcrpc_1st_date = as.Date(txt_nmcrpc_1st_date),
    txt_mcrpc_1st_date = as.Date(txt_mcrpc_1st_date),
    txt_mcrpc_2nd_date = as.Date(txt_mcrpc_2nd_date),
    # Find the earliest date of CRPC progression
    crpc_earliest_date = pmin(
      crpc_biochem_date,
      txt_nmcrpc_1st_date,
      txt_mcrpc_1st_date,
      txt_mcrpc_2nd_date,
      na.rm = TRUE
    )
  )
```


### Cohort C

For cohort C, lets get the HCI_cID and collection date. We also want to add the treatment date. 

```{r}
df_clin <- read_xlsx(meta_file, sheet = 5)

dd <- df_clin

HCI_cID = c(dd$Sample1_HCI_Num, dd$Sample2_HCI_Num, dd$Sample3_HCI_Num)

df_c <- data.frame(
  mrn = dd$`MRN (UUHSC)`,
  age = dd$Age_at_time_of_enrollment,
  HCI_cID = c(dd$Sample1_HCI_Num, dd$Sample2_HCI_Num, dd$Sample3_HCI_Num),
  date = c(dd$Sample1_date, dd$`Sample 2 Start date`, dd$`Sample3_Start date`),
  psa = c(dd$sample1_psa, dd$sample2_psa, dd$sample3_psa),
  ldh = c(dd$sample1_ldh, dd$sample2_ldh, dd$sample3_ldh),
  alk_phos = c(dd$sample1_alk_phos, dd$sample2_alk_phos, dd$sample3_alk_phos),
  alb = c(dd$`Albumin at s1`, rep(NA, (length(HCI_cID) - length(c(dd$`Albumin at s1`)  ) ) )),
  testo = c(dd$sample1_testos, dd$`Sample 2 Testos`, rep(NA, (length(HCI_cID) - length(c(dd$`sample1_testos`, dd$`Sample 2 Testos`)  ) ) )),
  hemoGlo = c(dd$`Hemoglobin at s1`, rep(NA, (length(HCI_cID) - length(dd$`Hemoglobin at s1`)  ) )),
  surgery_type = dd$Surgery_Type,
  surgery_date = dd$Surgery_date,
  radiation_type = dd$Radiation_Type,
  radiation_date = dd$Radiation_date,
  glsn_diag = dd$diag_glsn,
  mets_stage = dd$`De-Novo Metastatic (Y/N)`,
  death_date = c(dd$`Death date`),
  death_status = dd$`Deceased?`,
  txt_nmcrpc_1st_type = dd$`1st_nmcrpc_rx`,
  txt_nmcrpc_1st_date = dd$`1st_nmcrpc_rx_date`,
  txt_mcrpc_1st_type = dd$`1st_mcrpc_rx`,
  txt_mcrpc_1st_date = dd$`1st_mcrpc_rx_date`,
  txt_mcrpc_2nd_type = dd$`2nd_mcrpc_rx`,
  txt_mcrpc_2nd_date = dd$`2nd_mcrpc_rx_date`,
  txt_mhspc_1st_type = dd$`1st_mhspc_rx`,
  txt_mhspc_1st_date = dd$`1st_mhspc_rx_date`,
  txt_mhspc_2nd_type = dd$`2nd_mhspc_rx`,
  txt_mhspc_2nd_date = dd$`2nd_mhspc_rx_date`
) %>%
  filter(!is.na(HCI_cID)) %>%
  mutate(
    cohort = "C",
    psa = as.numeric(gsub("<", "", psa)),
    ldh = as.numeric(gsub("<", "", ldh)),
    alk_phos = as.numeric(gsub("<", "", alk_phos)),
    testo = as.numeric(gsub("<", "", testo))
  ) %>%
  filter(HCI_cID %in% df_hci_id$HCI_cID) %>% 
  # Convert dates to Date type
  mutate(
    date = as.Date(date),
    surgery_date = as.Date(surgery_date),
    radiation_date = as.Date(radiation_date),
    # Calculate days from treatment to sample (positive if treatment was before sample)
    days_since_surgery = abs(as.numeric(difftime(date, surgery_date, units = "days"))),
    days_since_radiation = abs(as.numeric(difftime(date, radiation_date, units = "days"))),
    # Create sur_or_rad column: pick treatment that happened before sample and is closest to sample date
    sur_or_rad = case_when(
    # Both treatments happened before sample - pick the closer one (smaller time difference)
    !is.na(days_since_surgery) & days_since_surgery > 0 &
    !is.na(days_since_radiation) & days_since_radiation > 0 ~
    ifelse(days_since_surgery < days_since_radiation, surgery_type, radiation_type),
    # Only surgery happened before sample
    !is.na(days_since_surgery) & days_since_surgery > 0 ~ surgery_type,
    # Only radiation happened before sample
    !is.na(days_since_radiation) & days_since_radiation > 0 ~ radiation_type,
    # No treatment before sample
    TRUE ~ NA_character_)
    )
```

We want to figure out when is the first CRPC treatment is given. Regardless of what CRPC treatment was given

```{r}
# If any treatment is given, find the earliest and latest treatment date.

# Cohort C: Find earliest and latest treatment date (any treatment)
df_c2 <- df_c %>%
  mutate(
    # Convert all treatment dates to Date type
    nmcrpc_date = as.Date(txt_nmcrpc_1st_date),
    mcrpc1_date = as.Date(txt_mcrpc_1st_date),
    mcrpc2_date = as.Date(txt_mcrpc_2nd_date)
  ) %>%
  rowwise() %>%
  mutate(
    # Find the earliest and latest date of any treatment given
    any_treatment_earliest_date = min(c(nmcrpc_date, mcrpc1_date, mcrpc2_date), na.rm = TRUE),
    any_treatment_latest_date   = max(c(nmcrpc_date, mcrpc1_date, mcrpc2_date), na.rm = TRUE),
    # Was any treatment given? (at least one non-NA date)
    any_treatment_given = any(!is.na(c(nmcrpc_date, mcrpc1_date, mcrpc2_date)))
  ) %>%
  ungroup() %>%
  mutate(
    # If no treatment was given, set dates to NA
    any_treatment_earliest_date = if_else(any_treatment_given, any_treatment_earliest_date, as.Date(NA)),
    any_treatment_latest_date   = if_else(any_treatment_given, any_treatment_latest_date, as.Date(NA))
  )
```

Now we want to create a column to indicate if the patient was given mCRPC treament within 30 days of the sample collection date. For samples wihtin 30 days (before or after the 30 days), label as Pre and after 30 day window label as post. 

Lets calcualte the time of sample collection to treatment

```{r}
df_c2 <- df_c2 %>%
  mutate(
    # Calculate days from sample collection to earliest and latest mCRPC (any) treatment
    days_to_mcrpc_early = as.numeric(difftime(date, any_treatment_earliest_date, units = "days")),
    days_to_mcrpc_late  = as.numeric(difftime(date, any_treatment_latest_date, units = "days")),
    # Label as Pre if sample is within 30 days before/after earliest mCRPC treatment, Post if >30 days after earliest mCRPC treatment
    mcrpc_status_early = case_when(
      is.na(any_treatment_earliest_date) ~ "Pre",                # No mCRPC treatment date available
      days_to_mcrpc_early >= -30 & days_to_mcrpc_early <= 30 ~ "Pre",  # Within 30 days before/after earliest mCRPC treatment
      days_to_mcrpc_early > 30 ~ "Post",                       # More than 30 days after earliest mCRPC treatment
      days_to_mcrpc_early < -30 ~ "Pre",                       # More than 30 days before earliest mCRPC treatment 
      TRUE ~ NA_character_),
    mcrpc_status_late = case_when(
      is.na(any_treatment_latest_date) ~ "Pre",                # No mCRPC treatment date available
      days_to_mcrpc_late >= -30 & days_to_mcrpc_late <= 30 ~ "Pre",  # Within 30 days before/after latest mCRPC treatment
      days_to_mcrpc_late > 30 ~ "Post",                       # More than 30 days after latest mCRPC treatment
      days_to_mcrpc_late < -30 ~ "Pre",                       # More than 30 days before latest mCRPC treatment 
      TRUE ~ NA_character_)
  )
```

Now do the same for mHSPC treatments

```{r}
# Cohort C: Find earliest and latest mHSPC treatment date
df_c2 <- df_c2 %>%
  mutate(
    # Convert mHSPC treatment dates to Date type
    mhspc1_date = as.Date(txt_mhspc_1st_date),
    mhspc2_date = as.Date(txt_mhspc_2nd_date)
  ) %>%
  rowwise() %>%
  mutate(
    # Find the earliest and latest date of mHSPC treatment
    mhspc_earliest_date = min(c(mhspc1_date, mhspc2_date), na.rm = TRUE),
    mhspc_latest_date   = max(c(mhspc1_date, mhspc2_date), na.rm = TRUE),
    # Was any mHSPC treatment given?
    mhspc_treatment_given = any(!is.na(c(mhspc1_date, mhspc2_date)))
  ) %>%
  ungroup() %>%
  mutate(
    # If no treatment was given, set dates to NA
    mhspc_earliest_date = if_else(mhspc_treatment_given, mhspc_earliest_date, as.Date(NA)),
    mhspc_latest_date   = if_else(mhspc_treatment_given, mhspc_latest_date, as.Date(NA))
  )
```

Calculate days from sample collection to mHSPC treatment

```{r}
df_c2 <- df_c2 %>%
  mutate(
    # Calculate days from sample collection to earliest and latest mHSPC treatment
    days_to_mhspc_early = as.numeric(difftime(date, mhspc_earliest_date, units = "days")),
    days_to_mhspc_late  = as.numeric(difftime(date, mhspc_latest_date, units = "days")),
    # Label as Pre if sample is within 30 days before/after earliest mHSPC treatment, Post if >30 days after
    mhspc_status_early = case_when(
      is.na(mhspc_earliest_date) ~ "Pre",                # No mHSPC treatment date available
      days_to_mhspc_early >= -30 & days_to_mhspc_early <= 30 ~ "Pre",  # Within 30 days before/after earliest mHSPC treatment
      days_to_mhspc_early > 30 ~ "Post",                       # More than 30 days after earliest mHSPC treatment
      days_to_mhspc_early < -30 ~ "Pre",                       # More than 30 days before earliest mHSPC treatment
      TRUE ~ NA_character_),
    mhspc_status_late = case_when(
      is.na(mhspc_latest_date) ~ "Pre",                # No mHSPC treatment date available
      days_to_mhspc_late >= -30 & days_to_mhspc_late <= 30 ~ "Pre",  # Within 30 days before/after latest mHSPC treatment
      days_to_mhspc_late > 30 ~ "Post",                       # More than 30 days after latest mHSPC treatment
      days_to_mhspc_late < -30 ~ "Pre",                       # More than 30 days before latest mHSPC treatment
      TRUE ~ NA_character_)
  )
```

Number of samples in pre- or post-

```{r}
table(df_c2$mcrpc_status_early)
```

Lets check how many pre- post- pairs we have

```{r}
mrn_pre_post <- df_c2 %>%
  group_by(mrn) %>%
  summarise(
    has_pre = as.integer(any(mcrpc_status_early == "Pre", na.rm = TRUE)),
    has_post = as.integer(any(mcrpc_status_early == "Post", na.rm = TRUE))
  )

# View only those with both Pre and Post
mrn_pre_post %>%
  filter(has_pre == 1 & has_post == 1) %>%
  dim()
```

### Simplify treatment types for Cohort C

Load the treatment simplification mapping and add simplified treatment column

```{r}
# Load treatment simplification mapping
treatment_map <- read_xlsx("/Users/zakiwilmot/Desktop/ZW_004_ProstateProteomics/data/updated_2024/clinical_data/mCRPC_treatment.xlsx")
colnames(treatment_map) <- c("original", "simplified")

# Create a named vector for easy lookup
treatment_lookup <- setNames(treatment_map$simplified, treatment_map$original)

# Add earliest mHSPC treatment type and simplified version to df_c2
df_c2 <- df_c2 %>%
  rowwise() %>%
  mutate(
    # Determine which mHSPC treatment is earliest based on earliest date
    earliest_mhspc_treatment_type = case_when(
      mhspc_earliest_date == mhspc1_date ~ txt_mhspc_1st_type,
      mhspc_earliest_date == mhspc2_date ~ txt_mhspc_2nd_type,
      TRUE ~ NA_character_
    ),
    # Map to simplified name using the lookup table
    earliest_mhspc_treatment_simplified = treatment_lookup[earliest_mhspc_treatment_type]
  ) %>%
  ungroup()
```

Check the simplified treatment types:

```{r}
# View the mHSPC mapping for all samples
df_c2 %>%
  select(mrn, HCI_cID, earliest_mhspc_treatment_type, earliest_mhspc_treatment_simplified, mhspc_status_early) %>%
  arrange(earliest_mhspc_treatment_simplified)
```

Check for unmapped mHSPC treatments:

```{r}
# Get all unique mHSPC treatment types from your data
all_mhspc_treatments <- df_c2 %>%
  select(txt_mhspc_1st_type, txt_mhspc_2nd_type) %>%
  pivot_longer(everything(), values_to = 'treatment') %>%
  filter(!is.na(treatment)) %>%
  distinct(treatment) %>%
  pull(treatment)

# Find which ones are NOT in the mapping
unmapped_mhspc <- setdiff(all_mhspc_treatments, treatment_map$original)
cat('Unmapped mHSPC treatments (need to be added to Excel file):\n')
print(sort(unmapped_mhspc))
cat('\nNumber of unmapped mHSPC treatment types:', length(unmapped_mhspc), '\n')
```


# Objective 3

*Viz of data*

Lets summarize the data with some plots

Create a master file to see the treatment status and collection date of all samples

```{r}
d1 <- df_a %>% 
  mutate(txt_stat = "Pre",
         time2txt = 0) 
d2 <- df_b3 %>% select(mrn, age, HCI_cID, date, cohort, updated_adt_early, days_to_adt_early) %>% 
  rename(txt_stat = updated_adt_early,
         time2txt= days_to_adt_early)
d3 <- df_c2 %>% select(mrn, age, HCI_cID, date, cohort, mcrpc_status_early, days_to_mcrpc_early) %>% 
  rename(txt_stat = mcrpc_status_early,
         time2txt = days_to_mcrpc_early)

df_meta_f <- rbind(d1,d2,d3) %>% 
  left_join(df_hci_id %>% select(-mrn, -X))
```

## Venn overlap

Lets get a Venn diagram of the number of samples in each batch

```{r}
un_a <- df_meta_f %>% filter(Psomagen == "Y") %>% pull(HCI_cID)
un_b <- df_meta_f %>% filter(Q_13356 == "Y") %>% pull(HCI_cID)
un_c <- df_meta_f %>% filter(Q_15806 == "Y") %>% pull(HCI_cID)

l_all <- list(
  Psomagen = un_a,
  Q_13356 = un_b,
  Q_15806 = un_c
)

overlap <- calculate.overlap(
x <- list("Psomagen"=un_a, 
          "Q_13356"=un_b,
          "Q_15806"=un_c))

pdf(file=file.path(wd$outCurr, "Samples_venn_overlap.pdf"), width = 5, height = 5)
ggvenn(
  x, 
  fill_color = pal.study,
  show_percentage = FALSE,
  stroke_size = 0.5, set_name_size = 4
)
dev.off()

un_a <- df_meta_f %>% filter(Psomagen == "Y") %>% pull(mrn) %>% unique()
un_b <- df_meta_f %>% filter(Q_13356 == "Y") %>% pull(mrn) %>% unique()
un_c <- df_meta_f %>% filter(Q_15806 == "Y") %>% pull(mrn) %>% unique()

l_all <- list(
  Psomagen = un_a,
  Q_13356 = un_b,
  Q_15806 = un_c
)

overlap <- calculate.overlap(
x <- list("Psomagen"=un_a, 
          "Q_13356"=un_b,
          "Q_15806"=un_c))

pdf(file=file.path(wd$outCurr, "Patient_venn_overlap.pdf"), width = 5, height = 5)
ggvenn(
  x, 
  fill_color = pal.study,
  show_percentage = FALSE,
  stroke_size = 0.5, set_name_size = 4
)
dev.off()
```

## Num samples

```{r}
# Step 1: Summarize the data (count samples for each cohort and txt_stat combination)
summary_df <- df_meta_f %>%
  mutate(txt_stat = gsub("[1-2]", "", txt_stat),
         txt_stat = gsub("post", "Post", txt_stat)) %>% 
  group_by(cohort, txt_stat) 

pr1.1 <- print(paste0("Num of patients in All is ", length(unique(summary_df$mrn))))
pr1.2 <- print(paste0("Num of samples in All is ", length(unique(summary_df$HCI_cID))))

summary_df1 <- summary_df %>% 
  summarise(count = n(), .groups = 'drop') %>% 
  # Create a new column that combines cohort and txt_stat
  mutate(cohort_txt_stat =
           paste(cohort, txt_stat, sep = "_"),
         cohort_txt_stat = factor(cohort_txt_stat, levels=c(
           "A_Pre", "A_Post", "B_Pre", "B_Post", "C_Pre", "C_Post"
         ))) %>% 
  mutate(Batch = "01_All_cohort")

# Custom colors: you can choose shades inspired by Tableau colors
custom_colors <- c(
  "A_Pre" = "#1f77b4", "A_Post" = "#aec7e8",  # Blue shades
  "B_Pre" = "#ff7f0e", "B_Post" = "#ffbb78",  # Orange shades
  "C_Pre" = "#2ca02c", "C_Post" = "#98df8a"  # Green shades
)

# Updated plot
p1 <- ggplot(summary_df1, aes(x = cohort_txt_stat, y = Batch, size = count, fill = cohort_txt_stat)) +
  geom_point(shape = 21, color = "black", position=position_dodge(width=0.5)) +  # Create circles with black border
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=0.5)) +  # Add text underneath circles
  scale_size_area(max_size = 15) +  # Adjust circle sizes
  scale_fill_manual(values = custom_colors) +  # Set custom colors
  labs(x = "Cohort", y = "", fill = "Cohort and txt_stat", size = "Sample Count") +  # Label x-axis and legend
  theme_bw() +  # Clean background
  theme(legend.position = "right", axis.ticks.y = element_blank(),
        panel.grid = element_blank())
ggsave(p1, file=file.path(wd$outCurr, "Treatment_breakdown.pdf"), height = 5, width = 7)
p1
```

## Longtidudinal 

Lets plot the distribution of samples. To see how many are longtiduinal samples. Ie - same mrn, profiled pre- and post- treatment. 

```{r}
# Identify patients groups ;
# Group 1 - mrn that occurs 1x
# Group 2 - mrn that occurs 2x
#   -2B - mrn 2x in cohort B 
#   -2C - mrn 2x in cohort C_D
#   -2D - mrn 1x in cohort B & 1x in cohort C_D
# Group 3 - mrn that occurs 3x

df_tally <- df_meta_f 

df_tally_tmp <- df_tally %>%
  select(mrn, HCI_cID, cohort) %>% distinct()

mrn_counts <- df_tally_tmp %>%
  group_by(mrn, cohort) %>%
  summarise(count = n(), .groups = 'drop') %>%
  pivot_wider(names_from = cohort, values_from = count, values_fill = 0)

tot_counts <- df_meta_f %>%
  group_by(mrn) %>% 
  summarise(total_observations = n())

mrn_counts <- mrn_counts %>% left_join(tot_counts)



group_1_mrns <- mrn_counts %>% filter(total_observations == 1) %>% pull(mrn)
group_2_mrns <- mrn_counts %>% filter(total_observations == 2)
group_2_cohort_B <- group_2_mrns %>% filter(B == 2) %>% pull(mrn)
group_2_cohort_C <- group_2_mrns %>% filter(C == 2) %>% pull(mrn)
group_2_mrns_set2 <- group_2_mrns %>% filter(B == 1 & B == 1) %>% pull(mrn)

# -- Group 2--- 
df_group2 <- df_meta_f %>%
  filter(mrn %in% group_2_cohort_B) %>% 
  select(mrn, HCI_cID, cohort, txt_stat) %>% arrange(mrn, cohort, txt_stat)


# Making a function, to avoid repetition
# Define a function that processes a cohort
process_cohort <- function(df_meta_f, cohort_mrns, prefix, start_group) {
  # Step 1: Filter and arrange
  df_cohort <- df_meta_f %>%
    filter(mrn %in% cohort_mrns) %>%
    select(mrn, HCI_cID, cohort, txt_stat) %>%
    arrange(mrn, cohort, txt_stat)
  
  # Step 2: Create the pattern for each mrn
  df_cohort <- df_cohort %>%
    group_by(mrn) %>%
    arrange(cohort, txt_stat) %>%
    mutate(pattern = paste(cohort, txt_stat, collapse = "-")) %>%
    ungroup()
  
  # Step 3: Assign groups based on unique patterns
  unique_patterns <- df_cohort %>%
    distinct(pattern) %>%
    mutate(group = paste0(prefix, LETTERS[start_group + row_number() - 1]))
  
  # Step 4: Merge group information back
  df_cohort <- df_cohort %>%
    left_join(unique_patterns, by = "pattern")
  
  # Return the dataframe and the last group number used
  df_cohort_tmp <- df_cohort %>% select(mrn, group) %>% distinct()
  last_group <- start_group + nrow(unique_patterns) - 1
  
  list(df_cohort_tmp = df_cohort_tmp, last_group = last_group)
}

# Process Group 2B (group_2_cohort_B)
result_B <- process_cohort(df_meta_f, group_2_cohort_B, "Group 2", 1)

# Process Group 2C starting from the next group after Group 2B
result_C <- process_cohort(df_meta_f, group_2_cohort_C, "Group 2", result_B$last_group + 1)

result_CD <- process_cohort(df_meta_f, group_2_mrns_set2, "Group 2", result_C$last_group + 1)

# Combine the results
df_group2_combined <- bind_rows(result_B$df_cohort_tmp, result_C$df_cohort_tmp, result_CD$df_cohort_tmp)


# ---- Group 3 ---- #
# Step 1: Get unique combinations of mrn, cohort, and txt_stat
group_4_mrns <- mrn_counts %>% filter(total_observations == 3) %>% pull(mrn)
df_group4 <- df_meta_f %>%
  filter(mrn %in% group_4_mrns) %>% 
  select(mrn, HCI_cID, cohort, txt_stat) %>% arrange(mrn, cohort, txt_stat)


# data_labeled <- df_group4 %>%
#   group_by(mrn, txt_stat) %>%
#   mutate(txt_stat = ifelse(row_number() == 1, txt_stat, paste0(txt_stat, row_number()))) %>%
#   ungroup()


# Step 1: Create a unique pattern by combining cohort and txt_stat for each mrn
# This creates a pattern for each mrn's cohort-txt_stat combination
df_group4 <- df_group4 %>%
  group_by(mrn) %>%
  arrange(cohort, txt_stat) %>%  # Ensure consistent order
  mutate(pattern = paste(cohort, txt_stat, collapse = "-")) %>%
  ungroup()

# Step 2: Assign groups based on unique patterns
# We create a lookup table where each unique pattern is assigned a group
unique_patterns <- df_group4 %>%
  distinct(pattern) %>%
  mutate(group = paste0("Group 3", LETTERS[row_number()]))


# Step 3: Merge the group information back into the original data
df_group4 <- df_group4 %>%
  left_join(unique_patterns, by = "pattern")
df_group4_tmp <- df_group4 %>% select(mrn, group) %>% distinct()

dat_group <- data.frame(
  mrn = c(group_1_mrns, 
          df_group2_combined$mrn,
          df_group4_tmp$mrn),
  group = c(
    rep("Group 1", each=length(group_1_mrns)),
    df_group2_combined$group,
    df_group4_tmp$group)
)

data <- df_meta_f %>% 
  #select(mrn, HCI_cID, cohort, txt_stat) %>% distinct() %>% 
  left_join(dat_group)

# Label consequtive pre or post
data <- data %>%
  group_by(mrn, txt_stat) %>%
  mutate(txt_stat = ifelse(row_number() == 1, txt_stat, paste0(txt_stat, row_number()))) %>%
  ungroup() %>% 
  mutate(txt_stat = factor(txt_stat, levels=c("Pre", "Pre2", "Post", "Post2", "Post3")))

df_pre_post <- data
df_pre_post_tmp <- df_pre_post %>%   
  dplyr::rename(SampleID = sample_id_psom) %>% 
  select(cohort, SampleID, txt_stat) %>% 
  mutate(SampleID = as.character(SampleID)) %>% 
  distinct()

data2 <- data %>% dplyr::group_by(group, cohort, txt_stat) %>% 
  arrange(group) %>% 
  dplyr::summarize(count = n()) %>% 
  mutate(group = factor(group, levels = (unique(group))))


ggplot(data2, aes(x=cohort, y=group, size=count, fill=txt_stat)) +
  geom_point(shape = 21, color = "black", position=position_dodge(width=0.5)) +  
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=0.5)) +  # Add text underneath circles
  theme_bw() +  # Clean background
  theme(legend.position = "right", panel.grid = element_blank()) +
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=0.5)) +  # Add text underneath circlesNULL
  NULL

# Custom colors
# Create a new column that combines cohort and txt_stat
data2 <- data2 %>% 
  mutate(cohort_txt_stat =paste(cohort, txt_stat, sep = "_"),
         cohort_txt_stat = factor(cohort_txt_stat, levels=unique(cohort_txt_stat))) 

# Custom colors: you can choose shades inspired by Tableau colors
custom_colors <- c(
  "A_Pre" = "#1f77b4", "A_Post" = "#aec7e8",  # Blue shades
  "B_Pre" = "#ff7f0e", "B_Post" = "#ffbb78", "B_Post2" = "#ffbb78", # Orange shades
  "C_Pre" = "#2ca02c", "C_Pre2" = "#2ca02c", 
  "C_Post" = "#98df8a",  "C_Post2" = "#98df8a", "C_Post3" = "#98df8a" # Green shades
)

ggplot(data2, aes(x=cohort, y=group, size=count, fill=cohort_txt_stat)) +
  geom_point(shape = 21, color = "black", position=position_dodge(width=0.5)) +  
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=0.5)) +  # Add text underneath circles
  theme_bw() +  # Clean background
  scale_fill_manual(values = custom_colors) +  # Set custom colors 
  theme(legend.position = "right", panel.grid = element_blank()) +
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=0.5)) +  # Add text underneath circlesNULL
  NULL

p2 <- 
ggplot(data2, aes(x=cohort, y=group, size=count, fill=txt_stat)) +
  geom_point(shape = 21, color = "black", position=position_dodge(width=0.5)) +  
  scale_size_area(max_size = 15) +  # Adjust circle sizes
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=0.5)) +  # Add text underneath circles
  theme_bw() +  # Clean background
  scale_fill_manual(values = c("Pre" = "lightblue", "Pre2" = "lightblue", 
                               "Post" = "orange", "Post2"= "orange")) +  # Set fill colors
  theme(legend.position = "right", panel.grid = element_blank()) +
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=0.5)) +  # Add text underneath circlesNULL
  NULL


ggsave(p2, file=file.path(wd$outCurr, "Treatment_grouping_pattern_updated.pdf"), height = 5, width = 12)

write.csv(data, file.path(wd$outCurr, "Treatment_grouping_pattern_updated.csv"))

df_pre_post_ori <- df_pre_post
```

In the local stage, lets group all together. So the plots are on the same scale. Also add the total number of samples

```{r}
# Sum of A (Pre + Post)
sum_A <- data2 %>%
  filter(cohort == "A") %>%
  summarise(total = sum(count)) %>%
  pull(total)

# Create Group 0
# First, summarize "pre/post" status for each cohort
group_0 <- data2 %>%
  mutate(
    txt_stat_grouped = case_when(
      str_detect(txt_stat, regex("pre", ignore_case = TRUE)) ~ "Pre",
      str_detect(txt_stat, regex("post", ignore_case = TRUE)) ~ "Post",
      TRUE ~ as.character(txt_stat)
    )
  ) %>%
  group_by(cohort, txt_stat_grouped) %>%
  summarise(count = sum(count), .groups = "drop") %>%
  
  # Special step: handle A separately
  mutate(
    count = if_else(cohort == "A",
                    sum(count[cohort == "A"]),
                    count)
  ) %>%
  
  mutate(
    group = "Group 01",
    cohort_txt_stat = paste0(cohort, "_", txt_stat_grouped)
  ) %>%
  select(group, cohort, txt_stat = txt_stat_grouped, count, cohort_txt_stat)



# Bind Group 1 and Group 0 together
df_new <- bind_rows(data2, group_0) %>% ungroup %>% distinct()

# Extend view
df_new <- df_new %>%
  bind_rows(tibble(
    group = "Group 00",
    cohort = "A",
    txt_stat = "Pre",
    count = NA,
    cohort_txt_stat = "A_Pre"
  )) %>% mutate(
    txt_stat = factor(txt_stat, levels=c("Pre", "Pre2", "Post", "Post2")),
    cohort_txt_stat = factor(cohort_txt_stat, 
                             levels=c("A_Pre", "A_Post",
                                      "B_Pre", "B_Post", "B_Post2",
                                      "C_Pre", "C_Pre2", "C_Post", "C_Post2, C_Post3"
                                      ))
  )

tmp_width <- 0
p2 <- 
ggplot(df_new, aes(x=cohort_txt_stat, y=group, size=count, fill=cohort_txt_stat)) +
  geom_point(shape = 21, color = "black", position=position_dodge(width=tmp_width)) +  
  scale_size_area(max_size = 15) +  # Adjust circle sizes
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=tmp_width)) +  # Add text underneath circles
  theme_bw() +  # Clean background
  scale_fill_manual(
    values = c("A_Pre" = "lightblue", "A_Pre2" = "lightblue", 
               "B_Pre" = "lightblue", "B_Pre2" = "lightblue", 
               "C_Pre" = "lightblue", "C_Pre2" = "lightblue", 
               "A_Post" = "orange", "A_Post2"= "orange",
               "B_Post" = "orange", "B_Post2"= "orange",
               "C_Post" = "orange", "C_Post2"= "orange", "C_Post"= "orange")
    ) +  # Set fill colors
  theme(legend.position = "right", panel.grid = element_blank()) +
  geom_text(aes(label = count), vjust = 2, size = 5, position=position_dodge(width=tmp_width)) +  # Add text underneath circlesNULL
  NULL


ggsave(p2, file=file.path(wd$outCurr, "Treatment_grouping_pattern_all.pdf"), height = 5, width = 12)
```


```{r}
# Number of cohort B unique patients
df_meta_f %>% filter(cohort == "B") %>% pull(mrn) %>% unique() %>% length()

# Number of cohort C unique patients
df_meta_f %>% filter(cohort == "C") %>% pull(mrn) %>% unique() %>% length()

# Number of cohort C samples
df_meta_f %>% filter(cohort == "C") %>% nrow()
```

Make a simple barplot to summarize the data. Two columns per cohort. Number of patients & number of samples. 

```{r}
# Calculate overlap between B and C_D cohorts
overlap_patients <- df_meta_f %>%
  filter(cohort %in% c("B", "C")) %>%
  group_by(mrn) %>%
  filter(n_distinct(cohort) > 1) %>%
  pull(mrn) %>%
  unique()

# Create summary data frame
summary_data <- df_meta_f %>%
  group_by(cohort) %>%
  summarise(
    n_patients = n_distinct(mrn),
    n_samples = n()
  ) %>%
  pivot_longer(
    cols = c(n_patients, n_samples),
    names_to = "metric",
    values_to = "count"
  )

# Create the barplot
p_bar <- ggplot(summary_data, aes(x = cohort, y = count, fill = metric)) +
  geom_bar(stat = "identity", position = position_dodge(width = 0.7), width = 0.6) +
  # Add overlay for C_D cohort showing overlap (only for patients)
  geom_bar(
    data = filter(summary_data, cohort == "C" & metric == "n_patients"),
    aes(y = length(overlap_patients), fill = "Overlap with B"),
    stat = "identity",
    position = position_dodge(width = 0.7),
    width = 0.6,
    alpha = 0.5
  ) +
  scale_fill_manual(
    values = c(
      "n_patients" = "#faedcd",
      "n_samples" = "grey60",
      "Overlap with B" = "darkred"
    ),
    labels = c("Number of Patients", "Number of Samples", "Overlap with B")
  ) +
  labs(
    x = "Cohort",
    y = "Count",
    fill = "Metric"
  ) +
  theme_bw() +
  theme(
    legend.position = "right",
    panel.grid.major.x = element_blank(),
    panel.grid.minor = element_blank()
  ) +
  geom_text(
    aes(label = count),
    position = position_dodge(width = 0.7),
    vjust = -0.5,
    size = 4
  ) +
  # Add annotation for overlap
  annotate(
    "text",
    x = "C_D",
    y = length(overlap_patients),
    label = paste("Overlap:", length(overlap_patients)),
    vjust = -0.5,
    size = 4,
    color = "darkred"
  )

ggsave(p_bar, file=file.path(wd$outCurr, "Patients_number_bar.pdf"), height = 5, width = 7)
```

## Swimmers plot 

Make a timeline for each patient

```{r edited_swimmer, message=FALSE, warning=FALSE}
# 1. Define the pattern assignment function
# In the functions.R script

# 2. Define the pattern order
# In the functions.R script

# 3. Relabel txt_stat as Pre1, Pre2, Post1, etc. by date within each patient/cohort/status
df_meta_f2 <- df_meta_f %>%
  group_by(mrn, cohort, txt_stat) %>%
  arrange(date, .by_group = TRUE) %>%
  mutate(
    txt_stat_ordered = paste0(txt_stat, row_number())
  ) %>%
  ungroup()

# 4. Assign pattern to each sample
patient_patterns <- get_patient_pattern(df_meta_f2 %>% mutate(txt_stat = txt_stat_ordered))
df_meta_f2 <- df_meta_f2 %>%
  left_join(patient_patterns %>% select(mrn, pattern, total_samples), by = "mrn")

# MRN-specific debug filters removed to avoid exposing identifiers

# 5. Order patients by pattern and total_samples
ordered_mrns <- df_meta_f2 %>%
  distinct(mrn, pattern, total_samples) %>%
  arrange(
    factor(pattern, levels = pattern_order),
    desc(total_samples)
  ) %>%
  pull(mrn)

# 6. Prepare swimmers_data with correct offsets and positions
swimmers_data <- df_meta_f2 %>%
  mutate(
    base_pos = case_when(
      cohort == "A" ~ 0,
      cohort == "B" ~ 1,
      cohort == "C" ~ 2
    ),
    # Extract trailing number from txt_stat_ordered, default to 1 if missing
    sample_num = as.numeric(str_extract(txt_stat_ordered, "\\d+$")),
    sample_num = ifelse(is.na(sample_num), 1, sample_num),
    # Offset for multiple samples of the same type
    x_offset = (sample_num - 1) * 0.2,
    # Treatment offset: Pre left, Post right
    treatment_offset = case_when(
      grepl("^Pre", txt_stat_ordered) ~ -0.2,
      grepl("^Post", txt_stat_ordered) ~ 0.2,
      TRUE ~ 0
    ),
    # Final x position, shifted to avoid 1/2 dashed line
    x_pos = base_pos + treatment_offset + x_offset - 0.05,
    # y position based on patient order
    y_pos = match(mrn, ordered_mrns),
    # For labeling
    sample_label = txt_stat_ordered
  )

# 7. Rectangle data for cohort boxes
rect_data <- data.frame(
  cohort = c("A", "B", "C"),
  xmin = c(-0.5, 0.5, 1.5),
  xmax = c(0.5, 1.5, 2.5)
)

# 8. Swimmers plot with time2txt above each point
p_swimmers_ordered <- ggplot(swimmers_data, aes(x = x_pos, y = y_pos)) +
  geom_rect(
    data = rect_data,
    inherit.aes = FALSE,
    aes(xmin = xmin, xmax = xmax, ymin = 0, ymax = max(swimmers_data$y_pos) + 1),
    fill = "grey95",
    alpha = 0.5
  ) +
  geom_vline(
    xintercept = c(0, 1, 2),
    linetype = "dashed",
    color = "grey50",
    alpha = 0.5
  ) +
  geom_point(
    aes(color = txt_stat_ordered),
    size = 3,
    shape = 16
  ) +
  geom_text(
    aes(label = time2txt),
    vjust = -1,
    size = 3
  ) +
  geom_line(
    aes(group = mrn),
    color = "grey50",
    alpha = 0.5
  ) +
  scale_x_continuous(
    breaks = 0:2,
    labels = c("Cohort A", "Cohort B", "Cohort C/D"),
    limits = c(-0.5, 2.5)
  ) +
  scale_color_manual(
    values = c(
      "Pre1" = "steelblue", "Pre2" = "steelblue", "Pre3" = "red",
      "Post1" = "orange", "Post2" = "orange", "Post3" = "orange"
    )
    #labels = c("Pre1", "Pre2", "Pre3", "Post1", "Post2", "Post3")
  ) +
  labs(
    x = "",
    y = "Patient",
    color = "Sample Type"
  ) +
  theme_bw() +
  theme(
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    legend.position = "right"
  )

print(p_swimmers_ordered)

# 9. Save the plot
ggsave(
  p_swimmers_ordered,
  file = file.path(wd$outCurr, "Sample_collection_timeline_ordered.pdf"),
  height = max(10, length(unique(swimmers_data$mrn)) * 0.3),
  width = 8,
  limitsize = FALSE
)
write.csv(swimmers_data, file.path(wd$outCurr, "SwimmersData.csv"))


# 11. Zommed section
p_swimmers_ordered3 <- ggplot(
  swimmers_data %>% filter(!pattern %in% c("A", "B_Pre_Only", "B_Post_Only",
                                           "C_Pre_Only", "C_Post_Only")), 
  aes(x = x_pos, y = y_pos)) +
  geom_vline(
    xintercept = c(0, 1, 2),
    linetype = "dashed",
    color = "grey50",
    alpha = 0.5
  ) +
  geom_point(
    aes(color = txt_stat_ordered),
    size = 3,
    shape = 16
  ) +
  geom_line(
    aes(group = mrn),
    color = "grey50",
    alpha = 0.5
  ) +
  geom_text(
    aes(label = time2txt),
    vjust = -1,
    size = 3
  ) +
  scale_x_continuous(
    breaks = 0:2,
    labels = c("Cohort A", "Cohort B", "Cohort C/D"),
    limits = c(-0.5, 2.5)
  ) +
  scale_color_manual(
    values = c(
      "Pre1" = "steelblue", "Pre2" = "steelblue", "Pre3" = "red",
      "Post1" = "orange", "Post2" = "orange", "Post3" = "orange"
    )
    #labels = c("Pre1", "Pre2", "Pre3", "Post1", "Post2", "Post3")
  ) +
  labs(
    x = "",
    y = "Patient",
    color = "Sample Type"
  ) +
  theme_bw() +
  theme(
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    legend.position = "right"
  )

ggsave(
  p_swimmers_ordered3,
  file = file.path(wd$outCurr, "Sample_collection_timeline_ordered_noTime_Zommed.pdf"),
  height = 15,
  width = 8,
  limitsize = FALSE
)

# 10. Save the plot without the time2txt text
p_swimmers_ordered <- ggplot(swimmers_data, aes(x = x_pos, y = y_pos)) +
  geom_rect(
    data = rect_data,
    inherit.aes = FALSE,
    aes(xmin = xmin, xmax = xmax, ymin = 0, ymax = max(swimmers_data$y_pos) + 1),
    fill = "grey95",
    alpha = 0.5
  ) +
  geom_vline(
    xintercept = c(0, 1, 2),
    linetype = "dashed",
    color = "grey50",
    alpha = 0.5
  ) +
  geom_point(
    aes(color = txt_stat_ordered),
    size = 3,
    shape = 16
  ) +
  geom_line(
    aes(group = mrn),
    color = "grey50",
    alpha = 0.5
  ) +
  scale_x_continuous(
    breaks = 0:2,
    labels = c("Cohort A", "Cohort B", "Cohort C/D"),
    limits = c(-0.5, 2.5)
  ) +
  scale_color_manual(
    values = c(
      "Pre1" = "steelblue", "Pre2" = "steelblue", "Pre3" = "red",
      "Post1" = "orange", "Post2" = "orange", "Post3" = "orange"
    )
    #labels = c("Pre1", "Pre2", "Pre3", "Post1", "Post2", "Post3")
  ) +
  labs(
    x = "",
    y = "Patient",
    color = "Sample Type"
  ) +
  theme_bw() +
  theme(
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    legend.position = "right"
  )

print(p_swimmers_ordered)

ggsave(
  p_swimmers_ordered,
  file = file.path(wd$outCurr, "Sample_collection_timeline_ordered_noTime.pdf"),
  height = max(10, length(unique(swimmers_data$mrn)) * 0.3),
  width = 8,
  limitsize = FALSE
)
```


## Clean up 

In some cases, we want to remove samples with 2x Pre or 2x Post. These samples are not helpful for our analysis. 

As there were many Post1 and Post2 in mCRPC. We will retain these.

Lets select the sample to keep

```{r}
mrn_with_2x <- df_meta_f2 %>% filter(txt_stat_ordered %in% c("Pre2", "Post2")) %>% pull(mrn) %>% unique()
# mrn to keep (Post1 and Post2 in c)
mrn_to_keep <- df_meta_f2 %>% filter(pattern %in% c("C_Post1_to_C_Post2")) %>% pull(mrn) %>% unique()

mrn_to_examine <- mrn_with_2x[!mrn_with_2x %in% mrn_to_keep]
```

For these mrns, retain the sample closest to treatment date. 

```{r}
df_tmp <- df_meta_f2 %>% filter(mrn %in% mrn_to_examine)
```

Lets select samples closest to treatment date. If "Pre" select earliest pre- ad "Post" select earliest post-

```{r}
# For each mrn and txt_stat, identify the sample to keep (closest to treatment date)
# - For "Pre", keep the sample with the maximum time2txt <= 0 (i.e., closest before treatment)
# - For "Post", keep the sample with the minimum time2txt >= 0 (i.e., closest after treatment)
# Then, get the rows to EXCLUDE (i.e., not the closest)

df_tmp_keep <- df_tmp %>%
  group_by(mrn, txt_stat) %>%
  mutate(
    keep_flag = case_when(
      txt_stat %in% c("Pre", "Pre1", "Pre2", "Pre3") ~ {
        # If any time2txt <= 0, keep the max of those; else, keep the min overall
        if (any(time2txt <= 0, na.rm = TRUE)) {
          time2txt == max(time2txt[time2txt <= 0], na.rm = TRUE)
        } else {
          time2txt == min(time2txt, na.rm = TRUE)
        }
      },
      txt_stat %in% c("Post", "Post1", "Post2", "Post3") ~ {
        # If any time2txt >= 0, keep the min of those; else, keep the max overall
        if (any(time2txt >= 0, na.rm = TRUE)) {
          time2txt == min(time2txt[time2txt >= 0], na.rm = TRUE)
        } else {
          time2txt == max(time2txt, na.rm = TRUE)
        }
      },
      TRUE ~ TRUE
    )
  ) %>%
  ungroup()

# Rows to EXCLUDE (i.e., not the closest to treatment date)
df_tmp_exclude <- df_tmp_keep %>%
  filter(!keep_flag)
```

Create a a df_meta_f for those we want to retain

```{r}
df_meta_f3 <- df_meta_f %>% filter(!HCI_cID %in% df_tmp_exclude$HCI_cID)
```

## Swimmers v2

Recreate the swimmers plot

```{r}
# 3. Relabel txt_stat as Pre1, Pre2, Post1, etc. by date within each patient/cohort/status
df_meta_f4 <- df_meta_f3 %>%
  group_by(mrn, cohort, txt_stat) %>%
  arrange(date, .by_group = TRUE) %>%
  mutate(
    txt_stat_ordered = paste0(txt_stat, row_number())
  ) %>%
  ungroup()

# 4. Assign pattern to each sample
patient_patterns <- get_patient_pattern(df_meta_f4 %>% mutate(txt_stat = txt_stat_ordered))
df_meta_f4 <- df_meta_f4 %>%
  left_join(patient_patterns %>% select(mrn, pattern, total_samples), by = "mrn")

# MRN-specific debug filters removed to avoid exposing identifiers

# 5. Order patients by pattern and total_samples
ordered_mrns <- df_meta_f4 %>%
  distinct(mrn, pattern, total_samples) %>%
  arrange(
    factor(pattern, levels = pattern_order),
    desc(total_samples)
  ) %>%
  pull(mrn)

# 6. Prepare swimmers_data with correct offsets and positions
swimmers_data <- df_meta_f4 %>%
  mutate(
    base_pos = case_when(
      cohort == "A" ~ 0,
      cohort == "B" ~ 1,
      cohort == "C" ~ 2
    ),
    # Extract trailing number from txt_stat_ordered, default to 1 if missing
    sample_num = as.numeric(str_extract(txt_stat_ordered, "\\d+$")),
    sample_num = ifelse(is.na(sample_num), 1, sample_num),
    # Offset for multiple samples of the same type
    x_offset = (sample_num - 1) * 0.2,
    # Treatment offset: Pre left, Post right
    treatment_offset = case_when(
      grepl("^Pre", txt_stat_ordered) ~ -0.2,
      grepl("^Post", txt_stat_ordered) ~ 0.2,
      TRUE ~ 0
    ),
    # Final x position, shifted to avoid 1/2 dashed line
    x_pos = base_pos + treatment_offset + x_offset - 0.05,
    # y position based on patient order
    y_pos = match(mrn, ordered_mrns),
    # For labeling
    sample_label = txt_stat_ordered
  )

# 7. Rectangle data for cohort boxes
rect_data <- data.frame(
  cohort = c("A", "B", "C"),
  xmin = c(-0.5, 0.5, 1.5),
  xmax = c(0.5, 1.5, 2.5)
)

write.csv(swimmers_data, file.path(wd$outCurr, "SwimmersData_cleaned.csv"))

# 10. All samples
p_swimmers_ordered4 <- ggplot(
  swimmers_data,
  aes(x = x_pos, y = y_pos)) +
  geom_vline(
    xintercept = c(0, 1, 2),
    linetype = "dashed",
    color = "grey50",
    alpha = 0.5
  ) +
  geom_point(
    aes(color = txt_stat_ordered),
    size = 3,
    shape = 16
  ) +
  geom_line(
    aes(group = mrn),
    color = "grey50",
    alpha = 0.5
  ) +
  # geom_text(
  #   aes(label = time2txt),
  #   vjust = -1,
  #   size = 3
  #) +
  scale_x_continuous(
    breaks = 0:2,
    labels = c("Cohort A", "Cohort B", "Cohort C/D"),
    limits = c(-0.5, 2.5)
  ) +
  scale_color_manual(
    values = c(
      "Pre1" = "steelblue", "Pre2" = "steelblue", "Pre3" = "red",
      "Post1" = "orange", "Post2" = "orange", "Post3" = "orange"
    )
    #labels = c("Pre1", "Pre2", "Pre3", "Post1", "Post2", "Post3")
  ) +
  labs(
    x = "",
    y = "Patient",
    color = "Sample Type"
  ) +
  theme_bw() +
  theme(
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    legend.position = "right"
  )

ggsave(
  p_swimmers_ordered4,
  file = file.path(wd$outCurr, "Sample_collection_timeline_ordered_noTime_cleaned.pdf"),
  height = 15,
  width = 8,
  limitsize = FALSE
)


# 11. Zommed section
p_swimmers_ordered4 <- ggplot(
  swimmers_data %>% filter(!pattern %in% c("A", "B_Pre_Only", "B_Post_Only",
                                           "C_Pre_Only", "C_Post_Only")), 
  aes(x = x_pos, y = y_pos)) +
  geom_vline(
    xintercept = c(0, 1, 2),
    linetype = "dashed",
    color = "grey50",
    alpha = 0.5
  ) +
  geom_point(
    aes(color = txt_stat_ordered),
    size = 3,
    shape = 16
  ) +
  geom_line(
    aes(group = mrn),
    color = "grey50",
    alpha = 0.5
  ) +
  # geom_text(
  #   aes(label = time2txt),
  #   vjust = -1,
  #   size = 3
  #) +
  scale_x_continuous(
    breaks = 0:2,
    labels = c("Cohort A", "Cohort B", "Cohort C/D"),
    limits = c(-0.5, 2.5)
  ) +
  scale_color_manual(
    values = c(
      "Pre1" = "steelblue", "Pre2" = "steelblue", "Pre3" = "red",
      "Post1" = "orange", "Post2" = "orange", "Post3" = "orange"
    )
    #labels = c("Pre1", "Pre2", "Pre3", "Post1", "Post2", "Post3")
  ) +
  labs(
    x = "",
    y = "Patient",
    color = "Sample Type"
  ) +
  theme_bw() +
  theme(
    panel.grid.major.y = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    legend.position = "right"
  )

ggsave(
  p_swimmers_ordered4,
  file = file.path(wd$outCurr, "Sample_collection_timeline_ordered_noTime_Zommed_cleaned.pdf"),
  height = 15,
  width = 8,
  limitsize = FALSE
)
```

## Bar plot v2

Make a simple barplot to summarize the data. Two columns per cohort. Number of patients & number of samples. 

```{r}
# Calculate overlap between B and C_D cohorts
overlap_patients <- df_meta_f4 %>%
  filter(cohort %in% c("B", "C")) %>%
  group_by(mrn) %>%
  filter(n_distinct(cohort) > 1) %>%
  pull(mrn) %>%
  unique()

# Create summary data frame
summary_data <- df_meta_f4 %>%
  group_by(cohort) %>%
  summarise(
    n_patients = n_distinct(mrn),
    n_samples = n()
  ) %>%
  pivot_longer(
    cols = c(n_patients, n_samples),
    names_to = "metric",
    values_to = "count"
  )

# Create the barplot
p_bar <- ggplot(summary_data, aes(x = cohort, y = count, fill = metric)) +
  geom_bar(stat = "identity", position = position_dodge(width = 0.7), width = 0.6) +
  # Add overlay for C_D cohort showing overlap (only for patients)
  geom_bar(
    data = filter(summary_data, cohort == "C" & metric == "n_patients"),
    aes(y = length(overlap_patients), fill = "Overlap with B"),
    stat = "identity",
    position = position_dodge(width = 0.7),
    width = 0.6,
    alpha = 0.5
  ) +
  scale_fill_manual(
    values = c(
      "n_patients" = "#faedcd",
      "n_samples" = "grey60",
      "Overlap with B" = "darkred"
    ),
    labels = c("Number of Patients", "Number of Samples", "Overlap with B")
  ) +
  labs(
    x = "Cohort",
    y = "Count",
    fill = "Metric"
  ) +
  theme_bw() +
  theme(
    legend.position = "right",
    panel.grid.major.x = element_blank(),
    panel.grid.minor = element_blank()
  ) +
  geom_text(
    aes(label = count),
    position = position_dodge(width = 0.7),
    vjust = -0.5,
    size = 4
  ) +
  # Add annotation for overlap
  annotate(
    "text",
    x = "C_D",
    y = length(overlap_patients),
    label = paste("Overlap:", length(overlap_patients)),
    vjust = -0.5,
    size = 4,
    color = "darkred"
  )

ggsave(p_bar, file=file.path(wd$outCurr, "Patients_number_bar_clean.pdf"), height = 5, width = 7)
```

## Swimmer + bar

We used Claude to fix some of the plot

```{r}
library(ggplot2)
library(dplyr)
library(tidyr)
library(patchwork)  # For combining plots


# Define pattern order with original names
pattern_order_original <- c(
  "A",
  "B_Pre_Only", "B_Post_Only",
  "C_Pre_Only", "C_Post_Only",
  "B_Serial_PrePost",
  "C_Serial_PrePost",
  "B_Pre_to_C_Pre",
  "B_Post_to_C_Post",
  "C_Post1_to_C_Post2",
  "B_Post_to_C_Pre",
  "C_Pre1_to_C_Pre2",
  "B_Pre1_to_B_Pre2",
  "B_Post_to_C_Pre_to_C_Post"
)

# Create mapping for better labels
pattern_labels <- c(
  "A" = "Localized\nBaseline only",
  
  # Single timepoint patterns
  "B_Pre_Only" = "mHSPC\nPre-treatment only",
  "B_Post_Only" = "mHSPC\nPost-treatment only",
  "C_Pre_Only" = "mCRPC\nPre-treatment only",
  "C_Post_Only" = "mCRPC\nPost-treatment only",
  
  # Serial within same disease state
  "B_Serial_PrePost" = "mHSPC\nPre → Post treatment",
  "C_Serial_PrePost" = "mCRPC\nPre → Post treatment",
  
  # Cross-disease state patterns (Pre-treatment)
  "B_Pre_to_C_Pre" = "Pre-treatment\nmHSPC → mCRPC",
  
  # Cross-disease state patterns (Post-treatment)
  "B_Post_to_C_Post" = "Post-treatment\nmHSPC → mCRPC",
  
  # Multiple timepoints within disease state
  "C_Post1_to_C_Post2" = "mCRPC\nPost → Post (serial)",
  "C_Pre1_to_C_Pre2" = "mCRPC\nPre → Pre (serial)",
  "B_Pre1_to_B_Pre2" = "mHSPC\nPre → Pre (serial)",
  
  # Complex cross-disease patterns
  "B_Post_to_C_Pre" = "mHSPC Post →\nmCRPC Pre",
  "B_Post_to_C_Pre_to_C_Post" = "mHSPC Post →\nmCRPC Pre → Post"
)

# Add any missing patterns with generic labels
for (pattern in unique(swimmers_data$pattern)) {
  if (!pattern %in% names(pattern_labels)) {
    pattern_labels[pattern] = pattern
  }
}

# Update cohort labels in the data
swimmers_data <- swimmers_data %>%
  mutate(
    cohort_display = case_when(
      cohort == "A" ~ "Localized",
      cohort == "B" ~ "mHSPC",
      cohort %in% c("C", "D") ~ "mCRPC",
      TRUE ~ cohort
    )
  )

# Create pattern factor with desired order
swimmers_data$pattern <- factor(swimmers_data$pattern, levels = pattern_order_original)

# Count patients per pattern for spacing
pattern_counts <- swimmers_data %>%
  group_by(pattern) %>%
  summarise(n_patients = n_distinct(mrn)) %>%
  arrange(pattern)

# INCREASED SPACING for better readability
pattern_spacing <- 3  # Increased from 1.5 to 3 for more separation

# Create new y-positions grouped by pattern with more spacing
swimmers_data <- swimmers_data %>%
  arrange(pattern, mrn) %>%
  group_by(pattern) %>%
  mutate(
    patient_within_pattern = as.numeric(factor(mrn)),
    pattern_num = as.numeric(pattern)
  ) %>%
  ungroup()

# Recalculate grouped_y_pos with increased spacing
pattern_positions <- pattern_counts %>%
  mutate(
    cumulative_patients = cumsum(n_patients),
    pattern_y_start = lag(cumulative_patients, default = 0) + (row_number() - 1) * pattern_spacing,
    pattern_y_end = pattern_y_start + n_patients - 1,
    pattern_label = pattern_labels[as.character(pattern)]
  )

swimmers_data <- swimmers_data %>%
  left_join(pattern_positions %>% select(pattern, pattern_y_start), by = "pattern") %>%
  mutate(grouped_y_pos = pattern_y_start + patient_within_pattern - 1)

# Calculate sample counts for bar graph
sample_counts <- swimmers_data %>%
  mutate(
    sample_type = case_when(
      txt_stat_ordered %in% c("Pre1", "Pre2", "Pre3") ~ "Pre",
      txt_stat_ordered %in% c("Post1", "Post2", "Post3") ~ "Post"
    )
  ) %>%
  group_by(cohort_display, sample_type) %>%
  summarise(n_samples = n(), .groups = "drop") %>%
  mutate(
    x_pos = case_when(
      cohort_display == "Localized" & sample_type == "Pre" ~ -0.2,
      cohort_display == "Localized" & sample_type == "Post" ~ 0.2,
      cohort_display == "mHSPC" & sample_type == "Pre" ~ 0.8,
      cohort_display == "mHSPC" & sample_type == "Post" ~ 1.2,
      cohort_display == "mCRPC" & sample_type == "Pre" ~ 1.8,
      cohort_display == "mCRPC" & sample_type == "Post" ~ 2.2
    )
  )

# Create the bar plot with fixed y-axis range
p_bars <- ggplot(sample_counts, aes(x = x_pos, y = n_samples)) +
  geom_col(aes(fill = sample_type), width = 0.3, alpha = 0.7) +
  geom_text(aes(label = n_samples), vjust = -0.5, size = 3.5, fontface = "bold") +
  scale_fill_manual(
    values = c("Pre" = "steelblue", "Post" = "orange"),
    name = "Sample Type",
    labels = c("Pre" = "Pre-treatment", "Post" = "Post-treatment")
  ) +
  scale_x_continuous(
    breaks = c(0, 1, 2),
    labels = c("", "", ""),  # Will add labels in main plot
    limits = c(-0.5, 3.5)
  ) +
  scale_y_continuous(
    limits = c(0, 100),  # Fixed range 0-100
    breaks = seq(0, 100, by = 20),
    expand = c(0, 0)
  ) +
  labs(
    x = "",
    y = "Number of Samples",
    title = "Sample Distribution by Disease State and Treatment Status"
  ) +
  theme_bw() +
  theme(
    panel.grid.major.x = element_blank(),
    panel.grid.minor.x = element_blank(),
    panel.grid.major.y = element_line(color = "grey90"),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    legend.position = "top",
    legend.box = "horizontal",
    plot.margin = margin(t = 5, r = 5, b = 0, l = 5),
    plot.title = element_text(size = 10, face = "italic")
  )

# Create the main swimmer plot with better separation
p_grouped <- ggplot(swimmers_data, aes(x = x_pos, y = grouped_y_pos)) +
  # Add alternating background rectangles for pattern groups
  geom_rect(
    data = pattern_positions[seq(1, nrow(pattern_positions), by = 2), ],
    aes(xmin = -0.5, xmax = 2.5, 
        ymin = pattern_y_start - 0.5, 
        ymax = pattern_y_end + 0.5),
    fill = "grey95",
    alpha = 0.5,
    inherit.aes = FALSE
  ) +
  # Add shaded regions for disease states
  annotate("rect", xmin = -0.5, xmax = 0.5, ymin = -Inf, ymax = Inf, 
           fill = "lightgreen", alpha = 0.05) +
  annotate("rect", xmin = 0.5, xmax = 1.5, ymin = -Inf, ymax = Inf, 
           fill = "lightblue", alpha = 0.05) +
  annotate("rect", xmin = 1.5, xmax = 2.5, ymin = -Inf, ymax = Inf, 
           fill = "lightyellow", alpha = 0.05) +
  # Add vertical lines for cohorts
  geom_vline(
    xintercept = c(0.5, 1.5),
    linetype = "solid",
    color = "grey70",
    alpha = 0.5,
    size = 0.5
  ) +
  # Add THICKER horizontal lines to separate pattern groups
  geom_hline(
    data = data.frame(y = pattern_positions$pattern_y_start[-1] - pattern_spacing/2),
    aes(yintercept = y),
    linetype = "solid",
    color = "grey50",
    alpha = 0.8,
    size = 1  # Thicker lines for better separation
  ) +
  # Add connecting lines for each patient
  geom_line(
    aes(group = mrn),
    color = "grey50",
    alpha = 0.5
  ) +
  # Add points colored by sample type
  geom_point(
    aes(color = txt_stat_ordered),
    size = 2.5,
    shape = 16
  ) +
  # Add pattern group labels with boxes
  geom_rect(
    data = pattern_positions,
    aes(xmin = 2.6, xmax = 3.4,
        ymin = pattern_y_start + (n_patients - 1) / 2 - 0.8,
        ymax = pattern_y_start + (n_patients - 1) / 2 + 0.8),
    fill = "white",
    color = "grey60",
    size = 0.5,
    inherit.aes = FALSE
  ) +
  geom_text(
    data = pattern_positions,
    aes(x = 3, y = pattern_y_start + (n_patients - 1) / 2,
        label = pattern_label),
    size = 2.8,
    hjust = 0.5,
    inherit.aes = FALSE
  ) +
  # Color scale
  scale_color_manual(
    values = c(
      "Pre1" = "steelblue", 
      "Pre2" = "steelblue", 
      "Pre3" = "red",
      "Post1" = "orange", 
      "Post2" = "orange", 
      "Post3" = "orange"
    ),
    name = "Sample Type",
    labels = c(
      "Pre1" = "Pre-treatment 1", 
      "Pre2" = "Pre-treatment 2", 
      "Pre3" = "Pre-treatment 3 (special)",
      "Post1" = "Post-treatment 1", 
      "Post2" = "Post-treatment 2", 
      "Post3" = "Post-treatment 3"
    )
  ) +
  # X-axis setup
  scale_x_continuous(
    breaks = 0:2,
    labels = c("Localized", "mHSPC", "mCRPC"),
    limits = c(-0.5, 3.5)
  ) +
  # Y-axis setup
  scale_y_continuous(
    trans = "reverse",  # Invert y-axis
    breaks = pattern_positions$pattern_y_start + pattern_positions$n_patients/2,
    labels = rep("", nrow(pattern_positions)),  # Remove y-axis labels since we have boxes
    sec.axis = sec_axis(
      ~.,
      breaks = pattern_positions$pattern_y_start + pattern_positions$n_patients/2,
      labels = paste0("(n=", pattern_positions$n_patients, ")")
    )
  ) +
  # Labels
  labs(
    x = "Disease State",
    y = "",
    color = "Sample Type"
  ) +
  # Theme
  theme_bw() +
  theme(
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.y = element_blank(),
    axis.ticks.y = element_blank(),
    axis.text.y.right = element_text(size = 7, color = "grey40"),
    axis.text.x = element_text(size = 11, face = "bold"),
    axis.title.x = element_text(size = 11, face = "bold"),
    legend.position = "right",
    plot.margin = margin(t = 0, r = 5, b = 5, l = 5),
    panel.border = element_rect(color = "grey50", size = 1)
  ) +
  coord_cartesian(clip = "off")  # Allow drawing outside plot area

# Combine the plots using patchwork
combined_plot <- p_bars / p_grouped + 
  plot_layout(heights = c(1, 5)) + 
  plot_annotation(
    title = "Sample Collection Timeline - Grouped by Pattern",
    subtitle = "Disease progression from Localized → mHSPC (metastatic Hormone-Sensitive) → mCRPC (metastatic Castration-Resistant)",
    theme = theme(
      plot.title = element_text(face = "bold", size = 14),
      plot.subtitle = element_text(size = 10, face = "italic", color = "grey40")
    )
  )

# Save the combined plot
ggsave(
  combined_plot,
  file=file.path(wd$outCurr, "swimmers_plot_improved_spacing.pdf"),
  height = 18,  # Increased height for better spacing
  width = 12,
  limitsize = FALSE
)


```

## A heatmap like-style plot

```{r}
pattern_order_edited <- c(
  "A",

  "B_Pre_Only", 
  "B_Serial_PrePost",
  "B_Post_Only",
  "B_Pre_to_C_Pre",
  "B_Post_to_C_Post",
  "B_Post_to_C_Pre_to_C_Post",
  "B_Post_to_C_Pre",
  
  
  "C_Pre_Only",
  "C_Serial_PrePost",
  "C_Post_Only",
  "C_Post1_to_C_Post2"
  
)


# Create pattern factor with desired order
swimmers_data$pattern <- factor(swimmers_data$pattern, levels = pattern_order_edited)

patient_x_order <- swimmers_data %>%
  distinct(mrn, pattern) %>%
  arrange(pattern, mrn) %>%
  mutate(x_pos_new = row_number())

# Create new y-positions grouped by pattern with more spacing

plot_df <- swimmers_data %>%
  mutate(
    fill_key = case_when(
      cohort == "A" & txt_stat_ordered == "Pre1"  ~ "A_Pre",
      cohort == "B" & txt_stat_ordered == "Pre1"  ~ "B_Pre",
      cohort == "C" & txt_stat_ordered == "Pre1"  ~ "C_Pre",
      
      cohort == "A" & txt_stat_ordered == "Post1" ~ "A_Post",
      cohort == "B" & txt_stat_ordered == "Post1" ~ "B_Post",
      cohort == "C" & txt_stat_ordered == "Post1" ~ "C_Post",
      cohort == "C" & txt_stat_ordered == "Post2" ~ "C_Post2"
    )
  ) %>%
  filter(!is.na(fill_key))


plot_df <- plot_df %>%
  left_join(
    patient_x_order %>% select(mrn, x_pos_new),
    by = "mrn"
  )

# Reverese the order
x_levels <- sort(unique(plot_df$x_pos))
spacing_factor <- 1.5  # try 1.5, 2, or 2.5

x_map <- setNames(
  rev(x_levels),
  x_levels
)

plot_df <- plot_df %>%
  mutate(x_pos_rev = x_map[as.character(x_pos)],
         x_pos_spaced = x_pos_rev * spacing_factor,
         x_pos_rev = gsub("-0.25", "0.5", x_pos_rev))


p_simple <- ggplot(
  plot_df,
  aes(x = x_pos_new, y = x_pos_rev)
) +
  geom_point(
    aes(fill = fill_key),
    shape = 21,
    size = 3,
    stroke = 0.25,
    color = "black"
  ) +
  scale_fill_manual(
    values = pal.all,
    name = "Cohort / Treatment"
  ) +
  labs(
    x = "Patients",
    y = "",
    title = "Sample availability by patient and disease stage"
  ) +
  theme_bw() +
  theme(
    panel.grid = element_blank(),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank(),
    axis.text.y = element_text(size = 10, face = "bold"),
    plot.title = element_text(face = "bold"),
    legend.position = "right"
  ) +
  geom_line(
    aes(group = mrn),
    color = "grey50",
    alpha = 0.5
  ) +
  NULL

p_simple

# Save the combined plot
ggsave(
  p_simple,
  file=file.path(wd$outCurr, "heatmap_style.pdf"),
  height = 4,  # Increased height for better spacing
  width = 20,
  limitsize = FALSE
)

```

## Sample number 

Lets get the sample numbers

```{r}
# Calculate total number of samples and patients for all cohorts (regardless of cohort)
n_total_samples <- df_meta_f4 %>% pull(HCI_cID) %>% unique() %>% length()
n_total_patients <- df_meta_f4 %>% pull(mrn) %>% unique() %>% length()

# Calculate total number of samples and patients for Cohorts A, B, and C (regardless of treatment)
n_a_samples <- df_meta_f4 %>% filter(cohort == "A") %>% pull(HCI_cID) %>% unique() %>% length()
n_a_patients <- df_meta_f4 %>% filter(cohort == "A") %>% pull(mrn) %>% unique() %>% length()

n_b_samples <- df_meta_f4 %>% filter(cohort == "B") %>% pull(HCI_cID) %>% unique() %>% length()
n_b_patients <- df_meta_f4 %>% filter(cohort == "B") %>% pull(mrn) %>% unique() %>% length()

n_c_samples <- df_meta_f4 %>% filter(cohort == "C") %>% pull(HCI_cID) %>% unique() %>% length()
n_c_patients <- df_meta_f4 %>% filter(cohort == "C") %>% pull(mrn) %>% unique() %>% length()

cat("All cohorts:\n")
cat("  Total samples:", n_total_samples, "\n")
cat("  Total patients:", n_total_patients, "\n\n")

cat("Cohort A:\n")
cat("  Total samples:", n_a_samples, "\n")
cat("  Total patients:", n_a_patients, "\n\n")

cat("Cohort B:\n")
cat("  Total samples:", n_b_samples, "\n")
cat("  Total patients:", n_b_patients, "\n\n")

cat("Cohort C:\n")
cat("  Total samples:", n_c_samples, "\n")
cat("  Total patients:", n_c_patients, "\n\n")

# Calculate number of Pre and Post samples and patients for Cohort B and C using df_meta_f4

# For Cohort B
n_pre_b_samples <- df_meta_f4 %>% filter(cohort == "B", txt_stat %in% c("Pre", "Pre1", "Pre2", "Pre3")) %>% pull(HCI_cID) %>% unique() %>% length()
n_post_b_samples <- df_meta_f4 %>% filter(cohort == "B", txt_stat %in% c("Post", "Post1", "Post2", "Post3")) %>% pull(HCI_cID) %>% unique() %>% length()
n_pre_b_patients <- df_meta_f4 %>% filter(cohort == "B", txt_stat %in% c("Pre", "Pre1", "Pre2", "Pre3")) %>% pull(mrn) %>% unique() %>% length()
n_post_b_patients <- df_meta_f4 %>% filter(cohort == "B", txt_stat %in% c("Post", "Post1", "Post2", "Post3")) %>% pull(mrn) %>% unique() %>% length()

# For Cohort C
n_pre_c_samples <- df_meta_f4 %>% filter(cohort == "C", txt_stat %in% c("Pre", "Pre1", "Pre2", "Pre3")) %>% pull(HCI_cID) %>% unique() %>% length()
n_post_c_samples <- df_meta_f4 %>% filter(cohort == "C", txt_stat %in% c("Post", "Post1", "Post2", "Post3")) %>% pull(HCI_cID) %>% unique() %>% length()
n_pre_c_patients <- df_meta_f4 %>% filter(cohort == "C", txt_stat %in% c("Pre", "Pre1", "Pre2", "Pre3")) %>% pull(mrn) %>% unique() %>% length()
n_post_c_patients <- df_meta_f4 %>% filter(cohort == "C", txt_stat %in% c("Post", "Post1", "Post2", "Post3")) %>% pull(mrn) %>% unique() %>% length()

cat("B cohort (by treatment):\n")
cat("  Pre samples:", n_pre_b_samples, "\n")
cat("  Post samples:", n_post_b_samples, "\n")
cat("  Pre patients:", n_pre_b_patients, "\n")
cat("  Post patients:", n_post_b_patients, "\n\n")

cat("C cohort (by treatment):\n")
cat("  Pre samples:", n_pre_c_samples, "\n")
cat("  Post samples:", n_post_c_samples, "\n")
cat("  Pre patients:", n_pre_c_patients, "\n")
cat("  Post patients:", n_post_c_patients, "\n")
```

# Objective 4: Table 

## One df with all cohorts


We want to get additional information about the cohorts to make a supplemental table. But we only select those that pass the final selection

## Cohort A

```{r}
# Age
summary(df_a2$age)

# Prostate treatment
df_a2 <- df_a2 %>% 
  mutate(sur_or_rad = case_when(
    str_detect(sur_or_rad, "^Surgery") ~ "Surgery",
    str_detect(sur_or_rad, "^Radiation") ~ "Radiation",
    TRUE ~ sur_or_rad))
  
# Display table with percentages for sur_or_rad
tbl <- table(df_a2$sur_or_rad, useNA = "ifany")
cat("  Treatment:", "\n")
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))

# Display table with percentages
cat("  Gleason:", "\n")
tbl <- table(case_when(is.na(df_a2$glsn_diag) ~ "Missing", df_a2$glsn_diag < 8 ~ "<8", df_a2$glsn_diag >= 8 ~ ">=8"))
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))

missing_n_pct <- function(x) {
  n_patients <- length(x)
  n_missing  <- sum(is.na(x))
  pct_missing <- round(100 * n_missing / n_patients, 1)

  paste0(n_missing, " (", pct_missing, "%)")
}

summary_with_missing <- function(x) {
  s <- summary(x)
  c(
    s,
    Missing = missing_n_pct(x)
  )
}

cat("  PSA:\n")
summary_with_missing(df_a2$psa)

cat("  ALP:\n")
summary_with_missing(df_a2$alk_phos)
```

## Cohort B

```{r}
# Since they are serials, we need to get the unique mrns

# Age
d1 <- df_b3 %>% dplyr::select(mrn, age) %>% distinct()
dim(d1)
summary(d1$age)

# Prostate treatment
df_b3 <- df_b3 %>% 
  mutate(sur_or_rad = case_when(
    str_detect(sur_or_rad, "^Surgery") ~ "Surgery",
    str_detect(sur_or_rad, "^Radiation") ~ "Radiation",
    TRUE ~ sur_or_rad))
d1 <- df_b3 %>% dplyr::select(mrn, age, sur_or_rad) %>% distinct()
dim(d1)
  
# Display table with percentages for sur_or_rad
tbl <- table(d1$sur_or_rad, useNA = "ifany")
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))

# Gleason
d1 <- df_b3 %>% 
  group_by(mrn) %>%
  slice_min(date, n = 1) %>%
  ungroup()
# Display table with percentages
tbl <- table(case_when(is.na(d1$glsn_diag) ~ "Missing", d1$glsn_diag < 8 ~ "<8", d1$glsn_diag >= 8 ~ ">=8"))
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))

# 1st PSA
cat("  PSA:\n")
summary_with_missing(d1$psa)

# 1st ALP
cat("  ALP:\n")
summary_with_missing(d1$alk_phos)

# De novo mets stage
tbl <- table(d1$mets_stage, useNA = "ifany")
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))

# Volumne at first mets
tbl <- table(d1$mets_volume, useNA = "ifany")
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))


# Find PSA of those serially sampeled
mrn_pre_post <- df_meta_f4 %>%
  filter(cohort == "B") %>% 
  group_by(mrn) %>%
  summarise(
    has_pre = as.integer(any(txt_stat == "Pre", na.rm = TRUE)),
    has_post = as.integer(any(txt_stat == "Post", na.rm = TRUE))
  )

# View only those with both Pre and Post
paired_mrn <- mrn_pre_post %>%
  filter(has_pre == 1 & has_post == 1)
dim(paired_mrn)

# PSA pre
d2 <- df_meta_f4 %>% filter(mrn %in% paired_mrn$mrn) %>% 
  filter(cohort == "B" & txt_stat == "Pre")
d3 <- df_b3 %>% filter(HCI_cID %in% d2$HCI_cID)
summary_with_missing(d3$psa)

# PSA post
d2 <- df_meta_f4 %>% filter(mrn %in% paired_mrn$mrn) %>% 
  filter(cohort == "B" & txt_stat == "Post")
d3 <- df_b3 %>% filter(HCI_cID %in% d2$HCI_cID)
summary_with_missing(d3$psa)

# Median time between pre and post
med_time <- df_meta_f4 %>%
  filter(mrn %in% paired_mrn$mrn) %>%
  group_by(mrn, txt_stat) %>%
  summarise(median_time = median(time2txt, na.rm = TRUE)) %>%
  pivot_wider(names_from = txt_stat, values_from = median_time) %>%
  mutate(diff = Post - Pre) %>%
  arrange(desc(diff))
summary_with_missing(med_time$diff)
```

## Cohort C

```{r}
# Age
d1 <- df_c2 %>% dplyr::select(mrn, age) %>% distinct()
dim(d1)
summary_with_missing(d1$age)

# Prostate treatment
df_c2 <- df_c2 %>% 
  mutate(sur_or_rad = case_when(
    str_detect(sur_or_rad, "^Surgery") ~ "Surgery",
    str_detect(sur_or_rad, "^Radiation") ~ "Radiation",
    TRUE ~ sur_or_rad))
d1 <- df_c2 %>% 
  filter(!HCI_cID == "2022-0449") %>% 
  dplyr::select(mrn, age, sur_or_rad) %>% distinct()
dim(d1)
# We got 120, one more because a patient - 2021-4689 and 2022-0449 has both surgery and radiation
# Surgery was more recent - so we take that
#d1 <- d1 %>% filter(!HCI_cID == "2022-0449")
# Display table with percentages for sur_or_rad
tbl <- table(d1$sur_or_rad, useNA = "ifany")
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))


# Gleason
d1 <- df_c2 %>% 
  group_by(mrn) %>%
  slice_min(date, n = 1) %>%
  ungroup()
# Display table with percentages
tbl <- table(case_when(is.na(d1$glsn_diag) ~ "Missing", d1$glsn_diag < 8 ~ "<8", d1$glsn_diag >= 8 ~ ">=8"))
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))

# 1st PSA
cat("  PSA:\n")
summary_with_missing(d1$psa)

# Mets stage
tbl <- table(d1$mets_stage, useNA = "ifany")
setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))

# Count by simplified mHSPC treatment for Pre (baseline) patients only
d1 <- df_c2 %>%
  dplyr::select(mrn, mcrpc_status_early, earliest_mhspc_treatment_simplified) %>% distinct() %>% 
  filter(mcrpc_status_early == "Pre")

tbl <- table(d1$earliest_mhspc_treatment_simplified, useNA = "ifany")
result <- setNames(paste0(tbl, " [", round(prop.table(tbl) * 100, 1), "%]"), names(tbl))

cat("Baseline (Pre) patients by simplified mHSPC treatment type:\n")
print(result)
cat("\nTotal Pre (baseline) patients:", sum(tbl), "\n")
```



# Save objects

```{r}
df_a_final <- df_a2 %>% 
  mutate(mrn = as.character(mrn))
df_b_final <- df_b3 %>% 
  mutate(mrn = as.character(mrn))
df_c_final <- df_c2 %>% 
  mutate(mrn = as.character(mrn))
```


For all cohorts

## Simplify all cohorts

We extract info from all cohorts

```{r}
# 1st - Create fake mrn 
mrn_key <- df_meta_f2 %>%
  distinct(mrn) %>%
  arrange(mrn) %>%
  mutate(fake_mrn = sprintf("mrn_%03d", row_number()))


# Add treatment info
df_txt <- df_meta_f2 %>% 
  mutate(mrn = as.character(mrn)) %>% 
  left_join(mrn_key, by = "mrn") %>%
  select(HCI_cID, cohort, txt_stat_ordered) %>% 
  dplyr::rename(txt_stat = txt_stat_ordered) %>% 
  mutate(txt_stat = case_when(txt_stat == "Pre1" ~ "Pre",
                              TRUE ~ txt_stat))


# 2nd - Metadata of clinical values

## Cohort A
df_meta_cohort_a <- df_a_final %>% 
  left_join(mrn_key, by = "mrn") %>%
  dplyr::select(-cohort) %>% 
  left_join(df_txt) %>% 
  mutate(
    death_status = ifelse(death_status == "Yes", "y", "n"),
    event_date = dplyr::case_when(
      !is.na(death_date) ~ death_date,
      TRUE ~ as.Date(censor_date)
    )
  ) %>% 
  mutate(OS = as.numeric(difftime(event_date, date, units = "days")),
         OS = round(OS/30.417, digit=2)) %>% 
  select(fake_mrn, HCI_cID, cohort, txt_stat, death_status, OS, age, glsn_diag, psa, alk_phos, ldh, hemoGlo)
  
## Cohort B
df_meta_cohort_b <- df_b_final %>%
  left_join(mrn_key, by = "mrn") %>%
  dplyr::select(-cohort) %>% 
  left_join(df_txt) %>% 
  mutate(
    death_status = ifelse(death_status == "Yes", "y", "n"),
    event_date = dplyr::case_when(
      !is.na(death_date) ~ death_date,
      TRUE ~ as.Date(censor_date)
    )
  ) %>% 
  mutate(OS = as.numeric(difftime(event_date, date, units = "days")),
         OS = round(OS/30.417, digit=2)) %>% 
  select(fake_mrn, HCI_cID, cohort, txt_stat, death_status, OS, age, glsn_diag, psa, alk_phos, ldh, hemoGlo)


## Cohort C
df_meta_cohort_c <- df_c_final %>%
  left_join(mrn_key, by = "mrn") %>%
  dplyr::select(-cohort) %>% 
  left_join(df_txt) %>% 
  mutate(
    death_status = ifelse(death_status == "Yes", "y", "n"),
    event_date = dplyr::case_when(
      !is.na(death_date) ~ death_date,
      TRUE ~ as.Date(censor_date)
    )
  ) %>% 
  mutate(OS = as.numeric(difftime(event_date, date, units = "days")),
         OS = round(OS/30.417, digit=2)) %>% 
  select(fake_mrn, HCI_cID, cohort, txt_stat, death_status, OS, age, glsn_diag, psa, alk_phos, ldh, hemoGlo)
```

## Per-batch NPX

We want to save the De-identified NPX, with the sample names listed sawpped with the HCI ID


```{r}
dat_NPX_save <- dat_NPX %>%
  left_join(
    df_hci_id %>%
      mutate(sample_id_psom = as.character(sample_id_psom)) %>%
      select(sample_id_psom, HCI_cID),
    by = c("SampleID" = "sample_id_psom")
  ) %>%
  mutate(SampleID = if_else(
    !is.na(HCI_cID),
    HCI_cID,
    SampleID
  )) %>%
  select(-HCI_cID)

  
# For data2, we exclude the repeated measurement
dat_NPX2_save <- dat_NPX2 %>%
  # Remove the duplciated sampels
  dplyr::filter(!SampleID == 78) %>% 
  left_join(
    df_hci_id %>%
      mutate(sample_id_Q13356 = as.character(sample_id_Q13356)) %>%
      select(sample_id_Q13356, HCI_cID),
    by = c("SampleID" = "sample_id_Q13356")
  ) %>%
  mutate(SampleID = if_else(!is.na(HCI_cID), HCI_cID, SampleID)) %>%
  select(-HCI_cID)


dat_NPX3_save <- dat_NPX3 %>%
  left_join(
    df_hci_id %>%
      mutate(sample_id_Q15806 = as.character(sample_id_Q15806)) %>%
      select(sample_id_Q15806, HCI_cID),
    by = c("SampleID" = "sample_id_Q15806")
  ) %>%
  mutate(SampleID = if_else(!is.na(HCI_cID), HCI_cID, SampleID)) %>%
  select(-HCI_cID)
```
## Follow-up

Get follow up time


```{r}
safe_min_date <- function(x) {
  x <- as.Date(x)
  if (all(is.na(x))) as.Date(NA) else min(x, na.rm = TRUE)
}

df_followup <- dplyr::bind_rows(
  df_a2 %>%
    dplyr::transmute(
      mrn = as.character(mrn),
      sample_date = as.Date(date),
      death_date = as.Date(death_date),
      death_status = tolower(as.character(death_status))
    ),
  df_b3 %>%
    dplyr::transmute(
      mrn = as.character(mrn),
      sample_date = as.Date(date),
      death_date = as.Date(death_date),
      death_status = tolower(as.character(death_status))
    ),
  df_c2 %>%
    dplyr::transmute(
      mrn = as.character(mrn),
      sample_date = as.Date(date),
      death_date = as.Date(death_date),
      death_status = tolower(as.character(death_status))
    )
) %>%
  dplyr::group_by(mrn) %>%
  dplyr::summarise(
    first_collection_date = safe_min_date(sample_date),
    first_death_date = safe_min_date(death_date),
    died = any(death_status %in% c("yes", "y", "dead", "1") | !is.na(death_date), na.rm = TRUE),
    event_date = dplyr::if_else(
      died & !is.na(first_death_date) & first_death_date <= censor_date,
      first_death_date,
      censor_date
    ),
    event = dplyr::if_else(
      died & !is.na(first_death_date) & first_death_date <= censor_date, 1L, 0L
    ),
    followup_days = as.numeric(event_date - first_collection_date),
    followup_months = round(followup_days / 30.4375, 2),
    .groups = "drop"
  )
```


## Progression

```{r}
is_yes <- function(x) {
  str_to_lower(trimws(as.character(x))) %in% c("yes", "y", "1", "true")
}

# 1) mHSPC cohort (B): how many progressed to mCRPC
mhspc_prog <- df_b3 %>%
  mutate(
    mrn = as.character(mrn),
    progressed_flag = is_yes(progressed_to_crpc) | !is.na(crpc_earliest_date)
  ) %>%
  group_by(mrn) %>%
  summarise(progressed = any(progressed_flag, na.rm = TRUE), .groups = "drop") %>%
  summarise(
    progressed_n = sum(progressed, na.rm = TRUE),
    mhspc_total = n(),
    .groups = "drop"
  )

# 2) mCRPC sub-cohort (C): xx/yy died
mcrpc_death <- df_c2 %>%
  mutate(
    mrn = as.character(mrn),
    death_flag = is_yes(death_status) | !is.na(death_date)
  ) %>%
  group_by(mrn) %>%
  summarise(died = any(death_flag, na.rm = TRUE), .groups = "drop") %>%
  summarise(
    died_n = sum(died, na.rm = TRUE),
    mcrpc_total = n(),
    .groups = "drop"
  )

mhspc_prog
mcrpc_death
```


## Extended results text

```{r}
# Analysis window
enroll_start <- as.Date("2020-08-01")
enroll_end <- as.Date("2025-04-01")
censor_date <- as.Date("2025-07-01")

safe_min_date <- function(x) {
  x <- as.Date(x)
  if (all(is.na(x))) as.Date(NA) else min(x, na.rm = TRUE)
}

safe_max_date <- function(x) {
  x <- as.Date(x)
  if (all(is.na(x))) as.Date(NA) else max(x, na.rm = TRUE)
}

fmt_med_iqr <- function(x, digits = 1) {
  x <- as.numeric(x)
  x <- x[!is.na(x)]
  if (length(x) == 0) return("NA")
  qs <- stats::quantile(x, probs = c(0.25, 0.50, 0.75), na.rm = TRUE)
  paste0(
    formatC(qs[[2]], format = "f", digits = digits),
    " (IQR ",
    formatC(qs[[1]], format = "f", digits = digits), "-",
    formatC(qs[[3]], format = "f", digits = digits), ")"
  )
}

fmt_med_iqr_range <- function(x, digits = 1) {
  x <- as.numeric(x)
  x <- x[!is.na(x)]
  if (length(x) == 0) return("NA")
  qs <- stats::quantile(x, probs = c(0.25, 0.50, 0.75), na.rm = TRUE)
  paste0(
    formatC(qs[[2]], format = "f", digits = digits),
    " (IQR ",
    formatC(qs[[1]], format = "f", digits = digits), "-",
    formatC(qs[[3]], format = "f", digits = digits),
    "; range ",
    formatC(min(x), format = "f", digits = digits), "-",
    formatC(max(x), format = "f", digits = digits),
    ")"
  )
}

fmt_n_pct <- function(n, d, digits = 1) {
  ifelse(d > 0, paste0(n, "/", d, " (", round(100 * n / d, digits), "%)"), paste0(n, "/", d, " (NA)"))
}

standardize_cohort <- function(df, cohort_label) {
  required_cols <- c(
    "mrn", "HCI_cID", "date", "death_date", "death_status", "age", "psa", "alk_phos",
    "ldh", "hemoGlo", "glsn_diag", "sur_or_rad", "mets_stage", "progressed_to_crpc", "crpc_earliest_date"
  )
  missing_cols <- setdiff(required_cols, names(df))
  for (nm in missing_cols) df[[nm]] <- NA

  df %>%
    dplyr::transmute(
      cohort = cohort_label,
      mrn = as.character(mrn),
      HCI_cID = as.character(HCI_cID),
      sample_date = as.Date(date),
      death_date = as.Date(death_date),
      death_status = tolower(as.character(death_status)),
      age = as.numeric(age),
      psa = as.numeric(psa),
      alk_phos = as.numeric(alk_phos),
      ldh = as.numeric(ldh),
      hemoGlo = as.numeric(hemoGlo),
      glsn_diag = as.character(glsn_diag),
      sur_or_rad = as.character(sur_or_rad),
      mets_stage = as.character(mets_stage),
      progressed_to_crpc = as.character(progressed_to_crpc),
      crpc_earliest_date = as.Date(crpc_earliest_date)
    )
}

cohort_long <- dplyr::bind_rows(
  standardize_cohort(df_a2, "Localized"),
  standardize_cohort(df_b3, "mHSPC"),
  standardize_cohort(df_c2, "mCRPC")
)

# Patient-level table (earliest sample date per patient)
patient_table <- cohort_long %>%
  dplyr::group_by(cohort, mrn) %>%
  dplyr::summarise(
    first_collection_date = safe_min_date(sample_date),
    last_collection_date = safe_max_date(sample_date),
    n_samples = dplyr::n_distinct(HCI_cID),
    first_death_date = safe_min_date(death_date),
    died = any(death_status %in% c("yes", "y", "dead", "1") | !is.na(death_date), na.rm = TRUE),
    progressed = any(is_yes(progressed_to_crpc) | !is.na(crpc_earliest_date), na.rm = TRUE),
    .groups = "drop"
  ) %>%
  dplyr::filter(first_collection_date >= enroll_start, first_collection_date <= enroll_end) %>%
  dplyr::mutate(
    event_date = dplyr::if_else(
      died & !is.na(first_death_date) & first_death_date <= censor_date,
      first_death_date,
      censor_date
    ),
    event = dplyr::if_else(
      died & !is.na(first_death_date) & first_death_date <= censor_date, 1L, 0L
    ),
    followup_months = as.numeric(event_date - first_collection_date) / 30.4375
  )

# Baseline sample-level characteristics (first sample per patient)
baseline <- cohort_long %>%
  dplyr::arrange(cohort, mrn, sample_date) %>%
  dplyr::group_by(cohort, mrn) %>%
  dplyr::slice_head(n = 1) %>%
  dplyr::ungroup() %>%
  dplyr::filter(sample_date >= enroll_start, sample_date <= enroll_end)

# Main cohort summary
cohort_summary <- patient_table %>%
  dplyr::group_by(cohort) %>%
  dplyr::summarise(
    n_patients = dplyr::n(),
    n_samples = sum(n_samples),
    n_longitudinal = sum(n_samples >= 2, na.rm = TRUE),
    n_died = sum(event == 1, na.rm = TRUE),
    med_followup = fmt_med_iqr_range(followup_months, digits = 2),
    .groups = "drop"
  ) %>%
  dplyr::left_join(
    baseline %>%
      dplyr::group_by(cohort) %>%
      dplyr::summarise(
        age_summary = fmt_med_iqr(age, digits = 1),
        psa_summary = fmt_med_iqr(psa, digits = 1),
        alk_summary = fmt_med_iqr(alk_phos, digits = 1),
        ldh_summary = fmt_med_iqr(ldh, digits = 1),
        hgb_summary = fmt_med_iqr(hemoGlo, digits = 1),
        glsn_known = sum(!is.na(glsn_diag) & glsn_diag != "", na.rm = TRUE),
        n_baseline = dplyr::n(),
        local_tx_n = sum(!is.na(sur_or_rad) & sur_or_rad != "", na.rm = TRUE),
        .groups = "drop"
      ),
    by = "cohort"
  )

# mHSPC-specific progression and de novo metastatic burden
mhspc_row <- patient_table %>%
  dplyr::filter(cohort == "mHSPC") %>%
  dplyr::summarise(
    progressed_n = sum(progressed, na.rm = TRUE),
    total_n = dplyr::n(),
    .groups = "drop"
  )

mhspc_denovo <- baseline %>%
  dplyr::filter(cohort == "mHSPC") %>%
  dplyr::summarise(
    denovo_n = sum(is_yes(mets_stage), na.rm = TRUE),
    total_n = dplyr::n(),
    .groups = "drop"
  )

# Pre/Post counts from final filtered sample set
pre_post_line <- ""
if (exists("df_meta_f4")) {
  pre_post <- df_meta_f4 %>%
    dplyr::mutate(
      cohort = dplyr::recode(cohort, "A" = "Localized", "B" = "mHSPC", "C" = "mCRPC"),
      txt_group = dplyr::case_when(
        stringr::str_detect(txt_stat, "^Pre") ~ "Pre",
        stringr::str_detect(txt_stat, "^Post") ~ "Post",
        TRUE ~ NA_character_
      )
    ) %>%
    dplyr::filter(!is.na(txt_group)) %>%
    dplyr::count(cohort, txt_group, name = "n_samples")

  pre_post_line <- pre_post %>%
    dplyr::arrange(cohort, txt_group) %>%
    dplyr::mutate(txt = paste0(cohort, " ", txt_group, " n=", n_samples)) %>%
    dplyr::pull(txt) %>%
    paste(collapse = "; ")
}
```


## Final meta

```{r}
df_meta_all <- rbind(df_meta_cohort_a, df_meta_cohort_b, df_meta_cohort_c)

write.csv(df_meta_all, file.path(wd$outData, "proteomics_metadata.csv"))
```

> TO DO - add the cohort B timing to mCRPC ...etc

## Zendo files

```{r}
df_meta_all2 <- df_meta_all %>% 
  dplyr::mutate(HCI_cID = paste0("c", HCI_cID),
                HCI_cID = gsub("-", "_", HCI_cID)) %>% 
  dplyr::rename(Patient_ID = fake_mrn) %>% 
  dplyr::mutate(Patient_ID = gsub("mrn", "P",Patient_ID),
                cohort = case_when(cohort == "A" ~ "Local",
                                   cohort == "B" ~ "mHSPC",
                                   cohort == "C" ~ "mCRPC"))
write.csv(df_meta_all2, file.path(wd$outData, "proteomics_metadata_all.csv"))
```

Save the individual NPX data and add the HCI_cID

```{r}
tmp_in <- df_meta_f %>% 
    dplyr::mutate(HCI_cID = paste0("c", HCI_cID),
                HCI_cID = gsub("-", "_", HCI_cID))


# 1st batch
dat_NPX_withID <- dat_NPX %>%
  left_join(
    tmp_in %>%
      mutate(sample_id_psom = as.character(sample_id_psom)) %>%
      select(HCI_cID, sample_id_psom),
    by = c("SampleID" = "sample_id_psom")
  ) %>%
  relocate(HCI_cID, .after = 1)
write.csv(dat_NPX_withID, file.path(wd$outData, "Batch_01_NPX.csv"))

# 2nd batch
dat_NPX2_withID <- dat_NPX2 %>%
  # Exclude sample 78 - missing data
  dplyr::filter(!SampleID == 78) %>% 
  left_join(
    tmp_in %>%
      mutate(sample_id_Q13356 = as.character(sample_id_Q13356)) %>%
      select(HCI_cID, sample_id_Q13356),
    by = c("SampleID" = "sample_id_Q13356")
  ) %>%
  relocate(HCI_cID, .after = 1)
write.csv(dat_NPX2_withID, file.path(wd$outData, "Batch_02_NPX.csv"))


# 3rd batch
dat_NPX3_withID <- dat_NPX3 %>%
  left_join(
    tmp_in %>%
      mutate(sample_id_Q15806 = as.character(sample_id_Q15806)) %>%
      select(HCI_cID, sample_id_Q15806),
    by = c("SampleID" = "sample_id_Q15806")
  ) %>%
  relocate(HCI_cID, .after = 1)
write.csv(dat_NPX3_withID, file.path(wd$outData, "Batch_03_NPX.csv"))

```





## Save objects

Save items for future analysis

```{r}
save(
  dat_NPX, dat_NPX2, dat_NPX3, 
  meta_file, df_meta_f2, df_meta_f4, df_id_pro,
  df_a_final, df_b_final, df_c_final,
  df_meta_all, 
  file = file.path(wd$outData, "01_Metadata.Rdata"))
```
