==> scripts/04_Pathway_analysis.Rmd <==
---
title: "Pathway analysis"
date: "2025-06-30"
output: 
  html_document:
    toc: true
    toc_float:
      collapsed: true
    toc_depth: 3
    number_sections: false
    theme: lumen
---

# Background

We identified the over-expressed DEPs, now we want to run pathyway analysis on these proteins.

# Objectives

1. Pathway analysis


# Conclusion

Pathway analysis was performed on the proteins upregulated in mCRPC. We ran pathway analysis on proteins on each overexprssed cohort

# Pre-processing 

## Loading packages

```{r, message=FALSE, warning=FALSE}
library(here)
library(tidyverse)
library(paletteer)
library(clusterProfiler)
library(org.Hs.eg.db)
library(msigdbr)
library(enrichplot)

library(stringr)
library(gridExtra)
```


## Directories

```{r}
set.seed(1234)
wd <- list()
wd$main <- here()
wd$data <- file.path(wd$main, "data")
wd$manishData <- file.path(wd$data, "raw/data_fromManish")
wd$d2024 <- file.path(wd$data, "updated_2024")
wd$d2024_npx <- file.path(wd$d2024, "NPX_data")
wd$output <- file.path(wd$main, "output")
wd$script <- file.path(wd$main, "scripts")

wd$outData <- file.path(wd$output, "data")
wd$outCurr <- file.path(wd$output, "04_Pathway")
```


Create directories

```{r}
if (!file.exists(wd$outCurr)) {
  dir.create(wd$outCurr)
} else {
  print("Directory already exists")
}
```

## Load data

Load the processed file

```{r}
load(file.path(wd$outData, "01_Metadata.Rdata"))
load(file.path(wd$outData, "02_data.Rdata"))
load(file.path(wd$outData, "03_data.Rdata"))
```

Load functions

```{r}
source(file.path(wd$script, "functions.R"))
```


# Obj 1: Pathway analysis

## Upregulated genes

We run pathway analysis first on the upregulated genes in mCRPC

Lets extract the upregulated genes in mCRPC specifically

```{r}
genes_of_interest <- overlap_df_limma %>% 
  filter(
    Category %in% c("B and CD", "CD")) %>% 
  pull(Assay) %>% unique()

```

We need a background set for pathway enrichment. So we use all the gene ids

```{r}
background_genes <- df_id_pro %>% pull(Assay) %>% unique()
```

Run the pathway analysis

```{r}
# Convert genes of interest to Entrez IDs
foreground_entrez <- bitr(genes_of_interest, 
                          fromType = "SYMBOL", 
                          toType = "ENTREZID", 
                          OrgDb = org.Hs.eg.db)

# Convert background genes to Entrez IDs
background_entrez <- bitr(background_genes, 
                          fromType = "SYMBOL", 
                          toType = "ENTREZID", 
                          OrgDb = org.Hs.eg.db)

# Remove any duplicates or NA values
foreground_entrez <- unique(na.omit(foreground_entrez))
background_entrez <- unique(na.omit(background_entrez))

# Retrieve Hallmark gene sets for Homo sapiens
msig_hallmark <- msigdbr(species = "Homo sapiens", category = "H")

# Prepare TERM2GENE as a data frame
msig_term2gene <- msig_hallmark[, c("gs_name", "entrez_gene")]


# Perform over-representation analysis with the correct TERM2GENE
enrich_results <- enricher(gene = foreground_entrez$ENTREZID,
                           TERM2GENE = msig_term2gene,
                           #universe = background_entrez$ENTREZID,
                           pAdjustMethod = "BH",
                           pvalueCutoff = 0.05,
                           qvalueCutoff = 0.2)

#barplot(enrich_results, showCategory = 10, title = "Top Enriched Hallmark Pathways")

enrich_up <- enrich_results
```

Lets generate a custom plot

```{r}
# Custom plot
top_x <- enrich_up@result %>% 
  #filter(p.adjust <0.2) %>% 
  mutate(logP = -log10(p.adjust)) %>% 
  arrange(desc(logP)) %>% 
  mutate(ID = factor(ID, levels=rev(ID))) %>% 
  slice_max(logP, n=10, with_ties = FALSE) 

p_c <-
  ggplot(top_x,
       aes(x = logP, y=ID)) +
  geom_bar(stat = "identity", fill = "#2CA02C") +  # Bars with color
  geom_vline(xintercept = -log10(0.2), linetype="dashed") +
  theme_custom +
  NULL

#ggsave(file=file.path(wd$outCurr, "enrichedPathway.pdf"), height = 3, width = 5)
```

Repeat for the other group of genes (high in A not in B)

```{r}
genes_of_interest <- overlap_df_limma %>% 
  filter(
    Category %in% c("A and B", "B and CD", "B")) %>% 
  pull(Assay) %>% unique()


# Convert genes of interest to Entrez IDs
foreground_entrez <- bitr(genes_of_interest, 
                          fromType = "SYMBOL", 
                          toType = "ENTREZID", 
                          OrgDb = org.Hs.eg.db)

# Remove any duplicates or NA values
foreground_entrez <- unique(na.omit(foreground_entrez))
background_entrez <- unique(na.omit(background_entrez))

# Retrieve Hallmark gene sets for Homo sapiens
msig_hallmark <- msigdbr(species = "Homo sapiens", category = "H")

# Prepare TERM2GENE as a data frame
msig_term2gene <- msig_hallmark[, c("gs_name", "entrez_gene")]


# Perform over-representation analysis with the correct TERM2GENE
enrich_results <- enricher(gene = foreground_entrez$ENTREZID,
                           TERM2GENE = msig_term2gene,
                           #universe = background_entrez$ENTREZID,
                           pAdjustMethod = "BH",
                           pvalueCutoff = 0.05,
                           qvalueCutoff = 0.2)

enrich_up_b <- enrich_results

# Custom plot
top_x_b <- enrich_up_b@result %>% 
  #filter(p.adjust <0.2) %>% 
  mutate(logP = -log10(p.adjust)) %>% 
  arrange(desc(logP)) %>% 
  mutate(ID = factor(ID, levels=rev(ID))) %>% 
  slice_max(logP, n=10) 

p_b <-
  ggplot(top_x_b,
       aes(x = logP, y=ID)) +
  geom_bar(stat = "identity", fill = "#FFBB78") +  # Bars with color
  geom_vline(xintercept = -log10(0.2), linetype="dashed") +
  theme_custom +
  NULL

```


Plot the two in one plot

```{r}
t1 <- top_x %>% filter(p.adjust <= 0.2) %>% mutate(class = "mCRPC") %>% dplyr::select(logP, ID, class) %>% 
  mutate(
ID = gsub("_", " ", ID),
ID = gsub("HALLMARK", "", ID),
ID = str_to_lower(ID),
ID = str_to_title(ID))
t2 <- top_x_b %>% filter(p.adjust <= 0.2) %>% mutate(class = "mHSPC") %>% dplyr::select(logP, ID, class) %>% 
    mutate(
ID = gsub("_", " ", ID),
ID = gsub("HALLMARK", "", ID),
ID = str_to_lower(ID),
ID = str_to_title(ID))

# Get the full set of IDs to ensure spacing consistency
combined_IDs <- unique(c(t1$ID, t2$ID))

# Convert `ID` to a factor with all possible levels
t1$ID <- factor(t1$ID, levels = rev(combined_IDs))
t2$ID <- factor(t2$ID, levels = rev(combined_IDs))

# Get the maximum logP value across both datasets
max_logP <- max(c(t1$logP, t2$logP))

# Ensure both plots have the same x-axis scale
p1 <- ggplot(t1, aes(x = logP, y = ID)) +
  geom_bar(stat = "identity", fill = "#2CA02C") +
  geom_vline(xintercept = -log10(0.2), linetype = "dashed") +
  facet_wrap(~class) +
  xlim(0, max_logP) +  # Force both plots to have the same x-axis range
  theme_custom +
  NULL

p2 <- ggplot(t2, aes(x = logP, y = ID)) +
  geom_bar(stat = "identity", fill = "#FFBB78") +
  geom_vline(xintercept = -log10(0.2), linetype = "dashed") +
  facet_wrap(~class) +
  xlim(0, max_logP) +  # Same x-axis limit
  theme_custom +
  NULL

p1 <- p1 + coord_fixed(ratio = 0.2)  # Adjust ratio as needed
p2 <- p2 + coord_fixed(ratio = 0.2)

p_all <- grid.arrange(p1, p2, ncol = 1)


ggsave(p_all, file=file.path(wd$outCurr, "enrichedPathway_highC_highB.pdf"), height = 6, width = 7)
```



## Cohort specific

We repeat the above, but we will use just the cohort specific proteins

### mCRPC

Lets extract the upregulated genes in mCRPC specifically

```{r}
genes_of_interest <- overlap_df_limma %>% 
  filter(
    Category %in% c("CD")) %>% 
  pull(Assay) %>% unique()

```

We need a background set for pathway enrichment. So we use all the gene ids

```{r}
background_genes <- df_id_pro %>% pull(Assay) %>% unique()
```

Run the pathway analysis

```{r}
# Convert genes of interest to Entrez IDs
foreground_entrez <- bitr(genes_of_interest, 
                          fromType = "SYMBOL", 
                          toType = "ENTREZID", 
                          OrgDb = org.Hs.eg.db)

# Convert background genes to Entrez IDs
background_entrez <- bitr(background_genes, 
                          fromType = "SYMBOL", 
                          toType = "ENTREZID", 
                          OrgDb = org.Hs.eg.db)

# Remove any duplicates or NA values
foreground_entrez <- unique(na.omit(foreground_entrez))
background_entrez <- unique(na.omit(background_entrez))

# Retrieve Hallmark gene sets for Homo sapiens
msig_hallmark <- msigdbr(species = "Homo sapiens", category = "H")

# Prepare TERM2GENE as a data frame
msig_term2gene <- msig_hallmark[, c("gs_name", "entrez_gene")]


# Perform over-representation analysis with the correct TERM2GENE
enrich_results <- enricher(gene = foreground_entrez$ENTREZID,
                           TERM2GENE = msig_term2gene,
                           #universe = background_entrez$ENTREZID,
                           pAdjustMethod = "BH",
                           pvalueCutoff = 0.05,
                           qvalueCutoff = 0.2)

#barplot(enrich_results, showCategory = 10, title = "Top Enriched Hallmark Pathways")

enrich_up <- enrich_results
```

Lets generate a custom plot

```{r}
# Custom plot
top_x <- enrich_up@result %>% 
  #filter(p.adjust <0.2) %>% 
  mutate(logP = -log10(p.adjust)) %>% 
  arrange(desc(logP)) %>% 
  mutate(ID = factor(ID, levels=rev(ID))) %>% 
  slice_max(logP, n=10, with_ties = FALSE) 

p_c <-
  ggplot(top_x,
       aes(x = logP, y=ID)) +
  geom_bar(stat = "identity", fill = "#2CA02C") +  # Bars with color
  geom_vline(xintercept = -log10(0.2), linetype="dashed") +
  theme_custom +
  NULL

#ggsave(file=file.path(wd$outCurr, "enrichedPathway.pdf"), height = 3, width = 5)
```

Repeat for the other group of genes (high in A not in B)

```{r}
genes_of_interest <- overlap_df_limma %>% 
  filter(
    Category %in% c("B")) %>% 
  pull(Assay) %>% unique()


# Convert genes of interest to Entrez IDs
foreground_entrez <- bitr(genes_of_interest, 
                          fromType = "SYMBOL", 
                          toType = "ENTREZID", 
                          OrgDb = org.Hs.eg.db)

# Remove any duplicates or NA values
foreground_entrez <- unique(na.omit(foreground_entrez))
background_entrez <- unique(na.omit(background_entrez))

# Retrieve Hallmark gene sets for Homo sapiens
msig_hallmark <- msigdbr(species = "Homo sapiens", category = "H")

# Prepare TERM2GENE as a data frame
msig_term2gene <- msig_hallmark[, c("gs_name", "entrez_gene")]


# Perform over-representation analysis with the correct TERM2GENE
enrich_results <- enricher(gene = foreground_entrez$ENTREZID,
                           TERM2GENE = msig_term2gene,
                           #universe = background_entrez$ENTREZID,
                           pAdjustMethod = "BH",
                           pvalueCutoff = 0.05,
                           qvalueCutoff = 0.2)

enrich_up_b <- enrich_results

# Custom plot
top_x_b <- enrich_up_b@result %>% 
  #filter(p.adjust <0.2) %>% 
  mutate(logP = -log10(p.adjust)) %>% 
  arrange(desc(logP)) %>% 
  mutate(ID = factor(ID, levels=rev(ID))) %>% 
  slice_max(logP, n=10) 

p_b <-
  ggplot(top_x_b,
       aes(x = logP, y=ID)) +
  geom_bar(stat = "identity", fill = "#FFBB78") +  # Bars with color
  geom_vline(xintercept = -log10(0.2), linetype="dashed") +
  theme_custom +
  NULL

```


Plot the two in one plot

```{r}
t1 <- top_x %>% filter(p.adjust <= 0.2) %>% mutate(class = "mCRPC") %>% dplyr::select(logP, ID, class) %>% 
  mutate(
ID = gsub("_", " ", ID),
ID = gsub("HALLMARK", "", ID),
ID = str_to_lower(ID),
ID = str_to_title(ID))
t2 <- top_x_b %>% filter(p.adjust <= 0.2) %>% mutate(class = "mHSPC") %>% dplyr::select(logP, ID, class) %>% 
    mutate(
ID = gsub("_", " ", ID),
ID = gsub("HALLMARK", "", ID),
ID = str_to_lower(ID),
ID = str_to_title(ID))

# Get the full set of IDs to ensure spacing consistency
combined_IDs <- unique(c(t1$ID, t2$ID))

# Convert `ID` to a factor with all possible levels
t1$ID <- factor(t1$ID, levels = rev(combined_IDs))
t2$ID <- factor(t2$ID, levels = rev(combined_IDs))

# Get the maximum logP value across both datasets
max_logP <- max(c(t1$logP, t2$logP))

# Ensure both plots have the same x-axis scale
p1 <- ggplot(t1, aes(x = logP, y = ID)) +
  geom_bar(stat = "identity", fill = "#2CA02C") +
  geom_vline(xintercept = -log10(0.2), linetype = "dashed") +
  facet_wrap(~class) +
  xlim(0, max_logP) +  # Force both plots to have the same x-axis range
  theme_custom +
  NULL

p2 <- ggplot(t2, aes(x = logP, y = ID)) +
  geom_bar(stat = "identity", fill = "#FFBB78") +
  geom_vline(xintercept = -log10(0.2), linetype = "dashed") +
  facet_wrap(~class) +
  xlim(0, max_logP) +  # Same x-axis limit
  theme_custom +
  NULL

p1 <- p1 + coord_fixed(ratio = 0.2)  # Adjust ratio as needed
p2 <- p2 + coord_fixed(ratio = 0.2)

p_all2 <- grid.arrange(p1, p2, ncol = 1)


ggsave(p_all2, file=file.path(wd$outCurr, "enrichedPathway_highC_highB_unique.pdf"), height = 6, width = 7)
```


# Obj 2: KLK3 PSA

We check the KLK3 PSA association. 

## Overall

```{r}
# Get KLK3 OlinkID
klk3_id <- df_id_pro %>%
  filter(Assay == "KLK3") %>%
  pull(OlinkID)

# Link HCI_cID to final_sample_id for all mHSPC patients
df_link <- df_meta_f5 %>%
  select(HCI_cID, txt_stat_ordered) %>%
  left_join(df_final_ids %>% select(HCI_cID, final_sample_id))

# Use all patients 
psa_protein_data <- df_meta_all %>%
  filter(HCI_cID %in% df_meta_f5$HCI_cID) %>%
  mutate(logPSA = log10(psa)) %>% 
  select(HCI_cID, fake_mrn, logPSA, txt_stat, psa, cohort, glsn_diag) %>%
  left_join(df_link, by = "HCI_cID") %>%
  filter(!is.na(final_sample_id)) %>%
  filter(final_sample_id %in% colnames(exprMatrix_corrected_sub))

# Extract KLK3 expression for these samples
klk3_expression <- exprMatrix_corrected_sub[klk3_id, psa_protein_data$final_sample_id]
psa_protein_data$KLK3_NPX <- as.numeric(klk3_expression)

# Remove any rows with missing values
psa_protein_data <- psa_protein_data %>% filter(!is.na(logPSA), !is.na(KLK3_NPX))

# Calculate overall correlation
cor_overall <- cor.test(psa_protein_data$logPSA, psa_protein_data$KLK3_NPX, method = "spearman")

# Print results
cat("=== Correlation between log(PSA) and KLK3 Protein Expression ===\n")
cat("All mHSPC patients (not just paired samples)\n\n")
cat("Overall (Pre + Post combined):\n")
cat("  Spearman rho =", round(cor_overall$estimate, 3), "\n")
cat("  P-value =", formatC(cor_overall$p.value, format = "e", digits = 2), "\n")
cat("  N =", nrow(psa_protein_data), "\n\n")


# Scatter plot
p_psa_klk3 <- ggplot(psa_protein_data, aes(x = KLK3_NPX, y = logPSA)) +
  geom_point(size = 5, alpha = 0.7) +
  geom_smooth(method = "lm", se = TRUE) +
  #scale_color_manual(values = c("#ff7f0e", "gold4")) +
  labs(
    title = sprintf("log(PSA) vs KLK3 Expression (rho = %.2f, p = %.2e)",
                    cor_overall$estimate, cor_overall$p.value),
    x = "KLK3 NPX",
    y = "log10(PSA)",
    color = "Treatment Status"
  ) +
  theme_bw() +
  theme_custom

# Save plot
ggsave(p_psa_klk3, file = file.path(wd$outCurr, "logPSA_vs_KLK3_allSamples.pdf"), height = 4, width = 6)
print(p_psa_klk3)
```

Lets do it per cohort

```{r}
cor_by_cohort <- psa_protein_data %>%
  group_by(cohort) %>%
  summarise(
    rho = cor(KLK3_NPX, logPSA, method = "spearman", use = "complete.obs"),
    p = cor.test(KLK3_NPX, logPSA, method = "spearman")$p.value
  ) %>%
  mutate(
    label = sprintf("rho = %.2f\np = %.2e", rho, p)
  )


p_psa_klk3.1 <- ggplot(psa_protein_data, aes(x = KLK3_NPX, y = logPSA, fill=cohort)) +
  geom_point(size = 5, alpha = 0.7, shape=21) +
  geom_smooth(method = "lm", se = TRUE) +
  scale_fill_manual(values=pal.cohort.n) +
  facet_wrap(~cohort) +
  geom_text(
    data = cor_by_cohort,
    aes(label = label),
    x = Inf, y = Inf,
    hjust = 1.1, vjust = 1.2,
    inherit.aes = FALSE
  ) +
  labs(
    title = "log10(PSA) vs KLK3 Expression by Cohort",
    x = "KLK3 NPX",
    y = "log10(PSA)"
  ) +
  theme_bw() +
  theme_custom

# Save plot
ggsave(p_psa_klk3.1, file = file.path(wd$outCurr, "logPSA_vs_KLK3_perCohort.pdf"), height = 4, width = 12)
print(p_psa_klk3.1)

```



## Per-cohort

Local state. Lets do the local state, pre-and post treatment

```{r}
df_plot_a <- psa_protein_data %>% 
  dplyr::filter(cohort == "A") %>% 
  dplyr::mutate(glsn_grp = case_when(glsn_diag >=8 ~ ">8",
                                     glsn_diag <8 ~ "<8"))

p_psa_a1 <- 
  df_plot_a %>%
  ggplot(aes(x = txt_stat, y = KLK3_NPX, fill = glsn_grp)) +
  geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(outlier.alpha = 0,  width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  scale_fill_manual(values=c("lightblue", "orange")) +
  NULL

p_psa_a2 <- 
  df_plot_a %>%
  ggplot(aes(x = txt_stat, y = logPSA, fill = glsn_grp)) +
  geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(outlier.alpha = 0,  width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  scale_fill_manual(values=c("lightblue", "orange")) +
  NULL

p_psa_a1 + p_psa_a2
```


Lets do mHSPC group

```{r}
df_plot_b <- psa_protein_data %>% 
  dplyr::filter(cohort == "B") %>% 
  mutate(txt_stat = factor(txt_stat, levels = c("Pre", "Post1")))

p_psa_b1 <- 
  df_plot_b %>%
  ggplot(aes(x = txt_stat, y = KLK3_NPX, fill = txt_stat)) +
  geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(outlier.alpha = 0,  width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  scale_fill_manual(values=c("#ff7f0e", "gold4")) +
  geom_signif(
    comparisons = list(c("Pre", "Post1")),
    map_signif_level = FALSE,
    test = "t.test") +
  NULL

p_psa_b2 <- 
  df_plot_b %>%
  ggplot(aes(x = txt_stat, y = logPSA, fill = txt_stat)) +
  geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(outlier.alpha = 0,  width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  scale_fill_manual(values=c("#ff7f0e", "gold4")) +
    geom_signif(
    comparisons = list(c("Pre", "Post1")),
    map_signif_level = FALSE,
    test = "t.test") +
  NULL

p_psa_b1 + p_psa_b2


ggsave(p_psa_b1 + p_psa_b2, file = file.path(wd$outCurr, "logPSA_vs_KLK3_mHSPCt.pdf"), height = 4, width = 8)
```
