==> scripts/05_mHSPC_analysis.Rmd <==
---
title: "mHPSC analysis"
date: "2025-07-01"
output: 
  html_document:
    toc: true
    toc_float:
      collapsed: true
    toc_depth: 3
    number_sections: false
    theme: lumen
---

# Background

We bridge the different batches using limma. Now we want to focus the analysis on the mHSPC cohort.

# Objectives

1. Treatment effect on mHSPC cohort
2. Proteins associated with progression
3. Association with metastatic load
4. Baseline protein expression correlated with metastatic load


# Conclusion


# Pre-processing 

## Loading packages

```{r, message=FALSE, warning=FALSE}
library(here)
library(tidyverse)
library(readxl)   
library(readr)
library(stringr)
library(ggplot2)
library(paletteer)
library(ggvenn)
library(ggbeeswarm)
library(ggsignif)
library(patchwork)
# DE
library(limma)
library(survival)
library(survminer)
library(dplyr)
```

## Directories

```{r}
wd <- list()
wd$main <- here()
wd$data <- file.path(wd$main, "data")
wd$manishData <- file.path(wd$data, "raw/data_fromManish")
wd$d2024 <- file.path(wd$data, "updated_2024")
wd$d2024_npx <- file.path(wd$d2024, "NPX_data")
wd$output <- file.path(wd$main, "output")
wd$script <- file.path(wd$main, "scripts")

wd$outData <- file.path(wd$output, "data")
wd$outCurr <- file.path(wd$output, "05_mHSPC_analysis")
```


Create directories

```{r}
if (!file.exists(wd$outCurr)) {
  dir.create(wd$outCurr)
} else {
  print("Directory already exists")
}
```

Load functions

```{r}
source(file.path(wd$script, "functions.R"))
```

## Load data

Load the processed metadata file

```{r}
load(file.path(wd$outData, "01_Metadata.Rdata"))
load(file.path(wd$outData, "02_data.Rdata"))
load(file.path(wd$outData, "03_data.Rdata"))

# Print censor date
censor_date
```

# Obj 1: Txt effect

We first identify the paired samples in the mHSPC cohort.

```{r}
mrn_pre_post <- df_meta_f5 %>%
  filter(cohort == "B") %>%
  group_by(mrn) %>%
  summarise(
    has_pre = as.integer(any(txt_stat == "Pre", na.rm = TRUE)),
    has_post = as.integer(any(txt_stat == "Post", na.rm = TRUE))
  )

# View only those with both Pre and Post
paired_mrn <- mrn_pre_post %>%
  filter(has_pre == 1 & has_post == 1)
dim(paired_mrn)
```

We have 25 paired samples. Lets get the median time between pre and post.

```{r}
med_time <- df_meta_f5 %>%
  filter(mrn %in% paired_mrn$mrn) %>%
  group_by(mrn, txt_stat) %>%
  summarise(median_time = median(time2txt, na.rm = TRUE)) %>%
  pivot_wider(names_from = txt_stat, values_from = median_time) %>%
  mutate(diff = Post - Pre) %>%
  arrange(desc(diff))
to_exclude <- med_time %>% filter(diff >200) %>% pull(mrn) %>% unique()
med_time2 <- med_time %>% filter(!mrn %in% to_exclude)
```

Exclude samples with a median time between pre and post of more than 200 days.

```{r}
df_b_paired<- df_meta_f5 %>%
  filter(mrn %in% paired_mrn$mrn) %>%
  filter(!mrn %in% to_exclude)
```

## PSA

Lets plot the PSA as well of pre- and post-treatment. We noted there was one outlier, so lets log10 the PSA value

```{r}
# df_b_final - contain duplicated 2x pre samples / 2x post samples
df_plot2 <- df_b_final %>%
  filter(HCI_cID %in% df_meta_f5$HCI_cID) %>% 
  filter(mrn %in% df_b_paired$mrn) %>%
  rename(txt_stat = updated_adt_early) %>% 
  mutate(txt_stat = factor(txt_stat, levels = c("Pre", "Post")),
         logPSA = log10(psa)) 
  
# Create the plot with boxplot for txt_stat group
p_box_psa <-
ggplot(df_plot2, aes(x = txt_stat, y = logPSA, group = mrn)) +
  geom_boxplot(aes(group = txt_stat, fill = txt_stat), outlier.alpha = 0, width = 0.3, coef = 0, position = position_dodge(width = 0.9), alpha = 0.5) +
  geom_line(aes(group = mrn), color = "grey40", alpha = 0.7, position = position_dodge(width = 0.0)) +
  geom_point(aes(color = txt_stat), size = 2) +
  theme_bw() +
  labs(title = "PSA Pre and Post for Each Patient", x = "Treatment Status", y = "log10(PSA)") +
  theme(legend.position = "none") +
  scale_fill_manual(values=c("#ff7f0e", "gold4")) +
  scale_colour_manual(values=c("#ff7f0e", "gold4")) +
  theme_custom +
  NULL
p_box_psa

psa_df <- df_plot2 %>%
  dplyr::select(mrn, txt_stat, psa) %>% 
  group_by(mrn, txt_stat) %>%
  pivot_wider(names_from = txt_stat, values_from = psa) %>%
  mutate(diff = Post - Pre) %>%
  arrange(desc(diff))

psa_up <- psa_df %>% filter(diff > 0)

# --- New: Violin/box/quasirandom plot for PSA (Pre/Post) --- #

max_psa <- max(df_plot2$logPSA, na.rm = TRUE)
min_psa <- min(df_plot2$logPSA, na.rm = TRUE)

p_viol_psa <-
  df_plot2 %>%
  ggplot(aes(x = txt_stat, y = logPSA, fill = txt_stat)) +
  geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(outlier.alpha = 0, width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  scale_y_continuous(limits=c(min_psa, max_psa+0.5)) +
  labs(title = "PSA Pre and Post (Violin)", x = "Treatment Status", y = "log10(PSA)") +
  scale_fill_manual(values=c("#ff7f0e", "gold4")) +
  theme_custom +
  geom_signif(
    comparisons = list(c("Pre", "Post")),
    map_signif_level = FALSE,
    test = "t.test") +
  NULL
p_viol_psa

ggsave(p_box_psa, file=file.path(wd$outCurr, "PSA_boxplot.pdf"), height = 5, width = 5)
ggsave(p_viol_psa, file=file.path(wd$outCurr, "PSA_violin.pdf"), height = 5, width = 5)
```
Record those with decreased PSA

```{r}
df_b_paired2 <- df_b_paired %>% 
  filter(!mrn %in% psa_up$mrn)
```



## Proteomics

Now lets run limma to identify differentially expressed genes between pre and post in these samples

```{r}
# Generate the expression matrix
mat_in <- exprMatrix_corrected_sub 
# We also want to limit the proteins to proteins that were specific in mHSPC cohort
b_stage_pro <- overlap_df_limma %>% 
  filter(Category %in% c("A and B", "B", "B and CD")) %>% 
  pull(OlinkID)

exp_mat_sub <- mat_in[,df_b_paired2$final_sample_id]
exp_mat_sub <- exp_mat_sub[b_stage_pro,]

identical(colnames(exp_mat_sub), df_b_paired2$final_sample_id)

# Create design matrix (no intercept, just Pre/Post)
design <- model.matrix(~0 + factor(df_b_paired2$txt_stat, levels=c("Pre","Post")))
colnames(design) <- c("Pre","Post")

# Patient/blocking factor
block <- factor(df_b_paired2$mrn)

# Estimate correlation between pairs
corfit <- duplicateCorrelation(exp_mat_sub, design, block=block)

# Fit the model with blocking
fit <- lmFit(exp_mat_sub, design, block=block, correlation=corfit$consensus)

# Create contrast matrix
contrast.matrix <- makeContrasts(Post - Pre, levels=design)

# Run limma
fit2 <- contrasts.fit(fit, contrast.matrix)
fit2 <- eBayes(fit2)

# Get top differentially expressed genes
limma_b_pair <- topTable(fit2, number = Inf, sort.by = "none") %>% 
  tibble::rownames_to_column(var = "Gene") %>% 
  rename(OlinkID = Gene) %>% 
  left_join(df_id_pro) %>% 
  rename(Adjusted_pval = adj.P.Val,
         log2FC = logFC) %>% 
  # For table purpose keep it clean (round the number)
  mutate(log2FC = round(log2FC, 2),
         P.Value = as.numeric(sprintf("%.1e", P.Value)),
         Adjusted_pval = as.numeric(sprintf("%.1e", Adjusted_pval)))

# Check numner of significant proteins
limma_b_pair_sig <- limma_b_pair %>%
  filter(
    Adjusted_pval <= 0.05)
dim(limma_b_pair_sig)
# Save the results
save(limma_b_pair, file = file.path(wd$outCurr, "limma_b_pair.Rdata"))
```

Draw a violin plot of these significant proteins

```{r}
# Prepare long-format NPX data for plotting
df_npx_long <- exp_mat_sub %>%
  as.data.frame() %>% 
  tibble::rownames_to_column(var = "OlinkID") %>% 
  pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX")

# Filter for b_stage_pro proteins
sig_olink_ids <- limma_b_pair_sig

# Arrange limma_b_pair_sig by descending log2FC and create a factor for Assay with desired order
limma_b_pair_sig <- limma_b_pair_sig %>%
  arrange((log2FC)) %>%
  mutate(
    Assay = factor(Assay, levels = unique(Assay)),
    OlinkID = factor(OlinkID, levels = OlinkID) # OlinkID in descending logFC order
  )

# Merge and filter for plotting, and add logFC/adjp for annotation
df_plot <- df_npx_long %>% 
  left_join(df_id_pro, by = "OlinkID") %>% 
  dplyr::filter(OlinkID %in% sig_olink_ids$OlinkID) %>% 
  left_join(df_final_ids %>% select(final_sample_id, cohort, txt_stat), by = "final_sample_id") %>% 
  filter(final_sample_id %in% colnames(exp_mat_sub)) %>%
  left_join(limma_b_pair_sig %>% select(OlinkID, log2FC, Adjusted_pval), by = "OlinkID") %>%
  mutate(
    # For facetting, order by OlinkID factor (descending logFC)
    OlinkID = factor(OlinkID, levels = limma_b_pair_sig$OlinkID),
    # For facetting, use Assay as label, but add logFC and adjp to label
    facet_label = paste0(
      Assay, "\nlogFC=", log2FC, "\nadjP=", formatC(Adjusted_pval, format="e", digits=1)
    )
  ) %>% arrange((OlinkID))
fc_lvl <- unique(df_plot$facet_label)
df_plot <- df_plot %>% 
  mutate(facet_label = factor(df_plot$facet_label, levels=fc_lvl))

# Ensure txt_stat is a factor with Pre first, then Post
df_plot$txt_stat <- factor(df_plot$txt_stat, levels = c("Pre", "Post"))

# Show the plot
max_y <- max(df_plot$NPX, na.rm = TRUE)
min_y <- min(df_plot$NPX, na.rm = TRUE)

p_viol_gene <- 
  df_plot %>%
  ggplot(aes(x = txt_stat, y = NPX, fill = txt_stat)) +
  geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(outlier.alpha = 0,  width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  #scale_y_continuous(limits=c(min_y, max_y+3)) +
  facet_wrap(~facet_label, ncol=nrow(limma_b_pair_sig)/2, scales = "free_y") +
  scale_fill_manual(values=c("#ff7f0e", "gold4")) +
  NULL
p_viol_gene
ggsave(p_viol_gene, file=file.path(wd$outCurr, "PrePost_violin.pdf"), height = 6, width = 10)
```

Show as heatmap

```{r}
# Wide matrix: Assay x sample
heatmap_data <- df_plot %>%
  select(Assay, final_sample_id, NPX) %>%
  pivot_wider(names_from = final_sample_id, values_from = NPX) %>%
  column_to_rownames("Assay") %>%
  as.matrix()

annotation_col <- df_plot %>%
  select(final_sample_id, txt_stat) %>%
  distinct() %>%
  column_to_rownames("final_sample_id") %>%
  rename(Group = txt_stat)

annotation_col <- annotation_col[colnames(heatmap_data), , drop = FALSE]

annotation_colors <- list(
  Group = c(
    Pre  = "#ff7f0e",
    Post = "gold4"
  )
)


scale_data <- heatmap_data

pheatmap(
    heatmap_data,
    annotation_col = annotation_col,
    annotation_colors = annotation_colors,
    show_rownames = TRUE,
    show_colnames = TRUE,
    clustering_distance_rows = "euclidean",
    clustering_distance_cols = "euclidean",
    clustering_method = "complete",
    scale = "row",
    color = colorRampPalette(c("blue", "white", "red"))(100),
    main = paste0(
      "Pre vs Post Treatment"),
    fontsize_row = 8,
    fontsize_col = 8
  )

# Slight edit

library(dendextend)

col_dist <- dist(t(heatmap_data), method = "euclidean")
col_hclust <- hclust(col_dist, method = "complete")

col_dend <- as.dendrogram(col_hclust)

desired_order <- rownames(annotation_col)[
  order(annotation_col$Group)
]

col_dend <- rotate(col_dend, desired_order)


# Desired order: Pre first, then Post
desired_order <- rownames(annotation_col)[
  order(annotation_col$Group)
]

# Rotate dendrogram
col_dend <- rotate(col_dend, desired_order)

col_hclust_rot <- as.hclust(col_dend)


# plot
pheatmap(
  heatmap_data,
  annotation_col = annotation_col,
  annotation_colors = annotation_colors,
  cluster_cols = col_hclust_rot,   # <-- key line
  show_rownames = TRUE,
  show_colnames = FALSE,
  clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean",
  clustering_method = "complete",
  scale = "row",
  color = rev(RColorBrewer::brewer.pal(name = "RdBu", n = 11)),
  main = "Pre vs Post Treatment",
  fontsize_row = 8,
  fontsize_col = 8
)

pdf(file.path(wd$outCurr, "Pre_post_heatmap.pdf"), height = 4, width = 9)
pheatmap(
  heatmap_data,
  annotation_col = annotation_col,
  annotation_colors = annotation_colors,
  cluster_cols = col_hclust_rot,   # <-- key line
  show_rownames = TRUE,
  show_colnames = FALSE,
  clustering_distance_rows = "euclidean",
  clustering_distance_cols = "euclidean",
  clustering_method = "complete",
  scale = "row",
  color = rev(RColorBrewer::brewer.pal(name = "RdBu", n = 11)),
  main = "Pre vs Post Treatment",
  fontsize_row = 8,
  fontsize_col = 8
)
dev.off()
```



## Correlation

### Full mHSPC

In all samples

```{r}
# Get KLK3 OlinkID
klk3_id <- df_id_pro %>%
  filter(Assay == "KLK3") %>%
  pull(OlinkID)

# Link HCI_cID to final_sample_id for all mHSPC patients
df_link <- df_meta_f5 %>%
  select(HCI_cID, txt_stat_ordered) %>%
  left_join(df_final_ids %>% select(HCI_cID, final_sample_id))

# Use all mHSPC patients (df_b_final), not just paired
psa_protein_data <- df_b_final %>%
  filter(HCI_cID %in% df_meta_f5$HCI_cID) %>%
  mutate(logPSA = log10(psa),
         txt_stat = factor(updated_adt_early, levels = c("Pre", "Post"))) %>%
  select(HCI_cID, mrn, logPSA, txt_stat, psa) %>%
  left_join(df_link, by = "HCI_cID") %>%
  filter(!is.na(final_sample_id)) %>%
  filter(final_sample_id %in% colnames(exprMatrix_corrected_sub))

# Extract KLK3 expression for these samples
klk3_expression <- exprMatrix_corrected_sub[klk3_id, psa_protein_data$final_sample_id]
psa_protein_data$KLK3_NPX <- as.numeric(klk3_expression)

# Remove any rows with missing values
psa_protein_data <- psa_protein_data %>% filter(!is.na(logPSA), !is.na(KLK3_NPX))

# Calculate overall correlation
cor_overall <- cor.test(psa_protein_data$logPSA, psa_protein_data$KLK3_NPX, method = "spearman")

# Calculate correlation by treatment status
cor_pre <- cor.test(
  psa_protein_data %>% filter(txt_stat == "Pre") %>% pull(logPSA),
  psa_protein_data %>% filter(txt_stat == "Pre") %>% pull(KLK3_NPX),
  method = "spearman"
)

cor_post <- cor.test(
  psa_protein_data %>% filter(txt_stat == "Post") %>% pull(logPSA),
  psa_protein_data %>% filter(txt_stat == "Post") %>% pull(KLK3_NPX),
  method = "spearman"
)

# Print results
cat("=== Correlation between log(PSA) and KLK3 Protein Expression ===\n")
cat("All mHSPC patients (not just paired samples)\n\n")
cat("Overall (Pre + Post combined):\n")
cat("  Spearman rho =", round(cor_overall$estimate, 3), "\n")
cat("  P-value =", formatC(cor_overall$p.value, format = "e", digits = 2), "\n")
cat("  N =", nrow(psa_protein_data), "\n\n")

cat("Pre-treatment only:\n")
cat("  Spearman rho =", round(cor_pre$estimate, 3), "\n")
cat("  P-value =", formatC(cor_pre$p.value, format = "e", digits = 2), "\n")
cat("  N =", sum(psa_protein_data$txt_stat == "Pre"), "\n\n")

cat("Post-treatment only:\n")
cat("  Spearman rho =", round(cor_post$estimate, 3), "\n")
cat("  P-value =", formatC(cor_post$p.value, format = "e", digits = 2), "\n")
cat("  N =", sum(psa_protein_data$txt_stat == "Post"), "\n")

# Scatter plot
p_psa_klk3 <- ggplot(psa_protein_data, aes(x = KLK3_NPX, y = logPSA)) +
  geom_point(size = 5, alpha = 0.7) +
  geom_smooth(method = "lm", se = TRUE) +
  #scale_color_manual(values = c("#ff7f0e", "gold4")) +
  labs(
    title = sprintf("log(PSA) vs KLK3 Expression (rho = %.2f, p = %.2e)",
                    cor_overall$estimate, cor_overall$p.value),
    x = "KLK3 NPX",
    y = "log10(PSA)",
    color = "Treatment Status"
  ) +
  theme_bw() +
  theme_custom

# Save plot
ggsave(p_psa_klk3, file = file.path(wd$outCurr, "logPSA_vs_KLK3_allSamples.pdf"), height = 4, width = 6)
print(p_psa_klk3)
```

### Paired

We correlate PSA with KLK3 expression in the paried samples

```{r}
# Get KLK3 OlinkID
klk3_id <- df_id_pro %>%
  filter(Assay == "KLK3") %>%
  pull(OlinkID)

# Link HCI_cID to final_sample_id
df_link <- df_meta_f5 %>%
  select(HCI_cID, txt_stat_ordered) %>%
  left_join(df_final_ids %>% select(HCI_cID, final_sample_id))

# Merge PSA data with protein expression
# df_plot2 has HCI_cID and logPSA
# exprMatrix_corrected_sub has protein expression with final_sample_id as column names
psa_protein_data <- df_plot2 %>%
  select(HCI_cID, logPSA, txt_stat) %>%
  left_join(df_link, by = "HCI_cID") %>%
  filter(!is.na(final_sample_id)) %>%
  filter(final_sample_id %in% colnames(exprMatrix_corrected_sub))

# Extract KLK3 expression for these samples
klk3_expression <- exprMatrix_corrected_sub[klk3_id, psa_protein_data$final_sample_id]
psa_protein_data$KLK3_NPX <- as.numeric(klk3_expression)

# Remove any rows with missing values
psa_protein_data <- psa_protein_data %>% filter(!is.na(logPSA), !is.na(KLK3_NPX))

# Calculate overall correlation
cor_overall <- cor.test(psa_protein_data$logPSA, psa_protein_data$KLK3_NPX, method = "spearman")

# Calculate correlation by treatment status
cor_pre <- cor.test(
  psa_protein_data %>% filter(txt_stat == "Pre") %>% pull(logPSA),
  psa_protein_data %>% filter(txt_stat == "Pre") %>% pull(KLK3_NPX),
  method = "spearman"
)

cor_post <- cor.test(
  psa_protein_data %>% filter(txt_stat == "Post") %>% pull(logPSA),
  psa_protein_data %>% filter(txt_stat == "Post") %>% pull(KLK3_NPX),
  method = "spearman"
)

# Print results
cat("=== Correlation between log(PSA) and KLK3 Protein Expression ===\n\n")
cat("Overall (Pre + Post combined):\n")
cat("  Spearman rho =", round(cor_overall$estimate, 3), "\n")
cat("  P-value =", formatC(cor_overall$p.value, format = "e", digits = 2), "\n")
cat("  N =", nrow(psa_protein_data), "\n\n")

cat("Pre-treatment only:\n")
cat("  Spearman rho =", round(cor_pre$estimate, 3), "\n")
cat("  P-value =", formatC(cor_pre$p.value, format = "e", digits = 2), "\n")
cat("  N =", sum(psa_protein_data$txt_stat == "Pre"), "\n\n")

cat("Post-treatment only:\n")
cat("  Spearman rho =", round(cor_post$estimate, 3), "\n")
cat("  P-value =", formatC(cor_post$p.value, format = "e", digits = 2), "\n")
cat("  N =", sum(psa_protein_data$txt_stat == "Post"), "\n")

# Scatter plot
p_psa_klk3 <- ggplot(psa_protein_data, aes(x = KLK3_NPX, y = logPSA)) +
  geom_point(size = 3, alpha = 0.7) +
  geom_smooth(method = "lm", se = TRUE) +
  #scale_color_manual(values = c("#ff7f0e", "gold4")) +
  labs(
    title = sprintf("log(PSA) vs KLK3 Expression (rho = %.2f, p = %.2e)",
                    cor_overall$estimate, cor_overall$p.value),
    x = "KLK3 NPX",
    y = "log10(PSA)",
    color = "Treatment Status"
  ) +
  theme_bw() +
  theme_custom

# Save plot
ggsave(p_psa_klk3, file = file.path(wd$outCurr, "logPSA_vs_KLK3.pdf"), height = 6, width = 8)
print(p_psa_klk3)
```




Now plot KLK3 expression Pre vs Post for each patient (same style as PSA plot above).

```{r}
# Prepare data for KLK3 boxplot (matching df_plot2 structure)
# Add mrn to psa_protein_data
df_plot_klk3 <- psa_protein_data %>%
  left_join(df_b_final %>% select(HCI_cID, mrn), by = "HCI_cID") %>%
  filter(!is.na(mrn))

# Create the plot with boxplot for txt_stat group (matching PSA plot style)
p_box_klk3 <-
  ggplot(df_plot_klk3, aes(x = txt_stat, y = KLK3_NPX, group = mrn)) +
  geom_boxplot(aes(group = txt_stat, fill = txt_stat), outlier.alpha = 0, width = 0.3, coef = 0,
               position = position_dodge(width = 0.9), alpha = 0.5) +
  geom_line(aes(group = mrn), color = "grey40", alpha = 0.7, position = position_dodge(width = 0.0)) +
  geom_point(aes(color = txt_stat), size = 2) +
  theme_bw() +
  labs(title = "KLK3 Pre and Post for Each Patient", x = "Treatment Status", y = "KLK3 NPX") +
  theme(legend.position = "none") +
  scale_fill_manual(values = c("#ff7f0e", "gold4")) +
  scale_colour_manual(values = c("#ff7f0e", "gold4")) +
  theme_custom +
  geom_signif(
    comparisons = list(c("Pre", "Post")),
    map_signif_level = FALSE,
    test = "t.test") +
  NULL

p_box_klk3

# Save plot
pp <- p_box_psa + p_box_klk3
ggsave(pp, file = file.path(wd$outCurr, "KLK3_boxplot.pdf"), height = 4, width = 9)
```



# Obj 2: Progression

## Time to CRPC

Lets examine if the protein levels is associated with the disease progression. 

First lets define the time to CRPC. We define the time to CRPC as the earliest date of either the time of 

- CRPC biochemical progression (crpc_biochem_date)
- nmCRPC treatment date (txt_nmcrpc_1st_date)
- mCRPC treatment date (txt_mcrpc_1st_date)
- mCRPC treatment date (txt_mcrpc_2nd_date)

to the time of ADT treatment (adt_earliest_date). Lets calculate the time to event.

## Baseline

For this analysis, we want to use all samples that is at baseline, not samples that was treated with ADT. Also we added censoring date for patients that did not progress to CRPC.

```{r}
df_tmp1 <- df_meta_f5 %>% 
  select(HCI_cID, txt_stat_ordered)

df_tmp2 <- df_final_ids %>% 
  select(HCI_cID, final_sample_id)

df_tmp <- left_join(df_tmp1, df_tmp2)

df_b_cox <- df_b_final %>% 
  left_join(df_tmp, by = "HCI_cID") %>% 
  filter(txt_stat_ordered == "Pre1") %>% 
  filter(HCI_cID %in% df_meta_f5$HCI_cID) %>% 
  select(mrn, HCI_cID, progressed_to_crpc, adt_earliest_date, crpc_earliest_date, final_sample_id, psa) %>% 
  rename(SampleID = final_sample_id) %>% 
  mutate(sample_id = SampleID) %>% 
  arrange(adt_earliest_date) %>% 
  mutate(
    # If patient progressed, use crpc_earliest_date; otherwise, censor at censor_date
    event_date = dplyr::case_when(
      progressed_to_crpc == "y" & !is.na(crpc_earliest_date) ~ crpc_earliest_date,
      TRUE ~ as.Date(censor_date)
    ),
    time_to_crpc = as.numeric(difftime(event_date, adt_earliest_date, units = "days")),
    progressed_to_crpc = dplyr::case_when(
      progressed_to_crpc == "y" ~ 1L,
      is.na(progressed_to_crpc) | progressed_to_crpc != "y" ~ 0L
    )
  )
```

Save the file so we can discuss with Manish

```{r}
df_b_cox <- df_b_cox %>% arrange(desc(time_to_crpc))
write.csv(df_b_cox, file.path(wd$outCurr, "mHSPC_clin_info.csv"))
```

## Analysis Strategy: Full vs 5-year Censored Follow-up

We'll run the progression analysis using two approaches:
1. **Full follow-up**: Uses all available follow-up data
2. **5-year censored**: Administrative censoring at 5 years for clinical relevance

```{r}
df_b_cox <- df_b_cox %>% 
  filter(!is.na(adt_earliest_date))
# Create two datasets for comparison
# 1. Full follow-up (all data)
df_b_cox_full <- df_b_cox

# 2. 5-year administrative censoring
df_b_cox_5yr <- df_b_cox %>%
    mutate(
      # First check if original time > 1825, then cap and censor
      censored_at_5yr = time_to_crpc > 1825,
      time_to_crpc = pmin(time_to_crpc, 1825, na.rm = TRUE),
      progressed_to_crpc = ifelse(censored_at_5yr, 0, progressed_to_crpc)
    )

# Print summary of both datasets
cat("=== Full Follow-up Dataset ===\n")
cat("Time to CRPC summary:\n")
print(summary(df_b_cox_full$time_to_crpc))
cat("Progression events:", sum(df_b_cox_full$progressed_to_crpc, na.rm = TRUE), "\n")

cat("\n=== 5-Year Censored Dataset ===\n") 
cat("Time to CRPC summary:\n")
print(summary(df_b_cox_5yr$time_to_crpc))
cat("Progression events:", sum(df_b_cox_5yr$progressed_to_crpc, na.rm = TRUE), "\n")
```
Get the median time of progression in the patients that progressed

```{r}
df_b_cox %>% filter(progressed_to_crpc==1) %>% pull(time_to_crpc) %>% summary()
```

Now we'll run the protein progression analysis on both datasets.

### Full Follow-up Analysis

```{r}
# Prepare data for full follow-up analysis
meta_in_full <- df_b_cox_full %>% 
  rename(OS = time_to_crpc,
         death = progressed_to_crpc)

prot_ids <- limma_b_pair_sig %>% pull(OlinkID) %>% as.character()
# Define the expression matrix
mat_in <- exprMatrix_corrected_sub 
# Subset to the x proteins and the samples that are at baseline
mat_in_sub_full <- mat_in[prot_ids, df_b_cox_full$SampleID]

# Convert to long format
exprMatrix_long_full <- mat_in_sub_full %>% 
  as.data.frame() %>%
  rownames_to_column(var = "OlinkID") %>%  
  pivot_longer(cols = -OlinkID, names_to = "SampleID", values_to = "Value") %>% 
  rename(NPX = Value)

# Run analysis
tmp_npx_full <- exprMatrix_long_full 
results_base_full <- analyze_proteins_highest(prot_ids, npx_in=tmp_npx_full, meta_in=meta_in_full) %>% 
  left_join(df_id_pro) %>% 
  arrange(desc(HR)) %>% 
  mutate(P_Value_FDR = p.adjust(P_Value, method = "fdr"))
write.csv(results_base_full, file.path(wd$outCurr, "HR_values_mHSPC_univariate_pre_full.csv"))
```

### 5-Year Censored Analysis

```{r}
# Prepare data for 5-year censored analysis
meta_in_5yr <- df_b_cox_5yr %>% 
  rename(OS = time_to_crpc,
         death = progressed_to_crpc) 

# Subset to the x proteins and the samples that are at baseline
mat_in_sub_5yr <- mat_in[prot_ids, df_b_cox_5yr$SampleID]

# Convert to long format
exprMatrix_long_5yr <- mat_in_sub_5yr %>% 
  as.data.frame() %>%
  rownames_to_column(var = "OlinkID") %>%  
  pivot_longer(cols = -OlinkID, names_to = "SampleID", values_to = "Value") %>% 
  rename(NPX = Value)

# Run analysis
tmp_npx_5yr <- exprMatrix_long_5yr 
results_base_5yr <- analyze_proteins_highest(prot_ids, npx_in=tmp_npx_5yr, meta_in=meta_in_5yr) %>% 
  left_join(df_id_pro) %>% 
  arrange(desc(HR)) %>% 
  mutate(P_Value_FDR = p.adjust(P_Value, method = "fdr"))
write.csv(results_base_5yr, file.path(wd$outCurr, "HR_values_mHSPC_univariate_pre_5yr.csv"))
```

### Comparison of Results

```{r}
# Compare the results between full and 5-year analyses
comparison_results <- results_base_full %>%
  select(OlinkID, Assay, HR_full = HR, P_Value_full = P_Value) %>%
  left_join(
    results_base_5yr %>% select(OlinkID, HR_5yr = HR, P_Value_5yr = P_Value),
    by = "OlinkID"
  ) %>%
  mutate(
    HR_diff = HR_5yr - HR_full,
    P_diff = P_Value_5yr - P_Value_full
  ) %>%
  arrange(desc(abs(HR_diff)))

# Print top differences
cat("=== Top proteins with largest HR differences (5yr vs full) ===\n")
print(head(comparison_results, 10))

write.csv(comparison_results, file.path(wd$outCurr, "HR_comparison_full_vs_5yr.csv"))
```

### Forest Plots: Full vs 5-Year Censored Analysis

Now let's create forest plots comparing both approaches with PSA.

```{r}
# PSA analysis for both datasets
clin_var <- c("psa")
results_psa_full <- cox_extract(clin_var, meta_in_full)
res1_full <- results_psa_full$psa

results_psa_5yr <- cox_extract(clin_var, meta_in_5yr)
res1_5yr <- results_psa_5yr$psa
```

#### Full Follow-up Forest Plot

```{r}
# Prepare data for full follow-up forest plot
res_all_full <- data.frame(
  OlinkID = c(res1_full$OlinkID, results_base_full$OlinkID),
  Lower_Bound = c(res1_full$Lower_Bound, results_base_full$Lower_Bound),
  Upper_Bound = c(res1_full$Upper_Bound, results_base_full$Upper_Bound),
  HR = c(res1_full$HR, results_base_full$HR),
  P_Value = c(res1_full$P_Value, results_base_full$P_Value)
) %>% 
  left_join(df_id_pro) %>% 
  mutate(Assay = case_when(is.na(Assay) & OlinkID == "psa" ~ "PSA",
                           TRUE ~ Assay))

df_full <- res_all_full %>%
  mutate(
    Assay = factor(Assay, levels=rev(Assay)),
    HR_CI = sprintf("%.2f [%.2f, %.2f]", HR, Lower_Bound, Upper_Bound),
    P_Text = ifelse(P_Value < 0.001, formatC(P_Value, format = "e", digits = 2), sprintf("%.3f", P_Value)),
    P_Color = ifelse(P_Value < 0.05, "red", "black")
  )

# Forest Plot for Full Follow-up
p1_full <- ggplot(df_full, aes(x = HR, y = Assay)) +
  geom_errorbarh(aes(xmin = Lower_Bound, xmax = Upper_Bound), height = 0.3, color = "grey50") +
  geom_point(color = "black", size = 3, alpha=1) +
  geom_vline(xintercept = 1, linetype = "dashed") +
  theme_minimal(base_size = 14) +
  scale_x_continuous(limits=c(0,max(df_full$Upper_Bound, na.rm = TRUE))) +
  labs(x = "Hazard Ratio", y = NULL, title = "Full Follow-up Analysis") +
  theme(
    axis.text.y = element_text(face = "bold"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank()
  )

p2_full <- ggplot(df_full, aes(y = Assay)) +
  geom_text(aes(x = 1, label = HR_CI), hjust = 0, size = 4) +
  geom_text(aes(x = 2, label = P_Text), hjust = 0, size = 4, show.legend = FALSE) +
  scale_color_identity() +
  scale_x_continuous(limits = c(0.95, 2.9), breaks = NULL) +
  theme_void(base_size = 14) +
  theme(
    axis.text.y = element_blank(),
    plot.margin = margin(5.5, 20, 5.5, 5.5)
  )

final_plot_full <- p1_full + p2_full + plot_layout(widths = c(2, 1))
final_plot_full
ggsave(filename=file.path(wd$outCurr, "Top_x_HRs_full.pdf"), device="pdf", width=11, height=8)
```

#### 5-Year Censored Forest Plot

```{r}
# Prepare data for 5-year censored forest plot
res_all_5yr <- data.frame(
  OlinkID = c(res1_5yr$OlinkID, results_base_5yr$OlinkID),
  Lower_Bound = c(res1_5yr$Lower_Bound, results_base_5yr$Lower_Bound),
  Upper_Bound = c(res1_5yr$Upper_Bound, results_base_5yr$Upper_Bound),
  HR = c(res1_5yr$HR, results_base_5yr$HR),
  P_Value = c(res1_5yr$P_Value, results_base_5yr$P_Value)
) %>% 
  left_join(df_id_pro) %>% 
  mutate(Assay = case_when(is.na(Assay) & OlinkID == "psa" ~ "PSA",
                           TRUE ~ Assay))

df_5yr <- res_all_5yr %>%
  mutate(
    Assay = factor(Assay, levels=rev(Assay)),
    HR_CI = sprintf("%.2f [%.2f, %.2f]", HR, Lower_Bound, Upper_Bound),
    P_Text = ifelse(P_Value < 0.001, formatC(P_Value, format = "e", digits = 2), sprintf("%.3f", P_Value)),
    P_Color = ifelse(P_Value < 0.05, "red", "black")
  )

# Forest Plot for 5-Year Censored
p1_5yr <- ggplot(df_5yr, aes(x = HR, y = Assay)) +
  geom_errorbarh(aes(xmin = Lower_Bound, xmax = Upper_Bound), height = 0.3, color = "grey50") +
  geom_point(color = "black", size = 3, alpha=1) +
  geom_vline(xintercept = 1, linetype = "dashed") +
  theme_minimal(base_size = 14) +
  scale_x_continuous(limits=c(0,max(df_5yr$Upper_Bound, na.rm = TRUE))) +
  labs(x = "Hazard Ratio", y = NULL, title = "5-Year Censored Analysis") +
  theme(
    axis.text.y = element_text(face = "bold"),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank()
  )

p2_5yr <- ggplot(df_5yr, aes(y = Assay)) +
  geom_text(aes(x = 1, label = HR_CI), hjust = 0, size = 4) +
  geom_text(aes(x = 2, label = P_Text), hjust = 0, size = 4, show.legend = FALSE) +
  scale_color_identity() +
  scale_x_continuous(limits = c(0.95, 2.9), breaks = NULL) +
  theme_void(base_size = 14) +
  theme(
    axis.text.y = element_blank(),
    plot.margin = margin(5.5, 20, 5.5, 5.5)
  )

final_plot_5yr <- p1_5yr + p2_5yr + plot_layout(widths = c(2, 1))
final_plot_5yr
ggsave(filename=file.path(wd$outCurr, "Top_x_HRs_5yr.pdf"), device="pdf", width=11, height=8)
```


### Baseline KM stats

Using all the baseline sample, we check how many patients prgressed to mCRPC

```{r}
table(meta_in_full$death)
table(meta_in_5yr$death)
```


Get now the summary of the progression timeline

```{r}
summary(meta_in_full$OS)
```

Now lets generate a KM plot stratifying the protein expression to high and low (based on median),

```{r,eval=FALSE}
protein_list <- unique(results_base_5yr$OlinkID)
km_plots <- list()

# Merge with clinical data to get time to CRPC and event status
df_base <- tmp_npx_5yr %>%
  left_join(meta_in_5yr %>% select(mrn, OS, death, SampleID)) %>%
  filter(!is.na(NPX), !is.na(OS), !is.na(death)) %>% 
  left_join(df_id_pro)# %>% 
  # Remove outlier of patients with too long of a follow up
  #filter(OS <= 2000)


for (prot in protein_list) {
  df_sub <- df_base %>% filter(OlinkID == prot)
  assay_name <- df_sub$Assay
  
  # Stratify by median NPX
  median_val <- median(df_sub$NPX, na.rm = TRUE)
  df_sub <- df_sub %>%
    mutate(
      NPX_bin = ifelse(NPX > median_val, "high", "low"),
      NPX_bin = factor(NPX_bin, levels = c("low", "high"))
    )
  
  # Create survival object
  surv_object <- Surv(time = df_sub$OS, event = df_sub$death)
  
  # Fit KM curve
  fit <- survfit(surv_object ~ NPX_bin, data = df_sub)
  
  # Plot
  p <- ggsurvplot(
    fit, data = df_sub, pval = TRUE, risk.table = TRUE,
    conf.int = TRUE, palette = c("#7F7F7F", "#e41a1c"),
    title = paste("KM plot for", prot, " (", assay_name, ")")
  )
  
  km_plots[[prot]] <- p
  # Optionally, save to file:
  # ggsave(filename = paste0("KM_", prot, ".pdf"), plot = p$plot)
}

# To display a plot, for example the first one:
print(km_plots[[1]]$plot)
```

Save significant plot

```{r,eval=FALSE}
pdf(file.path(wd$outCurr, "KM_progression.pdf"), width = 7, height = 6)
for (p in km_plots) {
  print(p)
}
dev.off()
```


## Delta NPX

Lets try an approach where we take the change in protein expression, then associate with CRPC progression. But now we need to limit to the 23 patients that we have pre- and post- measurement.

```{r,eval=FALSE}
# Calculate change in protein expression (delta NPX) for each patient and protein
# Assuming df_npx_long has columns: SampleID, PatientID, Timepoint (e.g., "Pre", "Post"), OlinkID, NPX

# First, pivot to wide format to get Pre and Post NPX side by side
df_npx_in <- df_npx_long %>% 
  filter(final_sample_id %in% df_b_paired2$final_sample_id) %>% 
  left_join(df_id_pro, by = "OlinkID") %>% 
  dplyr::filter(OlinkID %in% limma_b_pair_sig$OlinkID) %>% 
  left_join(df_final_ids %>% select(mrn, final_sample_id, cohort, txt_stat), by = "final_sample_id") 

df_npx_wide <- df_npx_in %>%
  pivot_wider(
    id_cols = c(mrn, OlinkID),
    names_from = txt_stat,
    values_from = NPX
  )

# Calculate delta NPX (Post - Pre)
df_npx_wide <- df_npx_wide %>%
  mutate(delta_NPX = Post - Pre) 

# Merge with clinical data to get time to CRPC and event status
# meta_in: Patient-level clinical data with columns: PatientID, OS (time), death (event)
df_delta <- df_npx_wide %>%
  left_join(meta_in %>% select(mrn, OS, death)) %>%
  filter(!is.na(delta_NPX), !is.na(OS), !is.na(death))

# # Run Cox model for each protein
# cox_results <- df_delta %>%
#   group_by(OlinkID) %>%
#   do({
#     fit <- coxph(Surv(OS, death) ~ delta_NPX, data = .)
#     tidy_fit <- broom::tidy(fit)
#     tidy_fit
#   }) %>%
#   ungroup()
# 
# # Show results
# cox_results


# Alternate result table
result_df <- df_delta %>%
  group_by(OlinkID) %>%
  do({
    fit <- tryCatch(coxph(Surv(OS, death) ~ delta_NPX, data = .), error = function(e) NULL)
    if (is.null(fit)) {
      # Return NA row if model failed
      data.frame(
        OlinkID = unique(.$OlinkID),
        Coefficient = NA_real_,
        Lower_Bound = NA_real_,
        Upper_Bound = NA_real_,
        P_Value = NA_real_,
        HR = NA_real_
      )
    } else {
      coef_summary <- summary(fit)$coefficients
      confint_summary <- summary(fit)$conf.int
      data.frame(
        OlinkID = unique(.$OlinkID),
        Coefficient = coef_summary[,"coef"],
        Lower_Bound = confint_summary[,"lower .95"],
        Upper_Bound = confint_summary[,"upper .95"],
        P_Value = coef_summary[,"Pr(>|z|)"],
        HR = confint_summary[,"exp(coef)"]
      )
    }
  }) %>%
  ungroup() %>% 
  left_join(df_id_pro) %>% 
  arrange(desc(HR))

write.csv(result_df, file.path(wd$outCurr, "HR_values_mHSPC_univariate_delta.csv"))
```

Lets run the cox regression on the change of PSA

```{r,eval=FALSE}
clin_var <- c("psa")
results <- cox_extract(clin_var, meta_in)

```

## KM plot

> We only have 25 patietns. So it seems like it does not make sense to do the KM. Error values are too large. 

Now lets generate a KM plot stratifying the protein expression to high and low (based on median),

```{r,eval=FALSE}
protein_list <- unique(df_delta$OlinkID)
km_plots <- list()

for (prot in protein_list) {
  df_sub <- df_delta %>% filter(OlinkID == prot)
  
  # Stratify by median delta_NPX
  median_val <- median(df_sub$delta_NPX, na.rm = TRUE)
  df_sub <- df_sub %>%
    mutate(
      delta_bin = ifelse(delta_NPX > median_val, "high", "low"),
      delta_bin = factor(delta_bin, levels = c("low", "high"))
    )
  
  # Create survival object
  surv_object <- Surv(time = df_sub$OS, event = df_sub$death)
  
  # Fit KM curve
  fit <- survfit(surv_object ~ delta_bin, data = df_sub)
  
  # Plot
  p <- ggsurvplot(
    fit, data = df_sub, pval = TRUE, risk.table = TRUE,
    conf.int = TRUE, palette = c("#7F7F7F", "#e41a1c"),
    title = paste("KM plot for", prot)
  )
  
  km_plots[[prot]] <- p
  # Optionally, save to file:
  # ggsave(filename = paste0("KM_", prot, ".pdf"), plot = p$plot)
}

# To display a plot, for example the first one:
print(km_plots[[1]]$plot)
```


# Obj 3: Metastasis load

## Survival analysis by metastatic load

We'll examine how metastatic load affects progression to CRPC using both full follow-up and 5-year censored approaches.

### Full Follow-up Analysis

```{r}
# Prepare survival data with metastatic load information (full follow-up)
df_survival_mets_full <- df_b_cox_full %>%
  left_join(df_b_final %>% select(HCI_cID, mets_volume), by = "HCI_cID") %>%
  filter(!is.na(mets_volume)) %>%
  mutate(
    mets_volume = factor(mets_volume, levels = c("Low", "High"))
  )

# Check sample sizes
cat("=== Full Follow-up: Sample sizes by metastatic load ===\n")
print(table(df_survival_mets_full$mets_volume))
```

```{r}
# Create survival object and fit KM curve (full follow-up)
surv_object_full <- Surv(time = df_survival_mets_full$time_to_crpc, event = df_survival_mets_full$progressed_to_crpc)
fit_mets_full <- survfit(surv_object_full ~ mets_volume, data = df_survival_mets_full)

# Generate KM plot (full follow-up)
km_plot_mets_full <- ggsurvplot(
  fit_mets_full, 
  data = df_survival_mets_full, 
  pval = TRUE, 
  risk.table = TRUE,
  conf.int = TRUE, 
  palette = c("#2E8B57", "#DC143C"),  # SeaGreen for Low, Crimson for High
  title = "Kaplan-Meier: Time to CRPC by Metastatic Load (Full Follow-up)",
  xlab = "Time to CRPC (days)",
  ylab = "Progression-free survival probability",
  legend.title = "Metastatic Load",
  legend.labs = c("Low", "High"),
  risk.table.height = 0.3
)

km_plot_mets_full$plot
pdf(file = file.path(wd$outCurr, "KM_metastatic_load_full.pdf"), width = 12, height = 10)
km_plot_mets_full
dev.off()
```

```{r}
# Cox proportional hazards model (full follow-up)
cox_mets_full <- coxph(Surv(time_to_crpc, progressed_to_crpc) ~ mets_volume, data = df_survival_mets_full)
summary(cox_mets_full)
# Get confidence intervals
ci_mets <- exp(confint(cox_mets_full))
print(ci_mets)

cox_summary_full <- summary(cox_mets_full)
hr_mets_full <- exp(coef(cox_mets_full))
ci_mets_full <- exp(confint(cox_mets_full))
p_mets_full <- cox_summary_full$coefficients[,"Pr(>|z|)"]

cat("=== Full Follow-up Results ===\n")
cat("Hazard Ratio (High vs Low metastatic load):", round(hr_mets_full, 2), "\n")
cat("95% CI:", round(ci_mets_full[1], 2), "-", round(ci_mets_full[2], 2), "\n")
cat("P-value:", formatC(p_mets_full, format = "e", digits = 3), "\n")
```

### 5-Year Censored Analysis

```{r}
# Prepare survival data with metastatic load information (5-year censored)
df_survival_mets_5yr <- df_b_cox_5yr %>%
  left_join(df_b_final %>% select(HCI_cID, mets_volume), by = "HCI_cID") %>%
  filter(!is.na(mets_volume)) %>%
  mutate(
    mets_volume = factor(mets_volume, levels = c("Low", "High"))
  )

# Check sample sizes
cat("=== 5-Year Censored: Sample sizes by metastatic load ===\n")
print(table(df_survival_mets_5yr$mets_volume))
```

```{r}
# Create survival object and fit KM curve (5-year censored)
df_survival_mets_5yr <- df_survival_mets_5yr %>% 
  mutate(time_to_crpc_mnth = time_to_crpc / 30.417)
surv_object_5yr <- Surv(time = df_survival_mets_5yr$time_to_crpc_mnth, event = df_survival_mets_5yr$progressed_to_crpc)
fit_mets_5yr <- survfit(surv_object_5yr ~ mets_volume, data = df_survival_mets_5yr)

# Generate KM plot (5-year censored)
km_plot_mets_5yr <- ggsurvplot(
  fit_mets_5yr, 
  data = df_survival_mets_5yr, 
  pval = TRUE, 
  risk.table = TRUE,
  conf.int = TRUE, 
  palette = c("#2E8B57", "#DC143C"),
  title = "Kaplan-Meier: Time to CRPC by Metastatic Load (5-Year Censored)",
  xlab = "Time to CRPC (months)",
  ylab = "Progression-free survival probability",
  legend.title = "Metastatic Load",
  legend.labs = c("Low", "High"),
  risk.table.height = 0.3
)

km_plot_mets_5yr$plot

pdf(file = file.path(wd$outCurr, "KM_metastatic_load_5yr.pdf"), width = 12, height = 10)
km_plot_mets_5yr
dev.off()
```

```{r}
# Cox proportional hazards model (5-year censored)
cox_mets_5yr <- coxph(Surv(time_to_crpc, progressed_to_crpc) ~ mets_volume, data = df_survival_mets_5yr)
summary(cox_mets_5yr)

cox_summary_5yr <- summary(cox_mets_5yr)
hr_mets_5yr <- exp(coef(cox_mets_5yr))
ci_mets_5yr <- exp(confint(cox_mets_5yr))
p_mets_5yr <- cox_summary_5yr$coefficients[,"Pr(>|z|)"]

cat("=== 5-Year Censored Results ===\n")
cat("Hazard Ratio (High vs Low metastatic load):", round(hr_mets_5yr, 2), "\n")
cat("95% CI:", round(ci_mets_5yr[1], 2), "-", round(ci_mets_5yr[2], 2), "\n")
cat("P-value:", formatC(p_mets_5yr, format = "e", digits = 3), "\n")
```

### Comparison of Metastatic Load Results

```{r}
# Compare results between full and 5-year analyses
mets_comparison <- data.frame(
  Analysis = c("Full Follow-up", "5-Year Censored"),
  HR = c(round(hr_mets_full, 2), round(hr_mets_5yr, 2)),
  CI_Lower = c(round(ci_mets_full[1], 2), round(ci_mets_5yr[1], 2)),
  CI_Upper = c(round(ci_mets_full[2], 2), round(ci_mets_5yr[2], 2)),
  P_Value = c(formatC(p_mets_full, format = "e", digits = 3), 
              formatC(p_mets_5yr, format = "e", digits = 3))
)

cat("=== Comparison: Metastatic Load Analysis ===\n")
print(mets_comparison)

write.csv(mets_comparison, file.path(wd$outCurr, "metastatic_load_comparison.csv"))
```

## Combined prognostic analysis: Metastatic load + Protein expression

Let's examine if combining metastatic load with protein expression improves prognostic accuracy using both full follow-up and 5-year censored approaches.

### Full Follow-up Combined Analysis

```{r}
# Use all 10 proteins from the treatment response analysis
top_proteins_full <- limma_b_pair_sig %>% pull(OlinkID)

if(length(top_proteins_full) > 0) {
  # Create combined analysis for each top protein (full follow-up)
  combined_results_full <- list()
  
  for(prot in top_proteins_full) {
    # Get protein expression data
    prot_data <- exprMatrix_long_full %>%
      filter(OlinkID == prot) %>%
      select(SampleID, NPX) %>%
      rename(protein_expr = NPX)
    
    # Merge with survival and metastatic load data
    combined_data <- df_survival_mets_full %>%
      left_join(prot_data, by = c("SampleID")) %>%
      filter(!is.na(protein_expr)) %>%
      mutate(
        # Dichotomize protein expression at median
        protein_high = ifelse(protein_expr > median(protein_expr, na.rm = TRUE), "High", "Low"),
        protein_high = factor(protein_high, levels = c("Low", "High")),
        # Create combined risk groups
        risk_group = case_when(
          mets_volume == "Low" & protein_high == "Low" ~ "Low-Low",
          mets_volume == "Low" & protein_high == "High" ~ "Low-High", 
          mets_volume == "High" & protein_high == "Low" ~ "High-Low",
          mets_volume == "High" & protein_high == "High" ~ "High-High"
        ),
        risk_group = factor(risk_group, levels = c("Low-Low", "Low-High", "High-Low", "High-High"))
      )
    
    # Cox models
    cox_combined <- coxph(Surv(time_to_crpc, progressed_to_crpc) ~ mets_volume + protein_high, 
                         data = combined_data)
    cox_interaction <- coxph(Surv(time_to_crpc, progressed_to_crpc) ~ mets_volume * protein_high, 
                           data = combined_data)
    
    # Store results
    protein_name <- df_id_pro %>% filter(OlinkID == prot) %>% pull(Assay)
    combined_results_full[[prot]] <- list(
      protein = prot,
      protein_name = protein_name,
      data = combined_data,
      cox_combined = cox_combined,
      cox_interaction = cox_interaction
    )
    
    # Print summary
    # cat("\n=== Full Follow-up Combined Analysis for", protein_name, "===\n")
    # cat("Model with metastatic load + protein expression:\n")
    # print(summary(cox_combined))
    # 
    # cat("\nModel with interaction:\n") 
    # print(summary(cox_interaction))
  }
}
```

### 5-Year Censored Combined Analysis

```{r}
# Use all 10 proteins from the treatment response analysis
top_proteins_5yr <- limma_b_pair_sig %>% pull(OlinkID)

if(length(top_proteins_5yr) > 0) {
  # Create combined analysis for each top protein (5-year censored)
  combined_results_5yr <- list()
  
  for(prot in top_proteins_5yr) {
    # Get protein expression data
    prot_data <- exprMatrix_long_5yr %>%
      filter(OlinkID == prot) %>%
      select(SampleID, NPX) %>%
      rename(protein_expr = NPX)
    
    # Merge with survival and metastatic load data
    combined_data <- df_survival_mets_5yr %>%
      left_join(prot_data, by = c("SampleID")) %>%
      filter(!is.na(protein_expr)) %>%
      mutate(
        # Dichotomize protein expression at median
        protein_high = ifelse(protein_expr > median(protein_expr, na.rm = TRUE), "High", "Low"),
        protein_high = factor(protein_high, levels = c("Low", "High")),
        # Create combined risk groups
        risk_group = case_when(
          mets_volume == "Low" & protein_high == "Low" ~ "Low-Low",
          mets_volume == "Low" & protein_high == "High" ~ "Low-High", 
          mets_volume == "High" & protein_high == "Low" ~ "High-Low",
          mets_volume == "High" & protein_high == "High" ~ "High-High"
        ),
        risk_group = factor(risk_group, levels = c("Low-Low", "Low-High", "High-Low", "High-High"))
      )
    
    # Cox models
    cox_combined <- coxph(Surv(time_to_crpc, progressed_to_crpc) ~ mets_volume + protein_high, 
                         data = combined_data)
    cox_interaction <- coxph(Surv(time_to_crpc, progressed_to_crpc) ~ mets_volume * protein_high, 
                           data = combined_data)
    
    # Store results
    protein_name <- df_id_pro %>% filter(OlinkID == prot) %>% pull(Assay)
    combined_results_5yr[[prot]] <- list(
      protein = prot,
      protein_name = protein_name,
      data = combined_data,
      cox_combined = cox_combined,
      cox_interaction = cox_interaction
    )
    
    # # Print summary
    # cat("\n=== 5-Year Censored Combined Analysis for", protein_name, "===\n")
    # cat("Model with metastatic load + protein expression:\n")
    # print(summary(cox_combined))
    # 
    # cat("\nModel with interaction:\n") 
    # print(summary(cox_interaction))
  }
}
```

### Forest Plots: Metastatic Load Controlled Analysis

Now let's create forest plots showing the hazard ratios for proteins while controlling for metastatic load.

```{r}
# Extract results from combined models (full follow-up)
if(length(combined_results_full) > 0) {
  
  # Extract protein HRs from additive models (controlling for metastatic load)
  protein_hrs_full <- map_dfr(combined_results_full, function(result) {
    cox_summary <- summary(result$cox_combined)  # Using additive model (not interaction)
    
    # Extract protein coefficient (protein effect adjusted for metastatic load)
    protein_coef <- cox_summary$coefficients["protein_highHigh", , drop = FALSE]
    protein_confint <- cox_summary$conf.int["protein_highHigh", , drop = FALSE]
    
    data.frame(
      OlinkID = result$protein,
      protein_name = result$protein_name,
      HR = protein_confint[,"exp(coef)"],
      Lower_Bound = protein_confint[,"lower .95"],
      Upper_Bound = protein_confint[,"upper .95"],
      P_Value = protein_coef[,"Pr(>|z|)"],
      stringsAsFactors = FALSE
    )
  }, .id = "model") %>%
    arrange(desc(HR)) %>%
    mutate(P_Value_FDR = p.adjust(P_Value, method = "fdr"))
  
  # Create forest plot for full follow-up (metastatic load controlled)
  df_forest_full <- protein_hrs_full %>%
    mutate(
      protein_name = factor(protein_name, levels = rev(protein_name)),
      HR_CI = sprintf("%.2f [%.2f, %.2f]", HR, Lower_Bound, Upper_Bound),
      P_Text = ifelse(P_Value_FDR < 0.001, formatC(P_Value_FDR, format = "e", digits = 2), sprintf("%.3f", P_Value_FDR)),
      P_Color = ifelse(P_Value_FDR < 0.05, "red", "black")
    )
  
  # Forest Plot
  p1_forest_full <- ggplot(df_forest_full, aes(x = HR, y = protein_name)) +
    geom_errorbarh(aes(xmin = Lower_Bound, xmax = Upper_Bound), height = 0.3, color = "grey50") +
    geom_point(color = "black", size = 3, alpha = 1) +
    geom_vline(xintercept = 1, linetype = "dashed") +
    theme_minimal(base_size = 12) +
    scale_x_continuous(limits = c(0, max(df_forest_full$Upper_Bound, na.rm = TRUE) * 1.1)) +
    labs(x = "Hazard Ratio (Protein High vs Low)", y = NULL, 
         title = "Protein HRs Controlling for Metastatic Load (Full Follow-up)") +
    theme(
      axis.text.y = element_text(face = "bold"),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank()
    )
  
  # Table with HR and P-values
  p2_forest_full <- ggplot(df_forest_full, aes(y = protein_name)) +
    geom_text(aes(x = 1, label = HR_CI), hjust = 0, size = 3.5) +
    geom_text(aes(x = 2, label = P_Text), hjust = 0, size = 3.5) +
    scale_x_continuous(limits = c(0.95, 2.9), breaks = NULL) +
    theme_void(base_size = 12) +
    theme(
      axis.text.y = element_blank(),
      plot.margin = margin(5.5, 20, 5.5, 5.5)
    )
  
  # Combine plots
  forest_plot_full <- p1_forest_full + p2_forest_full + plot_layout(widths = c(2, 1))
  print(forest_plot_full)
  
  ggsave(filename = file.path(wd$outCurr, "Forest_plot_metastatic_controlled_full.pdf"), 
         plot = forest_plot_full, width = 12, height = 8)
  
  # Save results
  write.csv(protein_hrs_full, file.path(wd$outCurr, "Protein_HRs_metastatic_controlled_full.csv"), row.names = FALSE)
}
```

```{r}
# Extract results from combined models (5-year censored)
if(length(combined_results_5yr) > 0) {
  
  # Extract protein HRs from additive models (controlling for metastatic load)
  protein_hrs_5yr <- map_dfr(combined_results_5yr, function(result) {
    cox_summary <- summary(result$cox_combined)  # Using additive model (not interaction)
    
    # Extract protein coefficient (protein effect adjusted for metastatic load)
    protein_coef <- cox_summary$coefficients["protein_highHigh", , drop = FALSE]
    protein_confint <- cox_summary$conf.int["protein_highHigh", , drop = FALSE]
    
    data.frame(
      OlinkID = result$protein,
      protein_name = result$protein_name,
      HR = protein_confint[,"exp(coef)"],
      Lower_Bound = protein_confint[,"lower .95"],
      Upper_Bound = protein_confint[,"upper .95"],
      P_Value = protein_coef[,"Pr(>|z|)"],
      stringsAsFactors = FALSE
    )
  }, .id = "model") %>%
    arrange(desc(HR)) %>%
    mutate(P_Value_FDR = p.adjust(P_Value, method = "fdr"))
  
  # Create forest plot for 5-year censored (metastatic load controlled)
  df_forest_5yr <- protein_hrs_5yr %>%
    mutate(
      protein_name = factor(protein_name, levels = rev(protein_name)),
      HR_CI = sprintf("%.2f [%.2f, %.2f]", HR, Lower_Bound, Upper_Bound),
      P_Text = ifelse(P_Value_FDR < 0.001, formatC(P_Value_FDR, format = "e", digits = 2), sprintf("%.3f", P_Value_FDR)),
      P_Color = ifelse(P_Value_FDR < 0.05, "red", "black")
    )
  
  # Forest Plot
  p1_forest_5yr <- ggplot(df_forest_5yr, aes(x = HR, y = protein_name)) +
    geom_errorbarh(aes(xmin = Lower_Bound, xmax = Upper_Bound), height = 0.3, color = "grey50") +
    geom_point(color = "black", size = 3, alpha = 1) +
    geom_vline(xintercept = 1, linetype = "dashed") +
    theme_minimal(base_size = 12) +
    scale_x_continuous(limits = c(0, max(df_forest_5yr$Upper_Bound, na.rm = TRUE) * 1.1)) +
    labs(x = "Hazard Ratio (Protein High vs Low)", y = NULL, 
         title = "Protein HRs Controlling for Metastatic Load (5-Year Censored)") +
    theme(
      axis.text.y = element_text(face = "bold"),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank()
    )
  
  # Table with HR and P-values
  p2_forest_5yr <- ggplot(df_forest_5yr, aes(y = protein_name)) +
    geom_text(aes(x = 1, label = HR_CI), hjust = 0, size = 3.5) +
    geom_text(aes(x = 2, label = P_Text), hjust = 0, size = 3.5) +
    scale_x_continuous(limits = c(0.95, 2.9), breaks = NULL) +
    theme_void(base_size = 12) +
    theme(
      axis.text.y = element_blank(),
      plot.margin = margin(5.5, 20, 5.5, 5.5)
    )
  
  # Combine plots
  forest_plot_5yr <- p1_forest_5yr + p2_forest_5yr + plot_layout(widths = c(2, 1))
  print(forest_plot_5yr)
  
  ggsave(filename = file.path(wd$outCurr, "Forest_plot_metastatic_controlled_5yr.pdf"), 
         plot = forest_plot_5yr, width = 12, height = 8)
  
  # Save results
  write.csv(protein_hrs_5yr, file.path(wd$outCurr, "Protein_HRs_metastatic_controlled_5yr.csv"), row.names = FALSE)
}
```

```{r}
# Comparison of metastatic load controlled results
if(exists("protein_hrs_full") && exists("protein_hrs_5yr")) {
  
  # Compare metastatic load controlled results
  comparison_metastatic_controlled <- protein_hrs_full %>%
    select(OlinkID, protein_name, HR_full = HR, P_Value_full = P_Value, P_FDR_full = P_Value_FDR) %>%
    left_join(
      protein_hrs_5yr %>% select(OlinkID, HR_5yr = HR, P_Value_5yr = P_Value, P_FDR_5yr = P_Value_FDR),
      by = "OlinkID"
    ) %>%
    mutate(
      HR_diff = HR_5yr - HR_full,
      P_diff = P_Value_5yr - P_Value_full,
      P_FDR_diff = P_FDR_5yr - P_FDR_full
    ) %>%
    arrange(desc(abs(HR_diff)))
  
  cat("=== Comparison: Metastatic Load Controlled Analysis ===\n")
  cat("Top proteins with largest HR differences (5yr vs full):\n")
  print(head(comparison_metastatic_controlled, 10))
  
  write.csv(comparison_metastatic_controlled, 
            file.path(wd$outCurr, "Metastatic_controlled_comparison.csv"), 
            row.names = FALSE)
}
```

### Final forest plot

```{r}
# Combined Forest Plot: Univariate vs Metastatic-Controlled
if(exists("results_base_5yr") && exists("protein_hrs_5yr")) {

  # Add PSA analysis with log10 transformation for both univariate and metastatic-controlled
  # Univariate PSA analysis (5-year censored) with log10 transformation
  meta_in_psa_5yr <- df_b_cox_5yr %>%
    rename(OS = time_to_crpc, death = progressed_to_crpc) %>%
    mutate(log10_psa = log10(psa))

  clin_var_psa <- c("log10_psa")
  results_psa_univariate_5yr <- cox_extract(clin_var_psa, meta_in_psa_5yr)
  psa_univariate_5yr <- results_psa_univariate_5yr$log10_psa %>%
    mutate(OlinkID = "log10_psa", Assay = "log10(PSA)")

  # Add metastatic load univariate analysis
  # Univariate analysis for metastatic load (High vs Low)
  meta_in_mets_5yr <- df_b_cox_5yr %>%
    rename(OS = time_to_crpc, death = progressed_to_crpc) %>%
    left_join(df_b_final %>% select(HCI_cID, mets_volume), by = "HCI_cID") %>%
    filter(!is.na(mets_volume)) %>%
    mutate(mets_volume = factor(mets_volume, levels = c("Low", "High")))

  cox_mets_univariate <- coxph(Surv(OS, death) ~ mets_volume, data = meta_in_mets_5yr)
  cox_summary_mets <- summary(cox_mets_univariate)

  mets_univariate_5yr <- data.frame(
    OlinkID = "mets_volume",
    Assay = "Metastatic Load",
    HR = cox_summary_mets$conf.int["mets_volumeHigh", "exp(coef)"],
    Lower_Bound = cox_summary_mets$conf.int["mets_volumeHigh", "lower .95"],
    Upper_Bound = cox_summary_mets$conf.int["mets_volumeHigh", "upper .95"],
    P_Value = cox_summary_mets$coefficients["mets_volumeHigh", "Pr(>|z|)"],
    stringsAsFactors = FALSE
  )

  # Metastatic-controlled PSA analysis using combined results structure
  # Get PSA data for metastatic load controlled analysis
  if(length(combined_results_5yr) > 0) {
    # Use the first protein's data structure as template and replace with PSA
    psa_combined_data <- combined_results_5yr[[1]]$data %>%
      select(-protein_expr, -protein_high, -risk_group) %>%
      mutate(
        protein_expr = log10(psa),
        protein_high = ifelse(protein_expr > median(protein_expr, na.rm = TRUE), "High", "Low"),
        protein_high = factor(protein_high, levels = c("Low", "High"))
      )

    # Cox model with metastatic load control for PSA
    cox_psa_controlled <- coxph(Surv(time_to_crpc, progressed_to_crpc) ~ mets_volume + protein_high,
                               data = psa_combined_data)

    # Extract PSA coefficient (protein effect adjusted for metastatic load)
    cox_summary_psa <- summary(cox_psa_controlled)
    protein_coef_psa <- cox_summary_psa$coefficients["protein_highHigh", , drop = FALSE]
    protein_confint_psa <- cox_summary_psa$conf.int["protein_highHigh", , drop = FALSE]

    psa_controlled_5yr <- data.frame(
      OlinkID = "log10_psa",
      Assay = "log10(PSA)",
      HR = protein_confint_psa[,"exp(coef)"],
      Lower_Bound = protein_confint_psa[,"lower .95"],
      Upper_Bound = protein_confint_psa[,"upper .95"],
      P_Value = protein_coef_psa[,"Pr(>|z|)"],
      stringsAsFactors = FALSE
    )

    # Extract metastatic load coefficient from the multivariate model
    # This shows the effect of metastatic load when controlling for protein expression
    mets_coef_psa <- cox_summary_psa$coefficients["mets_volumeHigh", , drop = FALSE]
    mets_confint_psa <- cox_summary_psa$conf.int["mets_volumeHigh", , drop = FALSE]

    mets_controlled_5yr <- data.frame(
      OlinkID = "mets_volume",
      Assay = "Metastatic Load",
      HR = mets_confint_psa[,"exp(coef)"],
      Lower_Bound = mets_confint_psa[,"lower .95"],
      Upper_Bound = mets_confint_psa[,"upper .95"],
      P_Value = mets_coef_psa[,"Pr(>|z|)"],
      stringsAsFactors = FALSE
    )
  }
  
  # Combine univariate and metastatic-controlled results for all 10 proteins + PSA + Metastatic Load
  combined_forest_data <- results_base_5yr %>%
    select(OlinkID, Assay,
           HR_univariate = HR,
           Lower_univariate = Lower_Bound,
           Upper_univariate = Upper_Bound,
           P_univariate = P_Value_FDR) %>%
    # Add PSA and Metastatic Load univariate results
    bind_rows(
      psa_univariate_5yr %>%
        select(OlinkID, Assay,
               HR_univariate = HR,
               Lower_univariate = Lower_Bound,
               Upper_univariate = Upper_Bound,
               P_univariate = P_Value) %>%
        mutate(P_univariate = p.adjust(P_univariate, method = "fdr")),
      mets_univariate_5yr %>%
        select(OlinkID, Assay,
               HR_univariate = HR,
               Lower_univariate = Lower_Bound,
               Upper_univariate = Upper_Bound,
               P_univariate = P_Value) %>%
        mutate(P_univariate = p.adjust(P_univariate, method = "fdr"))
    ) %>%
    left_join(
      # Combine protein, PSA, and Metastatic Load controlled results
      bind_rows(
        protein_hrs_5yr %>% select(OlinkID,
                                  HR_controlled = HR,
                                  Lower_controlled = Lower_Bound,
                                  Upper_controlled = Upper_Bound,
                                  P_controlled = P_Value_FDR),
        psa_controlled_5yr %>% select(OlinkID,
                                     HR_controlled = HR,
                                     Lower_controlled = Lower_Bound,
                                     Upper_controlled = Upper_Bound,
                                     P_controlled = P_Value) %>%
                              mutate(P_controlled = p.adjust(P_controlled, method = "fdr")),
        mets_controlled_5yr %>% select(OlinkID,
                                      HR_controlled = HR,
                                      Lower_controlled = Lower_Bound,
                                      Upper_controlled = Upper_Bound,
                                      P_controlled = P_Value) %>%
                               mutate(P_controlled = p.adjust(P_controlled, method = "fdr"))
      ),
      by = "OlinkID"
    ) %>%
    # Reshape to long format for plotting
    pivot_longer(
      cols = c(HR_univariate, HR_controlled),
      names_to = "Analysis",
      names_prefix = "HR_",
      values_to = "HR"
    ) %>%
    # Add corresponding CI bounds and p-values
    mutate(
      Lower_Bound = ifelse(Analysis == "univariate", Lower_univariate, Lower_controlled),
      Upper_Bound = ifelse(Analysis == "univariate", Upper_univariate, Upper_controlled),
      P_Value = ifelse(Analysis == "univariate", P_univariate, P_controlled),
      
      # Format for display
      Analysis = factor(Analysis, levels = c("univariate", "controlled"), 
                       labels = c("Univariate", "Metastatic-Controlled")),
      
      # Format HR and p-value text
      HR_CI = sprintf("%.2f [%.2f, %.2f]", HR, Lower_Bound, Upper_Bound),
      P_Text = ifelse(P_Value < 0.001, formatC(P_Value, format = "e", digits = 2), sprintf("%.3f", P_Value)),
      P_Color = ifelse(P_Value < 0.05, "red", "black")
    ) %>%
    # Order by univariate HR for consistent plotting
    arrange(Assay, Analysis)
  
  # Create ordering matching limma_b_pair_sig order (by log2FC)
  # Get the protein order from limma_b_pair_sig
  limma_protein_order <- limma_b_pair_sig %>%
    #left_join(df_id_pro, by = "OlinkID") %>%
    arrange(log2FC) %>%  # Arrange by log2FC (most negative first)
    pull(Assay) %>% as.character()

  # Combine with PSA at top (and metastatic load - commented out)
  protein_order <- c(# "Metastatic Load",  # Commented out - not shown in plot
                     "log10(PSA)",
                     limma_protein_order)

  # Filter out Metastatic Load from the forest plot data
  combined_forest_data_plot <- combined_forest_data %>%
    filter(Assay != "Metastatic Load")

  # Update protein_order to exclude Metastatic Load
  protein_order <- protein_order[protein_order != "Metastatic Load"]

  # Apply ordering to combined data
  combined_forest_data_plot <- combined_forest_data_plot %>%
    mutate(
      Assay = factor(Assay, levels = protein_order),
      # Create numeric position for y-axis (2 positions per protein)
      protein_num = as.numeric(factor(Assay, levels = rev(protein_order))),
      y_position = protein_num * 2 + ifelse(Analysis == "Univariate", 0.25, -0.25)
    ) %>%
    arrange(desc(Assay), Analysis)

  # Create data for protein labels (only show once per protein)
  protein_labels <- combined_forest_data_plot %>%
    filter(Analysis == "Univariate") %>%
    mutate(y_label = y_position - 1)  # Center between the two analysis points (shift down)

  # Create data for separation lines
  separation_lines <- data.frame(
    x_start = rep(0, length(protein_order)-1),
    x_end = rep(max(combined_forest_data_plot$Upper_Bound, na.rm = TRUE) * 1.1, length(protein_order)-1),
    y_sep = seq(2.5, length(protein_order)*2 - 0.5, by = 2)
  )
  
  # Create forest plot
  p1_combined <- ggplot(combined_forest_data_plot, aes(x = HR, y = y_position, color = Analysis)) +
    # Add separation lines between proteins
    geom_segment(data = separation_lines,
                aes(x = x_start, xend = x_end, y = y_sep, yend = y_sep),
                color = "grey80", linetype = "dashed", alpha = 0.5, inherit.aes = FALSE) +
    # Add error bars and points
    geom_errorbarh(aes(xmin = Lower_Bound, xmax = Upper_Bound),
                   height = 0.2, alpha = 0.8) +
    geom_point(size = 3, alpha = 0.9) +
    geom_vline(xintercept = 1, linetype = "dashed", color = "black") +
    # Add protein labels on the left
    geom_text(data = protein_labels,
              aes(x = 0, y = y_label, label = Assay),
              hjust = 1, vjust = 0.5, size = 3.5, color = "black",
              inherit.aes = FALSE) +
    theme_minimal(base_size = 11) +
    scale_color_manual(values = c("Univariate" = "#2E86AB", "Metastatic-Controlled" = "#A23B72")) +
    scale_x_continuous(limits = c(0, max(combined_forest_data_plot$Upper_Bound, na.rm = TRUE) * 1.1),
                       expand = expansion(mult = c(0.25, 0.05))) +
    scale_y_continuous(breaks = protein_labels$y_label,
                       labels = rep("", length(protein_labels$y_label))) +
    labs(x = "Hazard Ratio", y = NULL,
         title = "Protein Hazard Ratios: Univariate vs Metastatic Load-Controlled",
         color = "Analysis") +
    theme(
      axis.text.y = element_blank(),
      axis.ticks.y = element_blank(),
      panel.grid.major = element_blank(),
      panel.grid.minor = element_blank(),
      legend.position = "top",
      plot.margin = margin(5.5, 5.5, 5.5, 40)
    )

  # Create table with HR and P-values
  p2_combined <- ggplot(combined_forest_data_plot, aes(y = y_position)) +
    geom_text(aes(x = 1, label = HR_CI, color = Analysis), hjust = 0, size = 3) +
    geom_text(aes(x = 2.8, label = P_Text, color = ifelse(P_Value < 0.05, Analysis, "black")),
              hjust = 0, size = 3, show.legend = FALSE) +
    scale_color_manual(values = c("Univariate" = "#2E86AB", "Metastatic-Controlled" = "#A23B72", "black" = "black"),
                       guide = "none") +
    scale_x_continuous(limits = c(0.95, 4), breaks = NULL) +
    scale_y_continuous(limits = range(combined_forest_data_plot$y_position)) +
    theme_void(base_size = 11) +
    theme(
      axis.text.y = element_blank(),
      plot.margin = margin(5.5, 20, 5.5, 5.5)
    )
  
  # Combine plots
  final_combined_plot <- p1_combined + p2_combined + plot_layout(widths = c(2.5, 1))
  
  print(final_combined_plot)
  
  # Save plot
  ggsave(filename = file.path(wd$outCurr, "Forest_plot_univariate_vs_controlled.pdf"), 
         plot = final_combined_plot, width = 7, height = 8)
  
  # Save data
  write.csv(combined_forest_data, file.path(wd$outCurr, "Combined_forest_data.csv"), row.names = FALSE)
  
  cat("=== Combined Forest Plot Created ===\n")
  cat("Shows all 10 proteins + log10(PSA) with both univariate and multivariate HRs\n")
  cat("Note: Proteins are ordered by log2FC from limma_b_pair_sig (most negative first)\n")
  cat("      In the multivariate model, proteins/PSA are adjusted for metastatic load\n")
  cat("      Metastatic Load data is calculated but not shown in the plot (commented out)\n")
}
```

### Combined Risk Group Kaplan-Meier Plots

```{r,eval=FALSE}
pal.km <- c("#2E8B57","#2E8B58", "#DC143C", "#DC143D")
#pal.km <- c("#228B22", "#98DF8A", "#a02c2c", "#df8a8a")

# KM plots for combined risk groups (full follow-up) - ALL PROTEINS IN ONE PDF
if(length(top_proteins_full) > 0 && length(combined_results_full) > 0) {
  
  # Open PDF device for all plots
  pdf(file = file.path(wd$outCurr, "KM_combined_all_proteins_full.pdf"), width = 12, height = 10)
  
  # Loop through all proteins
  for(protein_id in names(combined_results_full)) {
    
    top_result_full <- combined_results_full[[protein_id]]
    top_protein_name_full <- top_result_full$protein_name
    
    # KM plot for 4-group risk stratification (full follow-up)
    surv_obj_combined_full <- Surv(time = top_result_full$data$time_to_crpc, 
                                  event = top_result_full$data$progressed_to_crpc)
    
    fit_combined_full <- survfit(surv_obj_combined_full ~ risk_group, data = top_result_full$data)
    
    km_plot_combined_full <- ggsurvplot(
      fit_combined_full,
      data = top_result_full$data,
      pval = TRUE,
      risk.table = TRUE,
      conf.int = FALSE,
      palette = pal.km,
      linetype = c("solid", "dashed", "solid", "dashed"),
      title = paste0("KM Plot: Combined Risk Groups (Full Follow-up)\n(", top_protein_name_full, " + Metastatic Load)"),
      xlab = "Time to CRPC (days)",
      ylab = "Progression-free survival probability",
      legend.title = "Risk Group",
      legend.labs = c("Mets:Low, Protein:Low", "Mets:Low, Protein:High", 
                     "Mets:High, Protein:Low", "Mets:High, Protein:High"),
      risk.table.height = 0.4
    )
    
    cat("Generating KM plot for", top_protein_name_full, "\n")
    
    # Print plot to PDF (this preserves risk table)
    print(km_plot_combined_full)
  }
  
  # Close PDF device
  dev.off()
}

# KM plots for combined risk groups (5-year censored) - ALL PROTEINS IN ONE PDF
if(length(top_proteins_5yr) > 0 && length(combined_results_5yr) > 0) {
  
  # Open PDF device for all plots
  pdf(file = file.path(wd$outCurr, "KM_combined_all_proteins_5yr.pdf"), width = 12, height = 10)
  
  # Loop through all proteins
  for(protein_id in names(combined_results_5yr)) {
    
    top_result_5yr <- combined_results_5yr[[protein_id]]
    top_protein_name_5yr <- top_result_5yr$protein_name
    
    # KM plot for 4-group risk stratification (5-year censored)
    surv_obj_combined_5yr <- Surv(time = top_result_5yr$data$time_to_crpc, 
                                 event = top_result_5yr$data$progressed_to_crpc)
    
    fit_combined_5yr <- survfit(surv_obj_combined_5yr ~ risk_group, data = top_result_5yr$data)
    
    km_plot_combined_5yr <- ggsurvplot(
      fit_combined_5yr,
      data = top_result_5yr$data,
      pval = TRUE,
      risk.table = TRUE,
      conf.int = FALSE,
      palette = pal.km,
      linetype = c("solid", "dashed", "solid", "dashed"),
      title = paste0("KM Plot: Combined Risk Groups (5-Year Censored)\n(", top_protein_name_5yr, " + Metastatic Load)"),
      xlab = "Time to CRPC (days)",
      ylab = "Progression-free survival probability",
      legend.title = "Risk Group",
      legend.labs = c("Mets:Low, Protein:Low", "Mets:Low, Protein:High", 
                     "Mets:High, Protein:Low", "Mets:High, Protein:High"),
      risk.table.height = 0.4
    )
    
    cat("Generating 5-year KM plot for", top_protein_name_5yr, "\n")
    
    # Print plot to PDF (this preserves risk table)
    print(km_plot_combined_5yr)
  }
  
  # Close PDF device
  dev.off()
}
```

## Baseline mets expression

We look at the protein expression in the 58 samples at baseline

```{r}
# Subset to the x proteins and the samples that are at baseline
mat_in_58 <- mat_in[prot_ids, df_b_cox$SampleID]

# Prepare long-format NPX data for plotting
df_npx_long <- mat_in_58 %>%
  as.data.frame() %>% 
  tibble::rownames_to_column(var = "OlinkID") %>% 
  pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX")

# Filter for b_stage_pro proteins
sig_olink_ids <- b_stage_pro

# Merge and filter for plotting
df_plot <- df_npx_long %>% 
  left_join(df_id_pro, by = "OlinkID") %>% 
  dplyr::filter(OlinkID %in% b_stage_pro) %>% 
  left_join(df_final_ids %>% select(HCI_cID, final_sample_id, cohort, txt_stat), by = "final_sample_id") %>% 
  filter(final_sample_id %in% colnames(exp_mat_sub)) %>% 
  left_join(df_b_final %>% select(HCI_cID, mets_volume))

# Ensure txt_stat is a factor with Pre first, then Post
df_plot$txt_stat <- factor(df_plot$txt_stat, levels = c("Pre", "Post"))

# Show the plot
max_y <- max(df_plot$NPX, na.rm = TRUE)
min_y <- min(df_plot$NPX, na.rm = TRUE)

p_viol_gene <- 
  df_plot %>%
  filter(Assay == "TOP1") %>% 
  ggplot(aes(x = mets_volume, y = NPX, fill = mets_volume)) +
  #geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  #geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(#outlier.alpha = 0,  width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  scale_y_continuous(limits=c(min_y, max_y+3)) +
  geom_signif(
    comparisons = list(c("High", "Low")),
    map_signif_level = FALSE,
    test = "t.test") +
  NULL
p_viol_gene
```





## Protein expression by metastatic load

Now let's look at the protein expression based on the metastatic load

Generate boxplot to show the expression of mHSPC-specific proteins based on metastatic load

```{r}
# Info on mets - df_b_final$mets_volume
# b_stage_pro - proteins overexpressed in group B

# Prepare long-format NPX data for plotting
df_npx_long <- exp_mat_sub %>%
  as.data.frame() %>% 
  tibble::rownames_to_column(var = "OlinkID") %>% 
  pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX")

# Filter for b_stage_pro proteins
sig_olink_ids <- b_stage_pro

# Merge and filter for plotting
df_plot <- df_npx_long %>% 
  left_join(df_id_pro, by = "OlinkID") %>% 
  dplyr::filter(OlinkID %in% b_stage_pro) %>% 
  left_join(df_final_ids %>% select(HCI_cID, final_sample_id, cohort, txt_stat), by = "final_sample_id") %>% 
  filter(final_sample_id %in% colnames(exp_mat_sub)) %>% 
  left_join(df_b_final %>% select(HCI_cID, mets_volume))

# Ensure txt_stat is a factor with Pre first, then Post
df_plot$txt_stat <- factor(df_plot$txt_stat, levels = c("Pre", "Post"))

# Show the plot
max_y <- max(df_plot$NPX, na.rm = TRUE)
min_y <- min(df_plot$NPX, na.rm = TRUE)

p_viol_gene <- 
  df_plot %>%
  ggplot(aes(x = mets_volume, y = NPX, fill = mets_volume)) +
  #geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  #geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(#outlier.alpha = 0,  width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  scale_y_continuous(limits=c(min_y, max_y+3)) +
  geom_signif(
    comparisons = list(c("High", "Low")),
    map_signif_level = FALSE,
    test = "t.test") +
  NULL
p_viol_gene
# ... existing code ...
```


If we limit to the 10 genes or KLK 3 only

```{r}
# Prepare long-format NPX data for plotting
df_npx_long <- exp_mat_sub %>%
  as.data.frame() %>% 
  tibble::rownames_to_column(var = "OlinkID") %>% 
  pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX")

# Filter for b_stage_pro proteins
sig_olink_ids <- limma_b_pair_sig

# Merge and filter for plotting
df_plot <- df_npx_long %>% 
  left_join(df_id_pro, by = "OlinkID") %>% 
  #dplyr::filter(OlinkID %in% sig_olink_ids$OlinkID) %>% 
  dplyr::filter(Assay %in% "KLK3") %>% 
  left_join(df_final_ids %>% select(final_sample_id, HCI_cID, cohort, txt_stat), by = "final_sample_id") %>% 
  filter(final_sample_id %in% colnames(exp_mat_sub)) %>%
  mutate(
    # For facetting, order by OlinkID factor
    OlinkID = factor(OlinkID, levels = unique(OlinkID))
  ) %>% 
  left_join(df_b_final %>% select(HCI_cID, mets_volume))

# Ensure txt_stat is a factor with Pre first, then Post
df_plot$txt_stat <- factor(df_plot$txt_stat, levels = c("Pre", "Post"))

# Show the plot
max_y <- max(df_plot$NPX, na.rm = TRUE)
min_y <- min(df_plot$NPX, na.rm = TRUE)

p_viol_gene <- 
  df_plot %>%
  #filter(Assay == "TOP1") %>% 
  ggplot(aes(x = mets_volume, y = NPX, fill = mets_volume)) +
  geom_quasirandom(dodge.width=0.9, colour="grey30", alpha=0.5) +
  geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
  geom_boxplot(outlier.alpha = 0,  width = 0.2, coef=0,
               position = position_dodge(width = 0.9)) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) +
  scale_y_continuous(limits=c(min_y, max_y+3)) +
  geom_signif(
    comparisons = list(c("High", "Low")),
    map_signif_level = FALSE,
    test = "t.test") +
  facet_wrap(~Assay, ncol=10) +
  scale_fill_manual(values=c("#ff7f0e", "gold4")) +
  NULL
p_viol_gene
```


# Obj 4: Baseline protein expression correlated with metastatic load

In this analysis, we examine which proteins are differentially expressed between high and low metastatic load in baseline samples. Since metastatic load is binary (High/Low), we'll use statistical tests to identify proteins that correlate with metastatic load.

## Data preparation

First, let's prepare the baseline samples with metastatic load information:

```{r}
# Get baseline samples (Pre1 only) with metastatic load information
df_baseline_mets <- df_b_cox_5yr %>%
  left_join(df_b_final %>% select(HCI_cID, mets_volume), by = "HCI_cID") %>%
  filter(!is.na(mets_volume)) %>%
  select(SampleID, HCI_cID, mrn, mets_volume, time_to_crpc, progressed_to_crpc) %>%
  rename(final_sample_id = SampleID)

cat("=== Baseline samples with metastatic load information ===\n")
cat("Total samples:", nrow(df_baseline_mets), "\n")
print(table(df_baseline_mets$mets_volume))
```

## Statistical analysis of protein expression by metastatic load

We'll use limma to compare protein expression between high and low metastatic load groups:

```{r}
# Prepare protein expression data for baseline samples
#prot_ids <- limma_b_pair_sig %>% pull(OlinkID) %>% as.character()
prot_ids <- b_stage_pro
mat_baseline <- exprMatrix_corrected_sub[prot_ids, df_baseline_mets$final_sample_id]

# Ensure matrix and metadata are aligned
identical(colnames(mat_baseline), df_baseline_mets$final_sample_id)

# Create design matrix for metastatic load comparison
group_mets <- df_baseline_mets$mets_volume
design_mets <- model.matrix(~ 0 + group_mets)
colnames(design_mets) <- c("High", "Low")

# Create contrast: High vs Low metastatic load
contrast_mets <- makeContrasts(
  High_vs_Low = High - Low,  # +ve log2FC means higher in High mets
  levels = design_mets
)

# Fit linear model
fit_mets <- lmFit(mat_baseline, design_mets)
fit2_mets <- contrasts.fit(fit_mets, contrast_mets)
fit2_mets <- eBayes(fit2_mets)

# Get results for all proteins (no filtering)
protein_mets_results <- topTable(fit2_mets, coef = 1, number = Inf, sort.by = "none") %>%
  tibble::rownames_to_column(var = "OlinkID") %>%
  left_join(df_id_pro, by = "OlinkID") %>%
  rename(
    P_Value = P.Value,
    P_Value_FDR = adj.P.Val,
    log2FC = logFC
  ) %>%
  mutate(
    Significant = ifelse(P_Value_FDR < cut_fdr, "Significant", "Non-significant"),
    Direction = case_when(
      log2FC > 0 & P_Value_FDR < cut_fdr ~ "Higher in High Mets",
      log2FC < 0 & P_Value_FDR < cut_fdr ~ "Higher in Low Mets", 
      TRUE ~ "Non-significant"
    ),
    # Round values for display
    log2FC = round(log2FC, 3),
    P_Value = round(P_Value, 6),
    P_Value_FDR = round(P_Value_FDR, 6)
  ) %>%
  arrange(P_Value)

# Calculate mean expression by group for additional info
mean_expression <- df_baseline_mets %>%
  select(final_sample_id, mets_volume) %>%
  left_join(
    mat_baseline %>%
      as.data.frame() %>%
      tibble::rownames_to_column(var = "OlinkID") %>%
      pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX"),
    by = "final_sample_id"
  ) %>%
  group_by(OlinkID, mets_volume) %>%
  summarise(mean_NPX = mean(NPX, na.rm = TRUE), .groups = "drop") %>%
  pivot_wider(names_from = mets_volume, values_from = mean_NPX, names_prefix = "Mean_")

# Add mean expression to results
protein_mets_results <- protein_mets_results %>%
  left_join(mean_expression, by = "OlinkID")

# Display results
cat("=== Protein Expression vs Metastatic Load Results (limma) ===\n")
cat("Total proteins analyzed:", nrow(protein_mets_results), "\n")
cat("Significant proteins (FDR < 0.2):", sum(protein_mets_results$P_Value_FDR < cut_fdr, na.rm = TRUE), "\n")

# Show top results
cat("\n=== Top 10 proteins by p-value ===\n")
print(protein_mets_results %>% 
      select(Assay, log2FC, P_Value, P_Value_FDR, Direction) %>% 
      head(10))

# Save results
write.csv(protein_mets_results, file.path(wd$outCurr, "Protein_expression_vs_metastatic_load_limma.csv"), row.names = FALSE)
```

## Visualization

```{r}
# Box plots for significant proteins
significant_proteins <- protein_mets_results %>%
  filter(P_Value_FDR < cut_fdr) %>%
  arrange(P_Value) %>%
  head(9)  # Top 9 for a 3x3 plot

if(nrow(significant_proteins) > 0) {
  
  # Prepare data for plotting
  df_npx_baseline <- mat_baseline %>%
    as.data.frame() %>% 
    tibble::rownames_to_column(var = "OlinkID") %>% 
    pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX") %>%
    left_join(df_id_pro, by = "OlinkID") %>%
    left_join(df_baseline_mets, by = "final_sample_id") %>%
    filter(!is.na(mets_volume), !is.na(NPX))
  
  sig_plot_data <- df_npx_baseline %>%
    filter(OlinkID %in% significant_proteins$OlinkID) %>%
    left_join(significant_proteins %>% select(OlinkID, P_Value_FDR, log2FC), by = "OlinkID") %>%
    mutate(
      facet_label = paste0(Assay, "\n", 
                          "log2FC=", round(log2FC, 2), 
                          "\nFDR=", formatC(P_Value_FDR, format = "e", digits = 1))
    )
  
  sig_boxplot <- sig_plot_data %>%
    ggplot(aes(x = mets_volume, y = NPX, fill = mets_volume)) +
    geom_quasirandom(dodge.width = 0.9, colour = "grey30", alpha = 0.5) +
    geom_violin(alpha = 0.5, scale = "width", position = position_dodge(width = 0.9)) +
    geom_boxplot(outlier.alpha = 0, width = 0.2, coef = 0,
                 position = position_dodge(width = 0.9)) +
    facet_wrap(~facet_label, scales = "free_y", ncol = 3) +
    scale_fill_manual(values = c("Low" = "#2E8B57", "High" = "#DC143C")) +
    labs(x = "Metastatic Load", y = "NPX Expression",
         title = "Significant Proteins Associated with Metastatic Load") +
    theme_bw() +
    theme(
      legend.position = "none",
      panel.grid.major = element_blank(), 
      panel.grid.minor = element_blank(),
      strip.text = element_text(size = 8)
    )
  
  print(sig_boxplot)
  ggsave(sig_boxplot, file = file.path(wd$outCurr, "Significant_proteins_mets_load.pdf"), 
         height = 3, width = 4)
  
} else {
  cat("No significant proteins found at FDR < 0.05\n")
}
```

## Summary statistics

```{r}
# Summary of results
summary_stats <- protein_mets_results %>%
  summarise(
    total_proteins = n(),
    significant_05 = sum(P_Value_FDR < 0.05, na.rm = TRUE),
    significant_01 = sum(P_Value_FDR < 0.01, na.rm = TRUE),
    significant_20 = sum(P_Value_FDR < 0.2, na.rm = TRUE),
    higher_in_high_mets_05 = sum(log2FC > 0 & P_Value_FDR < 0.05, na.rm = TRUE),
    higher_in_low_mets_05 = sum(log2FC < 0 & P_Value_FDR < 0.05, na.rm = TRUE),
    higher_in_high_mets_20 = sum(log2FC > 0 & P_Value_FDR < 0.2, na.rm = TRUE),
    higher_in_low_mets_20 = sum(log2FC < 0 & P_Value_FDR < 0.2, na.rm = TRUE),
    median_log2fc = median(abs(log2FC), na.rm = TRUE),
    max_log2fc = max(abs(log2FC), na.rm = TRUE)
  )

cat("=== Summary Statistics ===\n")
print(summary_stats)

# Show significant proteins by direction (FDR < 0.2)
if(sum(protein_mets_results$P_Value_FDR < 0.2, na.rm = TRUE) > 0) {
  cat("\n=== Significant proteins by direction (FDR < 0.2) ===\n")
  direction_summary <- protein_mets_results %>%
    filter(P_Value_FDR < 0.2) %>%
    count(Direction) %>%
    arrange(desc(n))
  print(direction_summary)
  
  # List significant proteins
  cat("\n=== Significant proteins (FDR < 0.2) ===\n")
  sig_list <- protein_mets_results %>%
    filter(P_Value_FDR < 0.2) %>%
    arrange(P_Value) %>%
    select(Assay, log2FC, P_Value_FDR, Direction)
  print(sig_list)
}
```


## Metastatic load vs PSA

Lets look at PSA as well KLK3 NPX in high vs low metastatic load

```{r}
# Use all mHSPC patients (df_b_final), not just paired
psa_protein_data <- df_b_final %>%
  filter(HCI_cID %in% df_meta_f5$HCI_cID) %>%
  mutate(logPSA = log10(psa),
         txt_stat = factor(updated_adt_early, levels = c("Pre", "Post"))) %>%
  select(HCI_cID, mrn, logPSA, txt_stat, psa, mets_volume) %>%
  left_join(df_link, by = "HCI_cID") %>%
  filter(!is.na(final_sample_id)) %>% 
  filter(final_sample_id %in% df_baseline_mets$final_sample_id) %>% 
  mutate(mets_volume= factor(mets_volume, levels=c("Low", "High")))

# Get KLK3 OlinkID
klk3_id <- df_id_pro %>%
    filter(Assay == "KLK3") %>%
    pull(OlinkID)

# Extract KLK3 expression for these samples
klk3_expression <- exprMatrix_corrected_sub[klk3_id, psa_protein_data$final_sample_id]
psa_protein_data$KLK3_NPX <- as.numeric(klk3_expression)

# Plot correlation
# Scatter plot
p_psa_klk3 <- ggplot(psa_protein_data, aes(x = KLK3_NPX, y = logPSA)) +
  geom_point(size = 5, alpha = 0.7) +
  geom_smooth(method = "lm", se = TRUE) +
  #scale_color_manual(values = c("#ff7f0e", "gold4")) +
  labs(
    title = sprintf("log(PSA) vs KLK3 Expression (rho = %.2f, p = %.2e)",
                    cor_overall$estimate, cor_overall$p.value),
    x = "KLK3 NPX",
    y = "log10(PSA)",
    color = "Treatment Status"
  ) +
  theme_bw() +
  theme_custom



# Plot KLK3 NPX by metastatic burden
p_klk3_mets <- ggplot(psa_protein_data, aes(x = mets_volume, y = KLK3_NPX, fill = mets_volume)) +
  geom_violin(alpha = 0.7, trim = FALSE) +
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.5) +
  ggbeeswarm::geom_quasirandom(size = 2, alpha = 0.6) +
  ggsignif::geom_signif(
    comparisons = list(c("High", "Low")),
    map_signif_level = FALSE,
    test = "wilcox.test"
  ) +
  scale_fill_manual(values = c("Low" = "#2E8B57", "High" = "#DC143C")) +
  labs(
    title = "KLK3 Protein Expression by Metastatic Burden",
    subtitle = "Baseline (Pre-treatment) samples",
    x = "Metastatic Burden",
    y = "KLK3 NPX",
    fill = "Metastatic\nBurden"
  ) +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0))

# Plot PSA by metastatic burden
p_psa_mets <- ggplot(psa_protein_data, aes(x = mets_volume, y = logPSA, fill = mets_volume)) +
  geom_violin(alpha = 0.7, trim = FALSE) +
  geom_boxplot(width = 0.2, outlier.shape = NA, alpha = 0.5) +
  ggbeeswarm::geom_quasirandom(size = 2, alpha = 0.6) +
  ggsignif::geom_signif(
    comparisons = list(c("High", "Low")),
    map_signif_level = FALSE,
    test = "wilcox.test"
  ) +
  scale_fill_manual(values = c("Low" = "#2E8B57", "High" = "#DC143C")) +
  labs(
    title = "PSA Clinical Values by Metastatic Burden",
    subtitle = "Baseline (Pre-treatment) samples",
    x = "Metastatic Burden",
    y = "log10(PSA)",
    fill = "Metastatic\nBurden"
  ) +
  theme_bw() +
  theme(legend.position = "right",
        panel.grid.major = element_line(linetype = "blank"), 
        panel.grid.minor = element_line(linetype = "blank"),
        axis.text.x = element_text(angle = 0)) 

# Display plots
print(p_klk3_mets)
print(p_psa_mets)

# Print summary statistics
cat("\n=== KLK3 NPX by Metastatic Burden ===\n")
print(plot_data %>%
  group_by(mets_volume) %>%
  summarise(
    n = n(),
    mean = mean(KLK3_NPX, na.rm = TRUE),
    median = median(KLK3_NPX, na.rm = TRUE),
    sd = sd(KLK3_NPX, na.rm = TRUE)
  ))

cat("\n=== PSA by Metastatic Burden ===\n")
print(plot_data %>%
  group_by(mets_volume) %>%
  summarise(
    n = sum(!is.na(psa)),
    mean_psa = mean(psa, na.rm = TRUE),
    median_psa = median(psa, na.rm = TRUE),
    sd_psa = sd(psa, na.rm = TRUE)
  ))

# Perform statistical tests
cat("\n=== Statistical Tests ===\n")
cat("KLK3 NPX - Wilcoxon test:\n")
wilcox_klk3 <- wilcox.test(KLK3_NPX ~ mets_volume, data = plot_data)
print(wilcox_klk3)

cat("\nPSA - Wilcoxon test:\n")
wilcox_psa <- wilcox.test(logPSA ~ mets_volume, data = plot_data)
print(wilcox_psa)

# Save plots
ggsave(p_klk3_mets, file = file.path(wd$outCurr, "KLK3_metastatic_burden_baseline.pdf"),
       height = 6, width = 7)
ggsave(p_psa_mets, file = file.path(wd$outCurr, "PSA_metastatic_burden_baseline.pdf"),
       height = 6, width = 7)
```


## Cox regression analysis for metastatic load-associated proteins

For the proteins significantly associated with metastatic load (FDR < 0.2), we'll perform Cox regression analysis to assess their prognostic value using the full follow-up approach while controlling for metastatic load.

```{r}
# Get proteins significantly associated with metastatic load (FDR < 0.2)
mets_associated_proteins <- protein_mets_results %>%
  filter(P_Value_FDR < 0.2) %>%
  arrange(P_Value)

cat("=== Proteins associated with metastatic load (FDR < 0.2) ===\n")
cat("Number of proteins:", nrow(mets_associated_proteins), "\n")
print(mets_associated_proteins %>% select(Assay, log2FC, P_Value_FDR, Direction))

if(nrow(mets_associated_proteins) > 0) {
  
  # Prepare survival data (full follow-up)
  meta_cox_full <- df_b_cox_full %>% 
    rename(OS = time_to_crpc, death = progressed_to_crpc)
  
  # Get metastatic load information
  df_survival_mets_full <- meta_cox_full %>%
    left_join(df_b_final %>% select(HCI_cID, mets_volume), by = "HCI_cID") %>%
    filter(!is.na(mets_volume)) %>%
    mutate(mets_volume = factor(mets_volume, levels = c("Low", "High")))
  
  # Get protein expression data for these proteins
  mets_protein_ids <- mets_associated_proteins %>% pull(OlinkID)
  mat_mets_proteins <- exprMatrix_corrected_sub[mets_protein_ids, df_survival_mets_full$SampleID]
  
  # Convert to long format
  exprMatrix_mets_long <- mat_mets_proteins %>%
    as.data.frame() %>%
    rownames_to_column(var = "OlinkID") %>%
    pivot_longer(cols = -OlinkID, names_to = "SampleID", values_to = "NPX") %>%
    rename(sample_id = SampleID)
  
  # Perform Cox regression for each protein
  cox_results_mets <- list()
  
  for (prot in mets_protein_ids) {
    # Get protein expression data
    prot_data <- exprMatrix_mets_long %>%
      filter(OlinkID == prot) %>%
      select(sample_id, NPX) %>%
      rename(protein_expr = NPX)
    
    # Merge with survival and metastatic load data
    combined_data <- df_survival_mets_full %>%
      left_join(prot_data, by = c("SampleID" = "sample_id")) %>%
      filter(!is.na(protein_expr)) %>%
      mutate(
        # Dichotomize protein expression at median
        protein_high = ifelse(protein_expr > median(protein_expr, na.rm = TRUE), "High", "Low"),
        protein_high = factor(protein_high, levels = c("Low", "High"))
      )
    
    # Run Cox models
    # 1. Univariate protein model
    cox_univariate <- coxph(Surv(OS, death) ~ protein_expr, data = combined_data)
    
    # 2. Model controlling for metastatic load
    cox_controlled <- coxph(Surv(OS, death) ~ protein_expr + mets_volume, data = combined_data)
    
    # 3. Model with interaction
    cox_interaction <- coxph(Surv(OS, death) ~ protein_expr * mets_volume, data = combined_data)
    
    # Extract results
    protein_name <- df_id_pro %>% filter(OlinkID == prot) %>% pull(Assay)
    
    # Univariate results
    univ_summary <- summary(cox_univariate)
    univ_coef <- univ_summary$coefficients["protein_expr", ]
    univ_confint <- univ_summary$conf.int["protein_expr", ]
    
    # Controlled results
    ctrl_summary <- summary(cox_controlled)
    ctrl_coef <- ctrl_summary$coefficients["protein_expr", ]
    ctrl_confint <- ctrl_summary$conf.int["protein_expr", ]
    
    # Store results
    cox_results_mets[[prot]] <- data.frame(
      OlinkID = prot,
      Assay = protein_name,
      # Univariate
      HR_univariate = univ_confint["exp(coef)"],
      Lower_univariate = univ_confint["lower .95"],
      Upper_univariate = univ_confint["upper .95"],
      P_univariate = univ_coef["Pr(>|z|)"],
      # Metastatic load controlled
      HR_controlled = ctrl_confint["exp(coef)"],
      Lower_controlled = ctrl_confint["lower .95"],
      Upper_controlled = ctrl_confint["upper .95"],
      P_controlled = ctrl_coef["Pr(>|z|)"],
      # Sample info
      n_samples = nrow(combined_data),
      n_events = sum(combined_data$death),
      stringsAsFactors = FALSE
    )
    
    cat("\n=== Cox regression results for", protein_name, "===\n")
    cat("Sample size:", nrow(combined_data), "| Events:", sum(combined_data$death), "\n")
    cat("Univariate HR:", round(univ_confint["exp(coef)"], 3), 
        "[", round(univ_confint["lower .95"], 3), "-", round(univ_confint["upper .95"], 3), "]",
        "P =", formatC(univ_coef["Pr(>|z|)"], format = "e", digits = 2), "\n")
    cat("Controlled HR:", round(ctrl_confint["exp(coef)"], 3), 
        "[", round(ctrl_confint["lower .95"], 3), "-", round(ctrl_confint["upper .95"], 3), "]",
        "P =", formatC(ctrl_coef["Pr(>|z|)"], format = "e", digits = 2), "\n")
  }
  
  # Combine results
  cox_results_df <- do.call(rbind, cox_results_mets) %>%
    mutate(
      P_univariate_FDR = p.adjust(P_univariate, method = "fdr"),
      P_controlled_FDR = p.adjust(P_controlled, method = "fdr")
    ) %>%
    arrange(P_univariate)
  
  # Save results
  write.csv(cox_results_df, file.path(wd$outCurr, "Cox_regression_mets_associated_proteins.csv"), row.names = FALSE)
  
  # Display final results
  cat("\n=== Summary: Cox regression for metastatic load-associated proteins ===\n")
  print(cox_results_df %>% 
        select(Assay, HR_univariate, P_univariate, HR_controlled, P_controlled) %>%
        mutate(across(starts_with("HR"), ~ round(.x, 3)),
               across(starts_with("P_"), ~ formatC(.x, format = "e", digits = 2))))
  
} else {
  cat("No proteins significantly associated with metastatic load at FDR < 0.2\n")
}
```

## Pre/Post treatment violin plots by metastatic load

Create violin plots showing protein expression before and after treatment, stratified by metastatic load for the proteins associated with metastatic burden.

```{r}
# Create violin plots for the metastatic load-associated proteins
if(exists("mets_associated_proteins") && nrow(mets_associated_proteins) > 0) {
  
  # Get the protein IDs
  mets_protein_ids <- mets_associated_proteins %>% pull(OlinkID)
  
  # Prepare all Pre and Post treatment data with metastatic load information
  df_prepost_mets <- df_meta_f5 %>%
    filter(cohort == "B") %>%  # mHSPC cohort
    filter(txt_stat %in% c("Pre", "Post")) %>%  # All Pre and Post samples
    left_join(df_b_final %>% select(HCI_cID, mets_volume), by = "HCI_cID") %>%
    filter(!is.na(mets_volume), !is.na(final_sample_id))
  
  # Get protein expression matrix for all Pre/Post samples
  exp_mat_mets <- exprMatrix_corrected_sub[mets_protein_ids, df_prepost_mets$final_sample_id]
  
  # Prepare long-format data for plotting
  df_plot_mets <- exp_mat_mets %>%
    as.data.frame() %>% 
    tibble::rownames_to_column(var = "OlinkID") %>% 
    pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX") %>%
    left_join(df_id_pro, by = "OlinkID") %>%
    left_join(df_prepost_mets %>% select(final_sample_id, HCI_cID, mrn, txt_stat, mets_volume), 
              by = "final_sample_id") %>%
    filter(!is.na(mets_volume), !is.na(NPX)) %>%
    mutate(
      txt_stat = factor(txt_stat, levels = c("Pre", "Post")),
      mets_volume = factor(mets_volume, levels = c("Low", "High"))
    )
  
  cat("=== Pre/Post violin plots for metastatic load-associated proteins ===\n")
  cat("Number of proteins:", length(mets_protein_ids), "\n")
  cat("Number of samples with mets data:", nrow(df_plot_mets), "\n")
  cat("Number of patients:", length(unique(df_plot_mets$mrn)), "\n")
  
  # Show sample distribution
  sample_counts <- df_plot_mets %>%
    count(txt_stat, mets_volume) %>%
    pivot_wider(names_from = mets_volume, values_from = n, names_prefix = "Mets_")
  cat("Sample distribution:\n")
  print(sample_counts)
  
  # Create violin plot
  p_violin_mets <- df_plot_mets %>%
    ggplot(aes(x = txt_stat, y = NPX, fill = mets_volume)) +
    geom_quasirandom(aes(color = mets_volume), 
                     dodge.width = 0.8, alpha = 0.6, size = 1.5) +
    geom_violin(alpha = 0.7, scale = "width", 
                position = position_dodge(width = 0.8), 
                draw_quantiles = c(0.5)) +
    geom_boxplot(outlier.alpha = 0, width = 0.15, coef = 0,
                 position = position_dodge(width = 0.8),
                 alpha = 0.8, color = "white") +
    facet_wrap(~Assay, scales = "free_y", ncol = length(mets_protein_ids)) +
    scale_fill_manual(values = c("Low" = "#2E8B57", "High" = "#DC143C"),
                      name = "Metastatic Load") +
    scale_color_manual(values = c("Low" = "#1F5F3F", "High" = "#A01F1F"),
                       name = "Metastatic Load") +
    labs(x = "Treatment Status", 
         y = "NPX Expression",
         title = "Protein Expression Pre/Post Treatment by Metastatic Load",
         subtitle = "Paired samples from mHSPC cohort") +
    theme_bw() +
    theme(
      panel.grid.major = element_blank(), 
      panel.grid.minor = element_blank(),
      strip.text = element_text(size = 12, face = "bold"),
      axis.text = element_text(size = 10),
      axis.title = element_text(size = 12),
      legend.position = "top",
      legend.title = element_text(size = 12),
      legend.text = element_text(size = 10),
      plot.title = element_text(size = 14, face = "bold"),
      plot.subtitle = element_text(size = 12)
    )
  
  print(p_violin_mets)
  
  # Save the plot
  ggsave(p_violin_mets, 
         file = file.path(wd$outCurr, "PrePost_violin_by_mets_load.pdf"), 
         width = max(8, length(mets_protein_ids) * 4), 
         height = 6)
  
  # Additional analysis: Statistical comparison within each protein and mets group
  cat("\n=== Statistical analysis: Treatment effect by metastatic load ===\n")
  
  for(protein_id in mets_protein_ids) {
    protein_name <- df_id_pro %>% filter(OlinkID == protein_id) %>% pull(Assay)
    
    # Get data for this protein
    protein_data <- df_plot_mets %>% 
      filter(OlinkID == protein_id)
    
    cat("\n--- ", protein_name, " ---\n")
    
    # Test treatment effect within each metastatic load group (unpaired)
    for(mets_group in c("Low", "High")) {
      mets_data <- protein_data %>% filter(mets_volume == mets_group)
      
      pre_data <- mets_data %>% filter(txt_stat == "Pre") %>% pull(NPX)
      post_data <- mets_data %>% filter(txt_stat == "Post") %>% pull(NPX)
      
      if(length(pre_data) >= 3 & length(post_data) >= 3) {
        # Unpaired t-test for pre vs post within mets group
        test_result <- t.test(post_data, pre_data, paired = FALSE)
        
        cat(mets_group, "Mets - Pre: n=", length(pre_data), 
            " (mean=", round(mean(pre_data), 3), "), ",
            "Post: n=", length(post_data), 
            " (mean=", round(mean(post_data), 3), "), ",
            "P = ", formatC(test_result$p.value, format = "e", digits = 2), "\n")
      }
    }
    
    # Test interaction: Does treatment effect differ between metastatic load groups?
    # Using two-way ANOVA approach
    if(nrow(protein_data) >= 10) {
      tryCatch({
        # Two-way ANOVA: NPX ~ txt_stat * mets_volume
        aov_result <- aov(NPX ~ txt_stat * mets_volume, data = protein_data)
        aov_summary <- summary(aov_result)
        
        interaction_p <- aov_summary[[1]]["txt_stat:mets_volume", "Pr(>F)"]
        treatment_p <- aov_summary[[1]]["txt_stat", "Pr(>F)"]
        mets_p <- aov_summary[[1]]["mets_volume", "Pr(>F)"]
        
        cat("Two-way ANOVA - Treatment effect: P = ", 
            formatC(treatment_p, format = "e", digits = 2), "\n")
        cat("Two-way ANOVA - Mets effect: P = ", 
            formatC(mets_p, format = "e", digits = 2), "\n")
        cat("Two-way ANOVA - Interaction: P = ", 
            formatC(interaction_p, format = "e", digits = 2), "\n")
      }, error = function(e) {
        cat("ANOVA analysis failed: insufficient data\n")
      })
    }
  }
  
  # Summary statistics table
  summary_table <- df_plot_mets %>%
    group_by(Assay, txt_stat, mets_volume) %>%
    summarise(
      n = n(),
      mean_NPX = round(mean(NPX, na.rm = TRUE), 3),
      sd_NPX = round(sd(NPX, na.rm = TRUE), 3),
      median_NPX = round(median(NPX, na.rm = TRUE), 3),
      .groups = "drop"
    ) %>%
    arrange(Assay, mets_volume, txt_stat)
  
  cat("\n=== Summary statistics by protein, treatment, and metastatic load ===\n")
  print(summary_table)
  
  # Save summary table
  write.csv(summary_table, 
            file.path(wd$outCurr, "PrePost_summary_by_mets_load.csv"), 
            row.names = FALSE)
  
} else {
  cat("No metastatic load-associated proteins found for pre/post analysis\n")
}
```

## Expression of metastatic load-associated proteins in mCRPC cohort

Examine how the proteins associated with metastatic load in mHSPC behave in the mCRPC cohort (Pre and Post treatment). Note: metastatic load information is not available for mCRPC patients.

```{r}
# Create violin plots for the metastatic load-associated proteins in mCRPC cohort
if(exists("mets_associated_proteins") && nrow(mets_associated_proteins) > 0) {
  
  # Get the protein IDs
  mets_protein_ids <- mets_associated_proteins %>% pull(OlinkID)
  
  # Prepare mCRPC Pre and Post treatment data
  df_mcrpc_prepost <- df_meta_f5 %>%
    filter(cohort %in% c("C", "D")) %>%  # mCRPC cohorts
    filter(txt_stat %in% c("Pre", "Post")) %>%  # All Pre and Post samples
    filter(!is.na(final_sample_id))
  
  # Get protein expression matrix for mCRPC Pre/Post samples
  exp_mat_mcrpc <- exprMatrix_corrected_sub[mets_protein_ids, df_mcrpc_prepost$final_sample_id]
  
  # Prepare long-format data for plotting
  df_plot_mcrpc <- exp_mat_mcrpc %>%
    as.data.frame() %>% 
    tibble::rownames_to_column(var = "OlinkID") %>% 
    pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX") %>%
    left_join(df_id_pro, by = "OlinkID") %>%
    left_join(df_mcrpc_prepost %>% select(final_sample_id, HCI_cID, mrn, txt_stat, cohort), 
              by = "final_sample_id") %>%
    filter(!is.na(NPX)) %>%
    mutate(
      txt_stat = factor(txt_stat, levels = c("Pre", "Post")),
      cohort = factor(cohort, levels = c("C", "D"))
    )
  
  cat("=== mCRPC violin plots for metastatic load-associated proteins ===\n")
  cat("Number of proteins:", length(mets_protein_ids), "\n")
  cat("Number of mCRPC samples:", nrow(df_plot_mcrpc), "\n")
  cat("Number of mCRPC patients:", length(unique(df_plot_mcrpc$mrn)), "\n")
  
  # Show sample distribution
  sample_counts_mcrpc <- df_plot_mcrpc %>%
    count(txt_stat, cohort) %>%
    pivot_wider(names_from = cohort, values_from = n, names_prefix = "Cohort_")
  cat("Sample distribution by cohort:\n")
  print(sample_counts_mcrpc)
  
  # Total sample counts by treatment status
  total_counts <- df_plot_mcrpc %>%
    count(txt_stat)
  cat("Total samples by treatment status:\n")
  print(total_counts)
  
  # Create violin plot for mCRPC
  p_violin_mcrpc <- df_plot_mcrpc %>%
    ggplot(aes(x = txt_stat, y = NPX, fill = txt_stat)) +
    geom_quasirandom(alpha = 0.6, size = 1.5, color = "grey30") +
    geom_violin(alpha = 0.7, scale = "width", draw_quantiles = c(0.5)) +
    geom_boxplot(outlier.alpha = 0, width = 0.2, coef = 0,
                 alpha = 0.8, color = "white") +
    facet_wrap(~Assay, scales = "free_y", ncol = length(mets_protein_ids)) +
    scale_fill_manual(values = c("Pre" = "#FF7F0E", "Post" = "#D62728"),
                      name = "Treatment") +
    labs(x = "Treatment Status", 
         y = "NPX Expression",
         title = "Metastatic Load-Associated Proteins in mCRPC Cohort",
         subtitle = "Expression before and after mCRPC treatment") +
    theme_bw() +
    theme(
      panel.grid.major = element_blank(), 
      panel.grid.minor = element_blank(),
      strip.text = element_text(size = 12, face = "bold"),
      axis.text = element_text(size = 10),
      axis.title = element_text(size = 12),
      legend.position = "top",
      legend.title = element_text(size = 12),
      legend.text = element_text(size = 10),
      plot.title = element_text(size = 14, face = "bold"),
      plot.subtitle = element_text(size = 12)
    )
  
  print(p_violin_mcrpc)
  
  # Save the plot
  ggsave(p_violin_mcrpc, 
         file = file.path(wd$outCurr, "mCRPC_PrePost_violin.pdf"), 
         width = max(8, length(mets_protein_ids) * 4), 
         height = 6)
  
  # Statistical analysis for mCRPC
  cat("\n=== Statistical analysis: mCRPC treatment effect ===\n")
  
  for(protein_id in mets_protein_ids) {
    protein_name <- df_id_pro %>% filter(OlinkID == protein_id) %>% pull(Assay)
    
    # Get data for this protein
    protein_data_mcrpc <- df_plot_mcrpc %>% 
      filter(OlinkID == protein_id)
    
    cat("\n--- ", protein_name, " in mCRPC ---\n")
    
    # Test treatment effect in mCRPC (unpaired)
    pre_data <- protein_data_mcrpc %>% filter(txt_stat == "Pre") %>% pull(NPX)
    post_data <- protein_data_mcrpc %>% filter(txt_stat == "Post") %>% pull(NPX)
    
    if(length(pre_data) >= 3 & length(post_data) >= 3) {
      # Unpaired t-test for pre vs post
      test_result <- t.test(post_data, pre_data, paired = FALSE)
      
      cat("Pre: n=", length(pre_data), " (mean=", round(mean(pre_data), 3), 
          ", sd=", round(sd(pre_data), 3), ")\n")
      cat("Post: n=", length(post_data), " (mean=", round(mean(post_data), 3), 
          ", sd=", round(sd(post_data), 3), ")\n")
      cat("Treatment effect P = ", formatC(test_result$p.value, format = "e", digits = 2), "\n")
      cat("Effect size (Cohen's d): ", 
          round((mean(post_data) - mean(pre_data)) / sqrt(((length(post_data)-1)*var(post_data) + (length(pre_data)-1)*var(pre_data))/(length(post_data)+length(pre_data)-2)), 3), "\n")
    } else {
      cat("Insufficient data for statistical testing\n")
    }
    
    # Test cohort effect within treatment status
    if(length(unique(protein_data_mcrpc$cohort)) == 2) {
      cat("Cohort comparison:\n")
      
      for(tx_status in c("Pre", "Post")) {
        cohort_data <- protein_data_mcrpc %>% filter(txt_stat == tx_status)
        
        if(nrow(cohort_data) >= 6 && length(unique(cohort_data$cohort)) == 2) {
          cohort_test <- t.test(NPX ~ cohort, data = cohort_data)
          cat("  ", tx_status, "samples - Cohort C vs D: P = ", 
              formatC(cohort_test$p.value, format = "e", digits = 2), "\n")
        }
      }
    }
  }
  
  # Summary statistics table for mCRPC
  summary_table_mcrpc <- df_plot_mcrpc %>%
    group_by(Assay, txt_stat) %>%
    summarise(
      n = n(),
      mean_NPX = round(mean(NPX, na.rm = TRUE), 3),
      sd_NPX = round(sd(NPX, na.rm = TRUE), 3),
      median_NPX = round(median(NPX, na.rm = TRUE), 3),
      q25_NPX = round(quantile(NPX, 0.25, na.rm = TRUE), 3),
      q75_NPX = round(quantile(NPX, 0.75, na.rm = TRUE), 3),
      .groups = "drop"
    ) %>%
    arrange(Assay, txt_stat)
  
  cat("\n=== Summary statistics for mCRPC cohort ===\n")
  print(summary_table_mcrpc)
  
  # Save summary table
  write.csv(summary_table_mcrpc, 
            file.path(wd$outCurr, "mCRPC_PrePost_summary.csv"), 
            row.names = FALSE)
  
  # Compare mHSPC vs mCRPC baseline expression
  cat("\n=== Comparison: mHSPC vs mCRPC baseline expression ===\n")
  
  if(exists("df_plot_mets")) {
    # Get mHSPC baseline data
    mhspc_baseline <- df_plot_mets %>%
      filter(txt_stat == "Pre") %>%
      select(OlinkID, Assay, NPX) %>%
      mutate(cohort_type = "mHSPC")
    
    # Get mCRPC baseline data  
    mcrpc_baseline <- df_plot_mcrpc %>%
      filter(txt_stat == "Pre") %>%
      select(OlinkID, Assay, NPX) %>%
      mutate(cohort_type = "mCRPC")
    
    # Combine for comparison
    combined_baseline <- bind_rows(mhspc_baseline, mcrpc_baseline) %>%
      mutate(cohort_type = factor(cohort_type, levels = c("mHSPC", "mCRPC")))
    
    for(protein_id in mets_protein_ids) {
      protein_name <- df_id_pro %>% filter(OlinkID == protein_id) %>% pull(Assay)
      
      baseline_data <- combined_baseline %>% filter(OlinkID == protein_id)
      
      if(nrow(baseline_data) >= 6) {
        baseline_test <- t.test(NPX ~ cohort_type, data = baseline_data)
        
        mhspc_mean <- baseline_data %>% filter(cohort_type == "mHSPC") %>% pull(NPX) %>% mean()
        mcrpc_mean <- baseline_data %>% filter(cohort_type == "mCRPC") %>% pull(NPX) %>% mean()
        
        cat(protein_name, ": mHSPC baseline mean =", round(mhspc_mean, 3),
            ", mCRPC baseline mean =", round(mcrpc_mean, 3),
            ", P =", formatC(baseline_test$p.value, format = "e", digits = 2), "\n")
      }
    }
  }
  
} else {
  cat("No metastatic load-associated proteins found for mCRPC analysis\n")
}
```

## Enzalutamide-treated mCRPC patients analysis

Examine the expression of metastatic load-associated proteins specifically in mCRPC patients treated with Enzalutamide (Pre vs Post treatment).

```{r}
# Analyze proteins in Enzalutamide-treated mCRPC patients
if(exists("mets_associated_proteins") && nrow(mets_associated_proteins) > 0) {
  
  # Get the protein IDs
  mets_protein_ids <- mets_associated_proteins %>% pull(OlinkID)
  
  # Check if df_c_final exists and has treatment information
  if(exists("df_c_final") && "txt_mcrpc_1st_type" %in% colnames(df_c_final)) {
    
    # Prepare Enzalutamide-treated mCRPC data
    df_enza_patients <- df_c_final %>%
      filter(txt_mcrpc_1st_type == "Enzalutamide") %>%
      select(HCI_cID, txt_mcrpc_1st_type)
    
    cat("=== Enzalutamide-treated mCRPC patients ===\n")
    cat("Number of Enzalutamide-treated patients:", nrow(df_enza_patients), "\n")
    
    # Get Pre/Post samples for Enzalutamide-treated patients
    df_enza_prepost <- df_meta_f5 %>%
      filter(cohort %in% c("C", "D")) %>%
      filter(txt_stat %in% c("Pre", "Post")) %>%
      filter(HCI_cID %in% df_enza_patients$HCI_cID) %>%
      left_join(df_enza_patients, by = "HCI_cID") %>%
      filter(!is.na(final_sample_id))
    
    cat("Number of Pre/Post samples from Enzalutamide patients:", nrow(df_enza_prepost), "\n")
    
    if(nrow(df_enza_prepost) > 0) {
      
      # Get protein expression matrix for Enzalutamide Pre/Post samples
      exp_mat_enza <- exprMatrix_corrected_sub[mets_protein_ids, df_enza_prepost$final_sample_id]
      
      # Prepare long-format data for plotting
      df_plot_enza <- exp_mat_enza %>%
        as.data.frame() %>% 
        tibble::rownames_to_column(var = "OlinkID") %>% 
        pivot_longer(cols = -OlinkID, names_to = "final_sample_id", values_to = "NPX") %>%
        left_join(df_id_pro, by = "OlinkID") %>%
        left_join(df_enza_prepost %>% select(final_sample_id, HCI_cID, mrn, txt_stat, cohort, txt_mcrpc_1st_type), 
                  by = "final_sample_id") %>%
        filter(!is.na(NPX)) %>%
        mutate(txt_stat = factor(txt_stat, levels = c("Pre", "Post")))
      
      # Show sample distribution
      sample_counts_enza <- df_plot_enza %>%
        count(txt_stat) %>%
        mutate(percentage = round(100 * n / sum(n), 1))
      
      cat("Sample distribution for Enzalutamide patients:\n")
      print(sample_counts_enza)
      
      # Create violin plot for Enzalutamide-treated patients
      p_violin_enza <- df_plot_enza %>%
        ggplot(aes(x = txt_stat, y = NPX, fill = txt_stat)) +
        geom_quasirandom(alpha = 0.7, size = 2, color = "grey20") +
        geom_violin(alpha = 0.8, scale = "width", draw_quantiles = c(0.5)) +
        geom_boxplot(outlier.alpha = 0, width = 0.25, coef = 0,
                     alpha = 0.9, color = "white") +
        facet_wrap(~Assay, scales = "free_y", ncol = length(mets_protein_ids)) +
        scale_fill_manual(values = c("Pre" = "#1f77b4", "Post" = "#17becf"),
                          name = "Treatment") +
        labs(x = "Treatment Status", 
             y = "NPX Expression",
             title = "Metastatic Load-Associated Proteins in Enzalutamide-Treated mCRPC",
             subtitle = paste0("Pre vs Post Enzalutamide treatment (n=", 
                              length(unique(df_plot_enza$mrn)), " patients)")) +
        theme_bw() +
        theme(
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(),
          strip.text = element_text(size = 12, face = "bold"),
          axis.text = element_text(size = 11),
          axis.title = element_text(size = 12),
          legend.position = "top",
          legend.title = element_text(size = 12),
          legend.text = element_text(size = 11),
          plot.title = element_text(size = 14, face = "bold"),
          plot.subtitle = element_text(size = 12)
        )
      
      print(p_violin_enza)
      
      # Save the plot
      ggsave(p_violin_enza, 
             file = file.path(wd$outCurr, "Enzalutamide_PrePost_violin.pdf"), 
             width = max(8, length(mets_protein_ids) * 4.5), 
             height = 6)
      
      # Statistical analysis for Enzalutamide-treated patients
      cat("\n=== Statistical analysis: Enzalutamide treatment effect ===\n")
      
      for(protein_id in mets_protein_ids) {
        protein_name <- df_id_pro %>% filter(OlinkID == protein_id) %>% pull(Assay)
        
        # Get data for this protein
        protein_data_enza <- df_plot_enza %>% 
          filter(OlinkID == protein_id)
        
        cat("\n--- ", protein_name, " in Enzalutamide patients ---\n")
        
        # Test treatment effect (unpaired)
        pre_data <- protein_data_enza %>% filter(txt_stat == "Pre") %>% pull(NPX)
        post_data <- protein_data_enza %>% filter(txt_stat == "Post") %>% pull(NPX)
        
        if(length(pre_data) >= 3 & length(post_data) >= 3) {
          # Unpaired t-test
          test_result <- t.test(post_data, pre_data, paired = FALSE)
          
          # Calculate effect size (Cohen's d)
          pooled_sd <- sqrt(((length(post_data)-1)*var(post_data) + (length(pre_data)-1)*var(pre_data))/(length(post_data)+length(pre_data)-2))
          cohens_d <- (mean(post_data) - mean(pre_data)) / pooled_sd
          
          cat("Pre: n=", length(pre_data), " (mean=", round(mean(pre_data), 3), 
              ", sd=", round(sd(pre_data), 3), ")\n")
          cat("Post: n=", length(post_data), " (mean=", round(mean(post_data), 3), 
              ", sd=", round(sd(post_data), 3), ")\n")
          cat("Mean difference:", round(mean(post_data) - mean(pre_data), 3), "\n")
          cat("Enzalutamide effect P = ", formatC(test_result$p.value, format = "e", digits = 2), "\n")
          cat("Effect size (Cohen's d) = ", round(cohens_d, 3))
          
          # Interpret effect size
          if(abs(cohens_d) < 0.2) {
            cat(" (negligible effect)\n")
          } else if(abs(cohens_d) < 0.5) {
            cat(" (small effect)\n")
          } else if(abs(cohens_d) < 0.8) {
            cat(" (medium effect)\n")
          } else {
            cat(" (large effect)\n")
          }
          
          # 95% CI for mean difference
          ci_lower <- test_result$conf.int[1]
          ci_upper <- test_result$conf.int[2]
          cat("95% CI for difference: [", round(ci_lower, 3), ", ", round(ci_upper, 3), "]\n")
          
        } else {
          cat("Insufficient data for statistical testing\n")
          cat("Pre: n=", length(pre_data), ", Post: n=", length(post_data), "\n")
        }
        
        # Check for paired samples if available
        paired_data <- protein_data_enza %>%
          count(mrn) %>%
          filter(n == 2)  # Patients with both Pre and Post
        
        if(nrow(paired_data) >= 3) {
          cat("Paired analysis (n=", nrow(paired_data), " patients with both timepoints):\n")
          
          paired_wide <- protein_data_enza %>%
            filter(mrn %in% paired_data$mrn) %>%
            select(mrn, txt_stat, NPX) %>%
            pivot_wider(names_from = txt_stat, values_from = NPX) %>%
            filter(!is.na(Pre), !is.na(Post))
          
          if(nrow(paired_wide) >= 3) {
            paired_test <- t.test(paired_wide$Post, paired_wide$Pre, paired = TRUE)
            
            cat("Paired t-test P = ", formatC(paired_test$p.value, format = "e", digits = 2), "\n")
            cat("Mean paired change = ", round(mean(paired_wide$Post - paired_wide$Pre), 3), "\n")
          }
        }
      }
      
      # Summary statistics table for Enzalutamide patients
      summary_table_enza <- df_plot_enza %>%
        group_by(Assay, txt_stat) %>%
        summarise(
          n = n(),
          n_patients = n_distinct(mrn),
          mean_NPX = round(mean(NPX, na.rm = TRUE), 3),
          sd_NPX = round(sd(NPX, na.rm = TRUE), 3),
          median_NPX = round(median(NPX, na.rm = TRUE), 3),
          q25_NPX = round(quantile(NPX, 0.25, na.rm = TRUE), 3),
          q75_NPX = round(quantile(NPX, 0.75, na.rm = TRUE), 3),
          min_NPX = round(min(NPX, na.rm = TRUE), 3),
          max_NPX = round(max(NPX, na.rm = TRUE), 3),
          .groups = "drop"
        ) %>%
        arrange(Assay, txt_stat)
      
      cat("\n=== Summary statistics for Enzalutamide-treated patients ===\n")
      print(summary_table_enza)
      
      # Save summary table
      write.csv(summary_table_enza, 
                file.path(wd$outCurr, "Enzalutamide_PrePost_summary.csv"), 
                row.names = FALSE)
      
      # Compare Enzalutamide vs All mCRPC patients
      if(exists("df_plot_mcrpc")) {
        cat("\n=== Comparison: Enzalutamide vs All mCRPC patients ===\n")
        
        for(protein_id in mets_protein_ids) {
          protein_name <- df_id_pro %>% filter(OlinkID == protein_id) %>% pull(Assay)
          
          # Get baseline data for comparison
          enza_baseline <- df_plot_enza %>%
            filter(OlinkID == protein_id, txt_stat == "Pre") %>%
            pull(NPX)
          
          all_mcrpc_baseline <- df_plot_mcrpc %>%
            filter(OlinkID == protein_id, txt_stat == "Pre") %>%
            pull(NPX)
          
          if(length(enza_baseline) >= 3 & length(all_mcrpc_baseline) >= 3) {
            baseline_test <- t.test(enza_baseline, all_mcrpc_baseline)
            
            cat(protein_name, " baseline: Enzalutamide mean =", round(mean(enza_baseline), 3),
                " vs All mCRPC mean =", round(mean(all_mcrpc_baseline), 3),
                ", P =", formatC(baseline_test$p.value, format = "e", digits = 2), "\n")
          }
          
          # Compare treatment effects if possible
          enza_post <- df_plot_enza %>%
            filter(OlinkID == protein_id, txt_stat == "Post") %>%
            pull(NPX)
          
          all_mcrpc_post <- df_plot_mcrpc %>%
            filter(OlinkID == protein_id, txt_stat == "Post") %>%
            pull(NPX)
          
          if(length(enza_post) >= 3 & length(all_mcrpc_post) >= 3) {
            enza_change <- mean(enza_post) - mean(enza_baseline)
            all_change <- mean(all_mcrpc_post) - mean(all_mcrpc_baseline)
            
            cat(protein_name, " treatment effect: Enzalutamide Δ =", round(enza_change, 3),
                " vs All mCRPC Δ =", round(all_change, 3), "\n")
          }
        }
      }
      
    } else {
      cat("No Pre/Post samples found for Enzalutamide-treated patients\n")
    }
    
  } else {
    cat("Treatment information (txt_mcrpc_1st_type) not available in df_c_final\n")
  }
  
} else {
  cat("No metastatic load-associated proteins found for Enzalutamide analysis\n")
}
```

# Save
